% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R ez.R
\name{clean_chr}
\alias{clean_chr}
\title{Clean character}
\usage{
clean_chr(x)
}
\arguments{
\item{x}{Vector of type character}
}
\description{
This function clears a text string, removing accents and non-alphanumeric characters.
}
\examples{
x = c("Test Message (?)")
clean_chr(x)
}
