% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{color.reverse}
\alias{color.reverse}
\title{calculate reverse color given an opacity}
\usage{
color.reverse(fore, opacity = 0.5, back = c(255, 255, 255))
}
\arguments{
\item{fore}{top layer color in c(r,g,b) or #hex}

\item{opacity}{0...1}

\item{back}{bottom layer color in c(r,g,b) or #hex default white background}
}
\value{
A vector of colors. To see the colors, call show.color()
}
\description{
calculate reverse color given an opacity
}
\examples{
# ('#89DAF7')
}
