% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R ez.R
\name{ctrl_v}
\alias{ctrl_v}
\title{Quickly import a df to R}
\usage{
ctrl_v(nom = TRUE)
}
\description{
This function quickly imports a df that is contained in the clipboard to R.
}
\examples{
ctrl_v() # Use this function to import the df
}
