% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R ez.R
\name{date_R}
\alias{date_R}
\title{Convert dates from R}
\usage{
date_R(x)
}
\arguments{
\item{x}{numerical vector}
}
\description{
This function returns a vector of dates with R origin, which is "1970-01-01"
}
\examples{
x = c(18628,18629,18630)
date_R(x)
}
