% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R ez.R
\name{date_stata}
\alias{date_stata}
\title{Convert dates from Stata}
\usage{
date_stata(x)
}
\arguments{
\item{x}{numerical vector}
}
\description{
This function returns a vector of dates with Stata origin, which is "1960-01-01"
}
\examples{
x = c(22281,22282,22283)
date_stata(x)
}
