% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R ez.R
\name{duplicates_count}
\alias{duplicates_count}
\title{Count duplicates}
\usage{
duplicates_count(x)
}
\arguments{
\item{x}{Vector of numeric type or character}
}
\description{
Count duplicates
}
\examples{
data_example = data.frame(x = c(1,2,2,2,3,3))
}
