% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R ez.R
\name{expand_df}
\alias{expand_df}
\title{Expand data frame}
\usage{
expand_df(df, n)
}
\arguments{
\item{df}{Object of type dataframe}

\item{n}{Integer value}
}
\description{
This function expands a dataframe to a desired dimension.
}
\examples{
df_example = data.frame(var_1 = letters[1:3],var_2 = seq(1:3))

expand_df(df_example,5)
}
