% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.2char}
\alias{ez.2char}
\title{convert a column of factor type (or all factor columns) in a data frame into character type. Check with is.factor}
\usage{
ez.2char(x, col = NULL)
}
\arguments{
\item{x}{a data frame or a vector/col}

\item{col}{internally evaluated by eval('dplyr::select()')()
\cr        if x is a data frame, col is specified (e.g., "cond"), convert that col only.
\cr        if x is a data frame, col is unspecified (i.e., NULL default), convert all possible factor cols in x
\cr        if x is not a data frame, col is ignored}
}
\value{
returns a character vector or a data frame with changed col(s)
}
\description{
factor 2 char
}
\details{
Both value and variable label attributes will be removed when converting variables to characters.
}
\seealso{
\code{\link{ez.str}}

Other data.transformation.functions: \code{\link{ez.2factor}},
  \code{\link{ez.2label}}, \code{\link{ez.2long}},
  \code{\link{ez.2lower}}, \code{\link{ez.2value}},
  \code{\link{ez.2wide}}, \code{\link{ez.compute}},
  \code{\link{ez.del}}, \code{\link{ez.label.get}},
  \code{\link{ez.label.set}}, \code{\link{ez.leftjoin}},
  \code{\link{ez.move}}, \code{\link{ez.newcol}},
  \code{\link{ez.num}}, \code{\link{ez.recode}},
  \code{\link{ez.recols}}, \code{\link{ez.recol}},
  \code{\link{ez.rename}}, \code{\link{ez.replace}},
  \code{\link{ez.rncols}}, \code{\link{ez.rncol}},
  \code{\link{ez.select}}, \code{\link{ez.sort}},
  \code{\link{ez.split}}, \code{\link{ez.str}},
  \code{\link{ez.unique}}, \code{\link{ez.values.get}},
  \code{\link{ez.values.set}}
}

