% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.2wide}
\alias{ez.2wide}
\title{reconstruct to wide format, wrapper of \code{\link[stats]{reshape}}}
\usage{
ez.2wide(df, id, indexname, measure = NULL, drop = NULL, ...)
}
\arguments{
\item{id}{unique identification variable, or variable combination}

\item{indexname}{variable name for timing/repetition/index variable, such as "session"}

\item{measure}{column names that are the repeated measures, such as c("BDI_Pre","BDI_Post")}

\item{drop}{variables to drop before reshaping}
}
\value{
returns a new df. Because of the structural change, lable attributes would be lost and can NOT be copied with \code{\link[sjmisc]{copy_labels}}--not a subset.
}
\note{
refer to my spss syntax 'SUBJID * Time [School] - Measure2'
}
\examples{
set.seed(10)
df <- data_frame(
    Person = rep(c("greg", "sally", "sue"), each=2),
    Time = rep(c("Pre", "Post"), 3),
    Score1 = round(rnorm(6, mean = 80, sd=4), 0),
    Score2 = round(jitter(Score1, 15), 0),
    Score3 = 5 + (Score1 + Score2)/2
)
df = data.frame(df)
ez.2wide(df,
        "Person",
        "Time",
        c("Score1","Score2","Score3"))
}
\seealso{
\code{\link[tidyr]{gather}}, \code{\link[tidyr]{spread}}, \code{\link[tidyr]{separate}}, \code{\link[tidyr]{unite}}
\cr \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}
\cr \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{arrange}}
\cr \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{mutate}}
\cr \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{right_join}}, \code{\link[dplyr]{inner_join}}, \code{\link[dplyr]{full_join}}, \code{\link[dplyr]{semi_join}}, \code{\link[dplyr]{anti_join}}
\cr \code{\link[dplyr]{intersect}}, \code{\link[dplyr]{union}}, \code{\link[dplyr]{setdiff}}
\cr \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{bind_cols}}

Other data.transformation.functions: \code{\link{ez.2factor}},
  \code{\link{ez.2label}}, \code{\link{ez.2long}},
  \code{\link{ez.2lower}}, \code{\link{ez.2value}},
  \code{\link{ez.compute}}, \code{\link{ez.delete}},
  \code{\link{ez.del}}, \code{\link{ez.label.get}},
  \code{\link{ez.label.set}}, \code{\link{ez.leftjoin}},
  \code{\link{ez.move}}, \code{\link{ez.newcol}},
  \code{\link{ez.recode2}}, \code{\link{ez.recode}},
  \code{\link{ez.recols}}, \code{\link{ez.recol}},
  \code{\link{ez.rename}}, \code{\link{ez.replace}},
  \code{\link{ez.rncols}}, \code{\link{ez.rncol}},
  \code{\link{ez.select}}, \code{\link{ez.sort}},
  \code{\link{ez.split}}, \code{\link{ez.unique}},
  \code{\link{ez.values.get}}, \code{\link{ez.values.set}}
}

