% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.age}
\alias{ez.age}
\title{Function to calculate age from date of birth.}
\source{
This function was developed in part from this response on the R-Help mailing list.
}
\usage{
ez.age(dob, enddate = Sys.Date(), floorize = FALSE, units = "years",
  precise = TRUE)
}
\arguments{
\item{dob}{a vector of class \code{Date} representing the date of birth/start date}

\item{enddate}{a vector of class Date representing the when the observation's 
age is of interest, defaults to current date.}

\item{floorize}{TRUE/FALSE, age in integer, using \code{\link{floor}}}

\item{units}{character, which units of age should be calculated? allowed values are 
days, months, and years}

\item{precise}{logical indicating whether or not to calculate with leap year 
and leap second precision}
}
\value{
A numeric vector of ages the same length as the dob vector
}
\description{
his function calculates age in days, months, or years from a 
date of birth to another arbitrary date. This returns a numeric vector in 
the specified units.
}
\examples{
x <- as.Date(c("2011-01-01", "1996-02-29"))
age(x[1],x[2], units = "years")  # 3.893151 18.731507
floor(age(x[1],x[2], units = "years")) # 3 18

a <- as.Date(seq(as.POSIXct('1987-05-29 018:07:00'), len=26, by="21 day"))
b <- as.Date(seq(as.POSIXct('2002-05-29 018:07:00'), len=26, by="21 day"))
age <- age(a, units='years')
age
age <- age(a, units='months')
age
age <- age(a, as.Date('2005-09-01'))
age
}
\seealso{
See also \code{\link{difftime}} which this function uses and mimics 
some functionality but at higher unit levels.  \code{\link{ez.date}} \code{\link{ez.is.date}} \code{\link{ez.is.date.convertible}} \code{\link{ez.age}}
}
\author{
Jason P. Becker from package eeptools (sligthly modified by Jerry)
}
