% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.append}
\alias{ez.append}
\title{append a row to an exisiting data frame}
\usage{
ez.append(df, newrow, print2scr = TRUE)
}
\arguments{
\item{df}{df to be appended}

\item{newrow}{a list, e.g.,  list("Ted", 25)  <-- use list can preserve element data type}

\item{print2scr}{whether to print the new row to string (auto separated by tab), default TRUE}
}
\value{
returns a new df, old passed df does not change
}
\description{
could be slow, commonly use together with \code{\link{ez.header}}
}
\note{
although passing in newrow as a vector is fine, c(char,numeric) converts everything to char, bit when saving in excel, num formated as text
}
\examples{
# typical use:
results = ez.header(variable=character(),class=character(),n=numeric())
vars=colnames(x)
for (var in vars) {
    var=
    cls=
    n=
    results = ez.append(results,list(var,cls,n))
}
ez.savex(results,'x.xlsx')
}
\seealso{
\code{\link{ez.header}}
}
