% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.attrclean}
\alias{ez.attrclean}
\title{remove all attr}
\usage{
ez.attrclean(x, col = NULL, ...)
}
\arguments{
\item{x}{a data frame or a vector}

\item{col}{evaluated by \code{\link{ez.selcol}}(x,col). Or, NULL=all cols.}
}
\value{
returns a new data frame or vector
}
\description{
remove all attr
}

