% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.blank2na}
\alias{ez.blank2na}
\title{Replaces blank-ish elements of a factor or character vector to NA}
\usage{
ez.blank2na(x, na.strings = c("", ".", "NA", "na", "N/A", "n/a", "NaN",
  "nan"))
}
\arguments{
\item{x}{a vector of factor or character or any type}

\item{na.strings}{case sensitive strings that will be coverted to NA. The function will do a trimws(x,'both') before conversion. If NULL, do only trimws, no conversion to NA.}
}
\value{
Returns a vector trimws (always for factor, character) and NA converted (if matching na.strings). Attributes will also be kept ('label','labels', 'value.labels').
}
\description{
Replaces blank-ish elements of a factor or character vector to NA
}
\seealso{
\code{\link{ez.nan2na}}, \code{\link{ez.is.numeric.like}}, \code{\link{ez.num}}
}
