% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.boxcox}
\alias{ez.boxcox}
\title{box-cox power transformation}
\usage{
ez.boxcox(y, x = rep(1, length(y)), na.rm = FALSE, plot = TRUE,
  print2scr = TRUE, value = TRUE, value.force = TRUE,
  value.method = c("boxcox", "tukey.modified"), value.lambda = c("rounded",
  "raw"), ...)
}
\arguments{
\item{y}{a vector}

\item{x}{rep(1, length(y))}

\item{na.rm}{rm na from y,x (pairwise), if not, NA stays as is.}

\item{plot}{boxcox plot}

\item{print2scr}{print out transformation parameters}

\item{value}{return transformed y, or a list of transformation parameters}

\item{value.force}{transform regardless or only if p.lambda rounded is less than .05. otherwise original y}

\item{value.method}{boxcox (scaled tukey) or modified tukey, both methods keep the ordering. see \code{\link[car]{bcPower}}}

\item{value.lambda}{use rounded lambda one of c(1, 0, -1, 0.5, 0.33, -0.5, -0.33, 2, -2) or raw/calculated lambda}
}
\value{
returns transformed y, or a list of transformation parameters, otherwise original y (depending on value, value.force)
}
\description{
box-cox power transformation
}
