% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.citen}
\alias{ez.citen}
\title{retrieve article citation numbers from pubmed}
\usage{
ez.citen(xmlFile, outFile = NULL, index = NULL)
}
\arguments{
\item{xmlFile}{EndNote library file, contains exported articles with correct titles}

\item{outFile}{an excel file to store the results, if not provided, same base name as xmlFile}

\item{index}{1:10, same syntax titles[index] to choose a subset of titles in xmlFile to process, NULL=all}
}
\value{
returns invisible, save an excel file with results
}
\description{
retrieve article citation numbers from pubmed
}
\note{
get citation numbers cited by available pubmed central papers
}
