% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.cites}
\alias{ez.cites}
\title{retrieve article citation numbers from pubmed}
\usage{
ez.cites(xmlFile, outFile)
}
\arguments{
\item{xmlFile}{EndNote library file, contains exported articles with correct titles}

\item{outFile}{an excel file to store the results}
}
\value{
returns invisible, save an excel file with results
}
\description{
retrieve article citation numbers from pubmed
}
\note{
get citation numbers cited by available pubmed central papers
}
