% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.clcoldata}
\alias{ez.clcoldata}
\title{sanitize column-wise char data}
\usage{
ez.clcoldata(x, col = NULL, procedures = c("toupper", "removeleading0"))
}
\arguments{
\item{x}{a data frame or a vector}

\item{col}{evaluated by \code{\link{ez.selcol}}(x,col). Or, NULL=all cols.}

\item{procedures}{c('toupper','removeleading0') or 'tolower'}
}
\value{
returns a new data frame or vector
}
\description{
sanitize column-wise char data (if not numeric,logical,date), see also \code{\link{ez.clcolnames}}
}
\seealso{
\code{\link{ez.clcolnames}}
}
