% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.clcols}
\alias{ez.clcols}
\title{clean col names}
\usage{
ez.clcols(df, regex_search = "[[:space:][:punct:]]", regex_replace = "_",
  ignore.case = FALSE, perl = TRUE)
}
\arguments{
\item{df}{df}

\item{regex_search}{search}

\item{regex_replace}{replacement}

\item{ignore.case}{if FALSE, the pattern matching is case sensitive and if TRUE, case is ignored during matching.}

\item{perl}{Perl-compatible regexps be used, without perl, [[:space:][:punct:]] works, but not [\\s[:punct:]]
so seems always a good idea to turn on perl compatible. see \code{\link{gsub}}}
}
\value{
returns a new df with column names cleaned, old df does not change
}
\description{
replace certain characters in all column names, using regular expression and gsub()
}

