% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.clear}
\alias{ez.clear}
\alias{ez.clean}
\alias{ez.cl}
\title{clear, clean  console, workspace, plot or variable(s)}
\usage{
ez.clear(area = 0)

ez.clean(area = 0)

ez.cl(area = 0)
}
\arguments{
\item{area}{0 all, \cr
1 console only \cr
2 workspace only \cr
3 plot only \cr
'var' particular var \cr
c('var1','var2') particular vars}
}
\description{
clear, clean  console, workspace, plot or variable(s)
}
