% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.cnames}
\alias{ez.cnames}
\title{all column names, alias of \code{\link{names}}, there are also names(), colnames(), rownames(), row.names() but no col.names()}
\usage{
ez.cnames(x, do.NULL = TRUE, prefix = "col")
}
\description{
all column names, alias of \code{\link{names}}, there are also names(), colnames(), rownames(), row.names() but no col.names()
}
\seealso{
\code{\link{nrow}}, \code{\link{ncol}}, \code{\link{dim}}, \code{\link{length}}
\cr \code{\link{ez.len}}, \code{\link{ez.size}}
\cr \code{\link{names}}, \code{\link{colnames}}, \code{\link{rownames}}, \code{\link{row.names}},
\cr \code{\link{ez.rnames}}, \code{\link{ez.cnames}}, \code{\link{ez.names}}
}

