% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.corrmap}
\alias{ez.corrmap}
\title{plot a correlation matrix map}
\usage{
ez.corrmap(df, corr.type = "pearson", sig.level = 0.05, insig = "blank",
  method = "circle", tl.col = "black", tl.cex = 0.4, tl.pos = NULL,
  tl.offset = NULL, order = c("original", "AOE", "FPC", "hclust",
  "alphabet"), col = NULL, ...)
}
\arguments{
\item{df}{data frame in wide format, should be all numeric}

\item{corr.type}{"pearson" or "spearman", pairwise deletion for NA}

\item{sig.level}{sig.level. make it 1 to essentially ignore param insig}

\item{insig}{how to treat insig values, one of "pch"(show x, corrplot default),"p-value","blank", "n"(no change, as is)}

\item{method}{"circle", "square", "ellipse", "number", "pie", "shade" and "color". The areas of circles or squares show the absolute value of corresponding correlation coefficients. "color" = same areas}

\item{tl.col}{color of text label}

\item{tl.cex}{size of text label (can still be increased to gain some margins for the map when tl.pos='n')}

\item{tl.pos}{'n' for no text label. NULL=corrplot default=auto(??). Character or logical, position of text labels. If character, it must be one of "lt", "ld", "td", "d" or "n". "lt"(default if type=="full") means left and top, "ld"(default if type=="lower") means left and diagonal, "td"(default if type=="upper") means top and diagonal(near), "d" means diagonal, "n" means don't add textlabel.}

\item{tl.offset}{offset of text label, can somehow adjust margin if necessary (corrplot default 0.4)}

\item{order}{"original" for original order (default).
\cr "AOE" for the angular order of the eigenvectors.
\cr "FPC" for the first principal component order.
\cr "hclust" for the hierarchical clustering order.
\cr "alphabet" for alphabetical order.}

\item{col}{Vector, the color of glyphs. NULL=corrplot default. Jerry auto generates a default one if NULL.}

\item{...}{see \code{\link[corrplot]{corrplot}} for more parameters}
}
\value{
returns a list (r, p) r: a matrix representing the corrmap (p > sig.level, set to NA/blank), p: all raw p values
}
\description{
a wrapper of \code{\link[corrplot]{corrplot}}; the correlation and p values are calculated with \code{\link[Hmisc]{rcorr}}, in which Missing values are deleted in pairs rather than deleting all rows of x having any missing variables
}
