% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.corrmap2}
\alias{ez.corrmap2}
\title{corrmap with scatterplot}
\usage{
ez.corrmap2(x, group = NULL, smooth = TRUE, scale = TRUE,
  density = TRUE, ellipses = FALSE, digits = 2, method = "pearson",
  pch = 20, lm = FALSE, cor = TRUE, jiggle = FALSE, factor = 2,
  hist.col = "#00AFBB", show.points = TRUE, rug = TRUE,
  breaks = "Sturges", cex.cor = 2, wt = NULL, smoother = FALSE,
  stars = TRUE, ci = FALSE, alpha = 0.05, ...)
}
\arguments{
\item{x}{a data.frame or matrix (pairwise deletion for NA)}

\item{group}{factor col name, eg 'Species', to show different groups with different colors in scatterplot}

\item{smooth}{TRUE draws loess smooths}

\item{scale}{TRUE scales the correlation font by the size of the absolute correlation.}

\item{density}{TRUE shows the density plots as well as histograms}

\item{ellipses}{TRUE draws correlation ellipses}

\item{digits}{the number of digits to show}

\item{method}{method parameter for the correlation ("pearson","spearman","kendall")}

\item{pch}{The plot character (defaults to 20 which is a '.').}

\item{lm}{Plot the linear fit rather than the LOESS smoothed fits.}

\item{cor}{If plotting regressions, should correlations be reported?}

\item{jiggle}{Should the points be jittered before plotting?}

\item{factor}{factor for jittering (1-5)}

\item{hist.col}{What color should the histogram on the diagonal be? eg, cyan, light grey}

\item{show.points}{If FALSE, do not show the data points, just the data ellipses and smoothed functions}

\item{rug}{if TRUE (default) draw a rug under the histogram, if FALSE, don't draw the rug}

\item{breaks}{If specified, allows control for the number of breaks in the histogram (see the hist function)}

\item{cex.cor}{If this is specified, this will change the size of the text in the correlations. this allows one to also change the size of the points in the plot by specifying the normal cex values. If just specifying cex, it will change the character size, if cex.cor is specified, then cex will function to change the point size.}

\item{wt}{If specified, then weight the correlations by a weights matrix (see note for some comments)}

\item{smoother}{If TRUE, then smooth.scatter the data points  -- slow but pretty with lots of subjects}

\item{stars}{For those people who like to show the significance of correlations by using magic astricks, set stars=TRUE p-values(0, 0.001, 0.01, 0.05, 1) <=> symbols("***", "**", "*", ".", " ")}

\item{ci}{Draw confidence intervals for the linear model or for the loess fit, defaults to ci=FALSE. If confidence intervals are not drawn, the fitting function is lowess.}

\item{alpha}{The alpha level for the confidence regions, defaults to .05}

\item{...}{other options for pairs}
}
\description{
corrmap with scatterplot
}
