% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.corrmatrix}
\alias{ez.corrmatrix}
\title{calculate correlation matrix (for publication)}
\usage{
ez.corrmatrix(df, file = "CorrMatrix.xlsx")
}
\arguments{
\item{df}{data frame}

\item{file}{if not NULL, save matrix to an excel file}
}
\value{
returns invisible, save an excel file with results
}
\description{
calculate correlation matrix (for publication)
}
\note{
*** p < .001, ** p < .01, * < .05
}
