% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.count}
\alias{ez.count}
\title{Count the occurrence of a single value in data frame columnwise, or rowwise, or both}
\usage{
ez.count(x, val, col = NULL, dim = 3)
}
\arguments{
\item{x}{data frame or vector, if vector, parameters col, dim are ignored}

\item{val}{value to be counted, could be NA. Note would not differentiate 5.0 and 5}

\item{col}{column in string (not number), single or vector. If NULL, all columns used}

\item{dim}{1=along row (rowwise), 2=along col (colwse), 3=area, both, grand total (within specified cols/rows)}
}
\value{
returns a data frame, if dim=1/2; a single value if dim=3. 
\cr vector input x always outputs a single value.
}
\description{
count within one or more than one columns/rows, or entire data frame (ie, all columns/rows)
}
\examples{
sx = c("F", "F", "F", "M", "M", "M")
ht = c(69, 64, 67, 68, 72, 71)
wt = c(148, 132, 142, 149, 167, 165)
people = data.frame(sx, ht, wt)
ez.count(people,'M',dim=2)
}

