% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.countif}
\alias{ez.countif}
\title{Conditionally count the occurrence of a single value in data frame columnwise, or rowwise, or both}
\usage{
ez.countif(x, cnd, col = NULL, dim = 3, na.rm = FALSE)
}
\arguments{
\item{x}{data frame or vector, if vector, parameters col, dim are ignored}

\item{col}{column evaluated by eval('dplyr::select()'), single or vector. If NULL, all columns used}

\item{dim}{1=along row (rowwise), 2=along col (colwse), 3=area, both, grand total (within specified cols/rows)}

\item{cond}{a string like '.>=3', '.=="M"', 'is.na(.)', 'ifelse()'
\cr must use . to refer to each column of data frame (vectorized)
\cr must be able to be evaluated as TRUE/FALSE (essentially sum up TRUE)
\cr wrapper of mutate_all(funs(cnd)), so the same syntax
\cr Note, may not differentiate 5.0 and 5, ie. 5.0==5 TRUE
\cr > < etc, not meaningful for factors, return NA
\cr na.rm=FALSE to catch NA returned above}
}
\value{
returns a data frame, if dim=1/2; a single value if dim=3. 
\cr vector input x always outputs a single value.
}
\description{
count within one or more than one columns/rows, or entire data frame (ie, all columns/rows)
}
\examples{
sx = c("F", "F", "F", "M", "M", "M")
ht = c(69, 64, 67, 68, 72, 71)
wt = c(148, 132, 142, 149, 167, 165)
people = data.frame(sx, ht, wt)
ez.countif(people,'.=="M"','wt', dim=2)  # wt 0
ez.countif(people,'.>150', dim=2)   # sx ht wt   NA  0  2
ez.countif(people$wt, '.==165')
}
\seealso{
\code{\link{ez.countif}}
}

