% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.countplot}
\alias{ez.countplot}
\title{plot count data}
\usage{
ez.countplot(df, cmd, position = "stack", bar_color = "color", alpha = 1,
  stat_size = 5.5, bar_width = 0.7, ylab = NULL, xlab = NULL,
  zlab = NULL, legend_position = "top", legend_direction = "horizontal",
  legend_box = T, legend_size = c(0, 10), xangle = 0, vjust = NULL,
  hjust = NULL, print.cmd = FALSE)
}
\arguments{
\item{df}{data frame in long format (but be careful that standard error might be inaccurate depending on grouping in the long format)}

\item{cmd}{like "x, x|z, x|z a" where x z a are all discrete}

\item{position}{so far can only be "stack" or "fill" ('dodge' not supported yet)}

\item{bar_color}{"bw" or "color"  black/white or colorblind-friendly color}

\item{alpha}{bar alpha value}

\item{stat_size}{set to 0 to hide count/percentage}

\item{bar_width}{the width of bar itself}

\item{ylab}{y label NULL}

\item{xlab}{x label NULL}

\item{zlab}{z/a/fill/legend label, only applicable when there is z provided NULL}

\item{legend_position}{legend position 'top', 'bottom', 'left', 'right', 'none', c(x,y,two-element numeric vector)
\cr         c(0,0) corresponds to the "bottom left" and c(1,1) corresponds to the "top right" position.
\cr         if no z/a (legend) provided, auto force to 'none'}

\item{legend_direction}{horizontal or vertical}

\item{legend_box}{box of legend, T or F}

\item{legend_size}{c(0,10) the first number 0 controls the legend title, 0=hide; the second number controls legend.key.size, legend.text}

\item{xangle}{angle of x text 0}

\item{vjust}{vjust of x text NULL}

\item{hjust}{hjust of x text NULL}

\item{print.cmd}{T/F}
}
\value{
a ggplot object (+theme_apa() to get apa format plot), +scale_y_continuous(limits=c(-5,8),breaks=seq(-5,8,by=2),oob=scales::rescale_none)
\cr see http://stackoverflow.com/a/31437048/2292993 for discussion
}
\description{
plot count data, generate pie chart if one col passed, can also visualize continous data, e.g., 
\cr ez.countplot(iris, 'Species')
\cr ez.countplot(iris, 'Sepal.Length|Species',bar_width = 0.1,stat_size = 3)
}

