% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.csd}
\alias{ez.csd}
\title{current script dir (in full path)}
\usage{
ez.csd()
}
\arguments{
\item{works}{with Rscript, source() or in RStudio Run selection}
}
\description{
current script dir (in full path)
}
