% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.csf}
\alias{ez.csf}
\title{current script file (in full path)}
\usage{
ez.csf()
}
\arguments{
\item{works}{with Rscript, source() or in RStudio Run selection}
}
\description{
current script file (in full path)
}
