% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.date}
\alias{ez.date}
\title{convert to date}
\usage{
ez.date(x, ori = "Excel", format = "\%m/\%d/\%Y", ...)
}
\arguments{
\item{x}{a vector of char, num. param format for class 'character' (ignore ori); param ori for class 'numeric' (ignore format)
\cr a string factor treated as char (ie, using param format), a num factor cannot be processed}

\item{ori}{one of 'Excel', 'Matlab', 'R'}

\item{format}{specify date format. see examples}
}
\value{
returns a vector
}
\description{
convert to date with friendly ori (no need to remember exact origin date)
}
\examples{
# format one of c("\%d/\%m/\%Y", "\%d-\%m-\%Y", "\%Y/\%m/\%d", "\%m/\%d/\%Y", "\%Y-\%m-\%d"). \%y for two year digits
}
\seealso{
\code{\link{ez.date}} \code{\link{ez.is.date}} \code{\link{ez.is.date.convertible}} \code{\link{ez.age}}
}

