% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.debug}
\alias{ez.debug}
\title{debug mode on/off 1/0 T/F. NULL-> getOption('debug')}
\usage{
ez.debug(debugMode = NULL)
}
\value{
if debugMode=NULL, returns invisible T/F (unset=T). if debugMode=1/0, returns nothings
}
\description{
debug mode on/off 1/0 T/F. NULL-> getOption('debug')
}
