% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.describe}
\alias{ez.describe}
\title{print out summary statistics about a data frame or other object, alias of \code{\link[Hmisc]{describe}}}
\usage{
ez.describe(x)
}
\arguments{
\item{x}{a data frame or a vector or sth else that can be converted into a data frame}
}
\description{
print out summary statistics about a data frame or other object, alias of \code{\link[Hmisc]{describe}}
}

