% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.dprime}
\alias{ez.dprime}
\title{dprime and other Signal Detection Theory indices.}
\usage{
ez.dprime(hit, fa, miss, cr, adjusted = TRUE)
}
\arguments{
\item{hit}{Number of hits.}

\item{fa}{Number of false alarms.}

\item{miss}{Number of misses.}

\item{cr}{Number of correct rejections.}

\item{adjusted}{Should it use the Hautus (1995) adjustments for extreme values (hit rate of 1 and false alarm rate of 0). see for more the following notes and \href{https://stats.stackexchange.com/questions/134779}{stackexchange}}
}
\value{
Returns a list objects:
}
\description{
Computes Signal Detection Theory indices (percent of correct, d', beta, A', B''D, c, c').
}
\examples{
hit <- 9
fa <- 2
miss <- 1
cr <- 7

indices <- ez.dprime(hit, fa, miss, cr)


df <- data.frame(Participant = c("A", "B", "C"),
    hit = c(1, 2, 5), miss = c(9, 8, 5),
    fa = c(6, 8, 1), cr = c(4, 2, 9))

indices <- ez.dprime(hit=df$hit,
    fa=df$fa,
    miss=df$miss,
    cr=df$cr,
    adjusted=FALSE)


}
\author{
Jerry modified from \href{https://dominiquemakowski.github.io/}{Dominique Makowski}. See Pallier (2002) for the algorithms, Stanislaw & Todorov (1999) for a good tutorial.
}
