% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.duplicated}
\alias{ez.duplicated}
\title{find the duplicated rows in a data frame}
\usage{
ez.duplicated(df, col = NULL, vec = TRUE)
}
\arguments{
\item{df}{a data frame}

\item{col}{restrict to the columns where you would like to search for duplicates; e.g., 3, c(3), 2:5, "place", c("place","age")
\cr default is NULL, search for all columns}

\item{vec}{TRUE/FALSE, if TRUE, returns a vector of TRUE/FALSE indicating duplicates; 
\cr if FALSE, returns a df with one column 'Duplicated' of TRUE/FALSE}
}
\value{
return depends, see vec above
\cr this is different from the built-in R \code{\link{duplicated}}
\cr x <- c(1, 1, 4, 5, 4, 6)  duplicated(x) returns [1] FALSE TRUE FALSE FALSE TRUE FALSE
\cr but ez.duplicated(x) returns [1] TRUE TRUE TRUE FALSE TRUE FALSE
}
\description{
find the duplicated rows in a data frame
}

