% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.env}
\alias{ez.env}
\title{switch env}
\usage{
ez.env(env = NULL)
}
\arguments{
\item{env}{NULL=print out current env (.libPaths()[1]); file.symlink(sprintf('~/Dropbox/Apps/RStudio/R3.3_library/%s/', env),'/Library/Frameworks/R.framework/Versions/3.3/Resources/library')}
}
\value{
returns NULL
}
\description{
switch env
}

