% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.es.t.independent.msn}
\alias{ez.es.t.independent.msn}
\title{calculate effect size}
\usage{
ez.es.t.independent.msn(m1, s1, n1, m2, s2, n2)
}
\arguments{
\item{m1}{mean}

\item{s1}{standard deviation}

\item{n1}{numbers/subjects/samples group 1}

\item{m2}{mean}

\item{s2}{standard deviation}

\item{n2}{numbers/subjects/samples group 2}
}
\value{
returns invisible
}
\description{
calculate effect size
}
