% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.es.t.independent.tn}
\alias{ez.es.t.independent.tn}
\title{calculate effect size}
\usage{
ez.es.t.independent.tn(t, n1, n2)
}
\arguments{
\item{t}{t (equal variances assumed in SPSS), numbers/subjects/samples group 1 and 2}
}
\value{
returns invisible
}
\description{
calculate effect size
}
