% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.es.t.paired.msr}
\alias{ez.es.t.paired.msr}
\title{calculate effect size}
\usage{
ez.es.t.paired.msr(m1, s1, m2, s2, r)
}
\arguments{
\item{m1}{mean}

\item{s1}{standard deviation}

\item{m2}{mean}

\item{s2}{standard deviation}

\item{r}{correlation}
}
\value{
returns invisible
}
\description{
calculate effect size
}
