% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.eval}
\alias{ez.eval}
\title{wrapper of \code{\link{eval}}}
\usage{
ez.eval(cmd)
}
\arguments{
\item{cmd}{caller's envir, should be environment()  <-- hard to hard-code/default it in this wrapper function, R does not support call stack well.}

\item{env}{an R cmd in text, e.g., constructed with sprintf()}
}
\value{
this function evaluates the cmd in the caller's envir, so the actual return depends on the caller/cmd
}
\description{
wrapper of \code{\link{eval}}
}
\examples{
ez.eval('z=9',environment())
}

