% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.export}
\alias{ez.export}
\title{wrapper of \code{\link{dev.copy2pdf}}}
\usage{
ez.export(filename = "RPlot.pdf", pdf.width = 8.5, pdf.height = 11, ...)
}
\description{
wrapper of \code{\link{dev.copy2pdf}}
}
\note{
internally also use \code{\link[extrafont]{embed_fonts}} to embed fonts
}
\examples{
A4:     width 7(inches) height = 5
Letter: 8.5 x 11
}
\seealso{
\code{\link{pdf}} \code{\link{ez.pdfon}} \code{\link{ez.pdfoff}}
}

