% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.factorname}
\alias{ez.factorname}
\title{change factor level names in a df}
\usage{
ez.factorname(x, col, orn = NULL, print2scr = T)
}
\arguments{
\item{x}{data frame or vector, factor or non-factor}

\item{col}{a single column name, quoted string, ignored when x is not a data frame}

\item{orn}{new level names coresponding to levels(x), eg, c("one","two","three")}
}
\value{
returns a new df, factor (non-factor->factor)
}
\description{
change factor level names in a df
}
\note{
if x df, pass (x,col,orn);  if x not df, pass (x,orn), or (x,orn=). Because of the change, will remove value labels attr
}
\references{
\href{http://www.cookbook-r.com/Manipulating_data/Renaming_levels_of_a_factor/}{Cookbook R: Renaming levels of a factor}
}
\seealso{
\code{\link{ez.recode}}, \code{\link{ez.recode2}}, \code{\link{ez.replace}}, \code{\link{ez.replacewhen}}, \code{\link{ez.2label}}, \code{\link{ez.factorname}}, \code{\link{ez.strreplace}}, \code{\link{ez.strrev}}, \code{\link{ez.regexprep}}, \code{\link{ez.regexprepi}}
}
