% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.find}
\alias{ez.find}
\title{similar to matlab find, find non-zero in a vector}
\usage{
ez.find(v)
}
\arguments{
\item{v}{logical input}
}
\value{
returns index vector
}
\description{
similar to matlab find, find non-zero in a vector
}
\examples{
ez.find(a>1)
}

