% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.fisher.posthoc}
\alias{ez.fisher.posthoc}
\title{Pairwise tests of independence for nominal data}
\usage{
ez.fisher.posthoc(x, compare = "row", fisher = TRUE, gtest = TRUE,
  chisq = TRUE, method = "fdr", correct = "none", cramer = FALSE,
  digits = 3, ...)
}
\arguments{
\item{x}{A two-way contingency table. At least one dimension should have
more than two levels.}

\item{compare}{If \code{"row"}, treats the rows as the grouping variable.
If \code{"column"}, treats the columns as the grouping variable.}

\item{fisher}{If \code{"TRUE"}, conducts fisher exact test.}

\item{gtest}{If \code{"TRUE"}, conducts G-test.}

\item{chisq}{If \code{"TRUE"}, conducts Chi-square test of association.}

\item{method}{The method to adjust multiple p-values. 
See \code{stats::p.adjust}.}

\item{correct}{The correction method to pass to \code{DescTools::GTest}.}

\item{cramer}{If \code{"TRUE"}, includes and effect size, Cramer's V in the
output.
--jerry,does not support this because of cramerV() from the rcompansion package not included in this function}

\item{digits}{The number of significant digits in the output.}

\item{...}{Additional arguments, passed to \code{stats::fisher.test}, 
\code{DescTools::GTest}, or \code{stats::chisq.test}.}
}
\value{
A data frame of comparisons, p-values, and adjusted p-values.
}
\description{
Conducts pairwise tests for a 2-dimensional matrix,
             in which at at least one dimension has more than two
             levels, as a post-hoc test. Conducts Fisher exact, Chi-square,
             or G-test.
}
\examples{
# posthoc for fisher, essentially fisher and then apply fdr correction
# pairwiseNominalIndependence from rcompansion
# nbk \%>\% ez.factorelevel(c('dx','race'))->tmp
# table(tmp$dx,tmp$race) \%>\% ez.fisher.posthoc()
# note: table(tmp$race,tmp$dx) \%>\% ez.fisher.posthoc() is different
# also similar function from RVAideMemoire::fisher.multcomp
# but this funciton can do all (fisher, chisq and g-test)

### Independence test for a 4 x 2 matrix
data(Anderson)
fisher.test(Anderson)
Anderson = Anderson[(c("Heimlich", "Bloom", "Dougal", "Cobblestone")),]
PT = ez.fisher.posthoc(Anderson,
                                 fisher = TRUE,
                                 gtest  = FALSE,
                                 chisq  = FALSE,
                                 cramer = TRUE)
PT                                
cldList(comparison = PT$Comparison,
        p.value    = PT$p.adj.Fisher,
        threshold  = 0.05)                             
                                                              

}
\references{
\url{http://rcompanion.org/handbook/H_04.html}
}
\seealso{
\code{\link{pairwiseMcnemar}}, \code{\link{groupwiseCMH}},
          \code{\link{nominalSymmetryTest}}, 
          \code{\link{pairwiseNominalMatrix}}
}
\author{
Salvatore Mangiafico, \email{mangiafico@njaes.rutgers.edu}
}
\concept{
Chi-square G-test Fisher contingency table nominal
}
