% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.flipflop}
\alias{ez.flipflop}
\title{switch the names and values of a named vector, so that names->values, values->names}
\usage{
ez.flipflop(v)
}
\value{
returns a new named vector, if input is NULL or not named vector, return the inital input vector direclty.
}
\description{
switch the names and values of a named vector, so that names->values, values->names
}
