% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.format.vector}
\alias{ez.format.vector}
\title{format a vector for easy manual copy/processing.}
\usage{
ez.format.vector(vec, quote = TRUE, print2screen = TRUE)
}
\arguments{
\item{vec}{a vector}

\item{quote}{TRUE/FALSE, whether add a quote around each element (switch for string or number)}
}
\value{
nothing, only print out. 
\cr By default in R when you type a variable name, you get [1] "rs171440fwd" "rs1800497fwd"
\cr now with this function you get 'rs171440fwd','rs1800497fwd','rs180043'
}
\description{
format a vector for easy manual copy/processing.
}
\seealso{
\code{\link{ez.print}} \code{\link{ez.pprint}}
}
