% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.fullpath}
\alias{ez.fullpath}
\alias{ez.fp}
\title{Returns the full path by resolving ~ and relative path}
\usage{
ez.fullpath(...)

ez.fp(...)
}
\description{
Returns the full path by resolving ~ and relative path
}
