% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.github}
\alias{ez.github}
\title{alias of \code{\link[devtools]{install_github}}}
\usage{
ez.github(
  repo,
  ref = "HEAD",
  subdir = NULL,
  auth_token = github_pat(quiet),
  host = "api.github.com",
  dependencies = NA,
  upgrade = c("default", "ask", "always", "never"),
  force = FALSE,
  quiet = FALSE,
  build = TRUE,
  build_opts = c("--no-resave-data", "--no-manual", "--no-build-vignettes"),
  build_manual = FALSE,
  build_vignettes = FALSE,
  repos = getOption("repos"),
  type = getOption("pkgType"),
  ...
)
}
\description{
alias of \code{\link[devtools]{install_github}}
}
