% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.gmail}
\alias{ez.gmail}
\title{Send mail with Gmail}
\usage{
ez.gmail(to, subject, htmlbody, attachment = NULL)
}
\arguments{
\item{to}{email address, if multiple seperated by comma}

\item{subject}{'Gmail from R'}

\item{htmlbody}{'<b>Gmailr</b> is a <i>very</i> handy package!' or 'Email using R.'}

\item{attachment}{default NULL, 'BazaarQueriesforURLData.txt'}
}
\value{
returns nothing
}
\description{
Send mail with Gmail
}
\note{
first-time use, gmailr::gmail_auth('compose') to generate ~/.httr-oauth
}
