% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.header}
\alias{ez.header}
\title{create a header for a data frame; also create the data frame}
\usage{
ez.header(..., stringsAsFactors = FALSE)
}
\arguments{
\item{...}{e.g., col1=character(n), col2=numeric(n) where n defaults to 0, represents n of rows in the data frame}

\item{stringsAsFactors}{defaults to FALSE (data.frame() defaults to TRUE)}
}
\value{
returns an (empty) df with specified col names
}
\description{
wrapper of data.frame()
}

