% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.import}
\alias{ez.import}
\title{alias of \code{\link{library}}}
\usage{
ez.import(package, help, pos = 2, lib.loc = NULL, character.only = FALSE,
  logical.return = FALSE, warn.conflicts = TRUE, quietly = FALSE,
  verbose = getOption("verbose"))
}
\description{
alias of \code{\link{library}}
}

