% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.install}
\alias{ez.install}
\title{alias of \code{\link{install.packages}}}
\usage{
ez.install(pkgs, lib, repos = getOption("repos"),
  contriburl = contrib.url(repos, type), method, available = NULL,
  destdir = NULL, dependencies = NA, type = getOption("pkgType"),
  configure.args = getOption("configure.args"),
  configure.vars = getOption("configure.vars"), clean = FALSE,
  Ncpus = getOption("Ncpus", 1L), verbose = getOption("verbose"),
  libs_only = FALSE, INSTALL_opts, quiet = FALSE, keep_outputs = FALSE,
  ...)
}
\description{
alias of \code{\link{install.packages}}
}

