% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.installs}
\alias{ez.installs}
\title{install many packages at the same time}
\usage{
ez.installs(pkgs, load = FALSE, repos = NULL)
}
\arguments{
\item{pkgs}{c()}

\item{load}{if T also load pkgs: sapply(pkgs, require, character.only = TRUE)}

\item{repos}{"https://mran.revolutionanalytics.com/snapshot/2018-01-11"  (if NULL, getOption("repos"))}
}
\description{
install many packages at the same time
}

