% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.is.date.convertible}
\alias{ez.is.date.convertible}
\title{check if a vector of char, number is convertiable to date type}
\usage{
ez.is.date.convertible(x, format = "\%m/\%d/\%Y", ...)
}
\arguments{
\item{x}{a vector of char, number}

\item{format}{specify date format. see examples}

\item{...}{other parameters passed to as.Date(...)}
}
\value{
returns a vector of T/F (vectorized because of is.na() used).
}
\description{
check if a vector of char, number is convertiable to date type as.Date(as.character(x), format)
}
\examples{
mydate = c("10/11/2012","10-12-2012", 345)
# format one of c("\%d/\%m/\%Y", "\%d-\%m-\%Y", "\%Y/\%m/\%d", "\%m/\%d/\%Y", "\%Y-\%m-\%d"). \%y for two year digits
ez.is.date.convertible(mydate, format = "\%m/\%d/\%Y")  # T F F
}
\seealso{
\code{\link{ez.date}} \code{\link{ez.is.date}} \code{\link{ez.is.date.convertible}} \code{\link{ez.age}} \code{\link{ez.time}}
}
