% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.is.integer.like}
\alias{ez.is.integer.like}
\title{Check integers}
\usage{
ez.is.integer.like(x)
}
\arguments{
\item{x}{a numeric variable that is to be tested}
}
\value{
a logical value that indicates that tested variable "looks like" integer
}
\description{
This function tests if given variable "appears" to be an integer. To qualify as such, two conditions need to be satisfied: it should be stored as \code{\link{numeric}} object, and it should pass regular expression test if it consists only of digits.
}
\note{
copied from https://cran.r-project.org/web/packages/rapportools/index.html
\cr The built-in is.integer(x) does not test if x contains integer numbers! For that, use round, as in the function is.wholenumber(x) in its examples.
}
