% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.is.numeric.like}
\alias{ez.is.numeric.like}
\title{Check numeric}
\usage{
ez.is.numeric.like(x)
}
\arguments{
\item{x}{a str vector}
}
\value{
a logical vector
}
\description{
check if a str obj is actually numeric
}
\note{
Using regular expression, https://stackoverflow.com/a/21154566/2292993
\cr TRUE for:  c("-3","+4.4",   "-42","4L","3.9L",   "1.36e4","1.36E4",   "NA","NaN")
\cr positive or negative numbers with no more than one decimal c("-3","+4.4") OR
\cr positive or negative integers (e.g., c("-42","4L","3.9L")) OR
\cr positive or negative numbers in scientific notation c("1.36e4","1.36E4")
\cr 
\cr Special note:
\cr "NA","NaN" (but not "na","nan" etc--use ez.blank2na first) will be treated as numeric as well, ie, returns TRUE.
}
