% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.is.string}
\alias{ez.is.string}
\title{Strings}
\usage{
ez.is.string(x)
}
\arguments{
\item{x}{an object to check}
}
\value{
a logical value indicating whether provided object is a string
}
\description{
Checks if provided object is a string i.e. a length-one character vector.
}
\note{
copied from https://cran.r-project.org/web/packages/rapportools/index.html
}
\examples{
    is.string("foobar")          # [1] TRUE
    is.string(1)                 # [1] FALSE
    is.string(c("foo", "bar"))   # [1] FALSE
}
