% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.join}
\alias{ez.join}
\title{mimic python join, supports vectorization, wrapper of \code{\link{paste}}}
\usage{
ez.join(sep = "", ...)
}
\arguments{
\item{sep}{default is nothing}
}
\description{
mimic python join, supports vectorization, wrapper of \code{\link{paste}}
}
\examples{
ez.join('','a',' x ','b') # "a x b"
}
