% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.label.get}
\alias{ez.label.get}
\title{get variable label, wrapper of \code{\link{sjmisc_get_label}}}
\usage{
ez.label.get(x, cols = NULL)
}
\arguments{
\item{x}{a df: (efc), or a single var: (efc$e42dep)}

\item{cols}{a character (vector), 'e42dep' or c('e42dep','e42anx'). returns only labels for specified cols of the df (ignored if x is a var)}
}
\value{
returns character (if x is a single var), or an named character vector (if x is df). 
\cr If df has no variable label for all variables, returns NULL. If df has label for some variables, returns a string vector with some "". 
\cr However, when cols specified, it is possible to return a string vector with all ""--because cols only for slicing posthoc
}
\description{
get variable label, wrapper of \code{\link{sjmisc_get_label}}
}
\seealso{
\code{\link[tidyr]{gather}}, \code{\link[tidyr]{spread}}, \code{\link[tidyr]{separate}}, \code{\link[tidyr]{unite}}
\cr \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}
\cr \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{arrange}}
\cr \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{mutate}}
\cr \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{right_join}}, \code{\link[dplyr]{inner_join}}, \code{\link[dplyr]{full_join}}, \code{\link[dplyr]{semi_join}}, \code{\link[dplyr]{anti_join}}
\cr \code{\link[dplyr]{intersect}}, \code{\link[dplyr]{union}}, \code{\link[dplyr]{setdiff}}
\cr \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{bind_cols}}

Other data transformation functions: \code{\link{ez.2char}},
  \code{\link{ez.2factor}}, \code{\link{ez.2label}},
  \code{\link{ez.2long}}, \code{\link{ez.2lower}},
  \code{\link{ez.2value}}, \code{\link{ez.2wide}},
  \code{\link{ez.compute}}, \code{\link{ez.del}},
  \code{\link{ez.label.set}}, \code{\link{ez.leftjoin}},
  \code{\link{ez.move}}, \code{\link{ez.newcol}},
  \code{\link{ez.num}}, \code{\link{ez.recode}},
  \code{\link{ez.recols}}, \code{\link{ez.recol}},
  \code{\link{ez.rename}}, \code{\link{ez.replace}},
  \code{\link{ez.rncols}}, \code{\link{ez.rncol}},
  \code{\link{ez.select}}, \code{\link{ez.sort}},
  \code{\link{ez.split}}, \code{\link{ez.str}},
  \code{\link{ez.unique}}, \code{\link{ez.values.get}},
  \code{\link{ez.values.set}}
}
