% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.label.mode}
\alias{ez.label.mode}
\title{set variable label mode}
\usage{
ez.label.mode(df, mode = 1)
}
\arguments{
\item{mode}{1=attr(df,'variable.labels'); 2=attr(df[[col]],'label')-->better for left_join.}
}
\description{
set variable label mode
}
