% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.label.swap}
\alias{ez.label.swap}
\alias{ez.swaplabel}
\title{swap variable label mode}
\usage{
ez.label.swap(df, mode = 1)

ez.swaplabel(df, mode = 1)
}
\arguments{
\item{mode}{1=attr(df,'variable.labels')-->better for slice[1:10,]; 2=attr(df[[col]],'label')-->better for left_join.}
}
\description{
swap variable label mode
}
