% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjmisc18.R
\name{ez.labels.get}
\alias{ez.labels.get}
\alias{ez.getlabels}
\title{get value labels, wrapper of \code{\link{sjmisc_get_labels}}.}
\usage{
ez.labels.get(x, include.values = "n", attr.only = T,
  include.non.labelled = F, ...)

ez.getlabels(x, include.values = "n", attr.only = T,
  include.non.labelled = F, ...)
}
\arguments{
\item{x}{df or var (I generally use x as var)}
}
\description{
get value labels, wrapper of \code{\link{sjmisc_get_labels}}.
}
\details{
see also \code{\link{sjmisc_get_values}}
\cr include.values, how values are returned, as named vector = "as.name" or "n" or "as.prefix" or "p"
\cr attr.only = T, only search attributes, T desired (if F, also search factor levels, character strings)
\cr include.non.labelled, if T, values without labels (when less labels than values) will also be included in the returned labels
\cr returns NULL if no value labels attribute
}
\seealso{
\code{\link[tidyr]{gather}}, \code{\link[tidyr]{spread}}, \code{\link[tidyr]{separate}}, \code{\link[tidyr]{unite}}
\cr \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}
\cr \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{arrange}}
\cr \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{mutate}}
\cr \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{right_join}}, \code{\link[dplyr]{inner_join}}, \code{\link[dplyr]{full_join}}, \code{\link[dplyr]{semi_join}}, \code{\link[dplyr]{anti_join}}
\cr \code{\link[dplyr]{intersect}}, \code{\link[dplyr]{union}}, \code{\link[dplyr]{setdiff}}
\cr \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{bind_cols}}

Other data transformation functions: \code{\link{ez.2char}},
  \code{\link{ez.2factor}}, \code{\link{ez.2label}},
  \code{\link{ez.2long}}, \code{\link{ez.2lower}},
  \code{\link{ez.2value}}, \code{\link{ez.2wide}},
  \code{\link{ez.compute}}, \code{\link{ez.del}},
  \code{\link{ez.label.get}}, \code{\link{ez.label.set}},
  \code{\link{ez.labels.set}}, \code{\link{ez.move}},
  \code{\link{ez.newcol}}, \code{\link{ez.num}},
  \code{\link{ez.recode}}, \code{\link{ez.recols}},
  \code{\link{ez.recol}}, \code{\link{ez.rename}},
  \code{\link{ez.replace}}, \code{\link{ez.rncols}},
  \code{\link{ez.rncol}}, \code{\link{ez.select}},
  \code{\link{ez.sort}}, \code{\link{ez.split}},
  \code{\link{ez.str}}, \code{\link{ez.unique}},
  \code{\link{left_join}}
}
