% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.lineplot}
\alias{ez.lineplot}
\title{line plot with ggplot}
\usage{
ez.lineplot(df, cmd, line_size = 0.7, error_size = 0.7, error_gap = 0,
  error_width = 0.3, error_direction = "both", ylab = NULL, xlab = NULL,
  zlab = NULL, legend_position = "top", legend_direction = "horizontal",
  legend_box = T, legend_size = c(0, 10), xangle = 0, vjust = NULL,
  hjust = NULL)
}
\arguments{
\item{df}{data frame in long format (but be careful that standard error might be inaccurate depending on grouping in the long format)}

\item{cmd}{like "y|x, y|x z, y|x z a" where y (axis) is continous, x (axis) z/a (legend) are discrete, during plot x z a ->x z(za combined)}
}
\value{
a ggplot object (+theme_apa() to get apa format plot)
}
\description{
line plot with ggplot
}

