% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.logistics}
\alias{ez.logistics}
\title{glm(y~x+covar,family=binomial), for many y and/or many x}
\usage{
ez.logistics(df, y, x, covar = NULL, report = T, view = F, plot = F,
  pmethods = c("bonferroni", "fdr"), cols = 3, labsize = 2,
  textsize = 1.5, titlesize = 3, error = T, ...)
}
\arguments{
\item{df}{a data frame. Internally go through dropna (no ez.2value, scale)
\cr glm can handel factor by default (dummy coding)
\cr I do not do standarization here, because 
\cr While standardized coefficients in classic linear regression are well-defined, 
\cr logistic regression, like other generalized linear models, present additional complexity as a result of
\cr the non-linear link function (logit), and non-normal error function (binomial).
\cr https://think-lab.github.io/d/205/}

\item{y}{compatible with \code{\link{ez.selcol}}}

\item{x}{compatible with \code{\link{ez.selcol}}}

\item{covar}{NULL=no covar, compatible with \code{\link{ez.selcol}}}

\item{report}{print results (in APA format)}

\item{view}{call View(result)}

\item{plot}{T/F, the black dash line is bonferroni p = 0.05 (again for tests only with a non-NA p values), the grey black dash is uncorrected p = 0.05}

\item{pmethods}{c('bonferroni','fdr'), type p.adjust.methods for all methods. This correction applies for all possible tests that have been/could be done.}

\item{cols}{number of columns for multiplot. NULL=auto calculate}

\item{error}{whether show error message when error occurs}
}
\value{
an invisible data frame or list of data frame (if many y and many x)
\cr odds_ratio: odds ratio=exp(b), one unit increase in x result in the odds of being 1 for y "OR" times the odds of being 0 for y
\cr so that the variances of dependent and independent variables are 1.
\cr Therefore, standardized coefficients refer to how many standard deviations a dependent variable will change,
\cr per standard deviation increase in the predictor variable.
\cr
\cr dof
}
\description{
glm(y~x+covar,family=binomial), for many y and/or many x
}
