% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.logoff}
\alias{ez.logoff}
\title{Prints/Directs output to both terminal and a file (log.txt) globally, wrapper of \code{\link{ez.log}}}
\usage{
ez.logoff()
}
\value{
nothing
}
\description{
does not accept any parameter
}
\examples{
# logging on: logon()
# logging off: logoff()  # does not accept any parameter
}
\seealso{
\code{\link{ez.print}}
}
