% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.logon}
\alias{ez.logon}
\title{Prints/Directs output to both terminal and a file (log.txt) globally, wrapper of \code{\link{ez.log}}}
\usage{
ez.logon(file = "log.txt", mode = "a", status = TRUE, timestamp = TRUE)
}
\arguments{
\item{mode}{a=append; w=overwrite}

\item{status}{T=open the redirection/file, F=close the redirection}

\item{timestamp}{T=insert timestamp at the beginning and end, F=otherwise}
}
\value{
nothing
}
\description{
Prints/Directs output to both terminal and a file (log.txt) globally, wrapper of \code{\link{ez.log}}
}
\examples{
# logging on: logon()
# logging off: logoff()  # does not accept any parameter
}
\seealso{
\code{\link{ez.print}}
}

