% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.match}
\alias{ez.match}
\title{kinda like filter(), but the order of rows are in the order of vec}
\usage{
ez.match(df, col, vec, nomatch = 0)
}
\arguments{
\item{col}{a single col name}

\item{vec}{elements; see note}

\item{nomatch}{if 0, not return a row for the nomatch; if NA/NULL, return NA.}
}
\value{
returns a new data frame with rows in df[[col]] matching the vec elements. row names are normal (1:nrow)
}
\description{
kinda like filter(), but the order of rows are in the order of vec
}
\note{
works best if your vec contains exactly the same elements as df[[col]], and neither contain duplicate values
\cr Actually, duplicated elements in vec OK, will be mathched multiple times. Duplicates in df[[col]] will be picked using first match. For example, df[[col]] has two rows id1_col2, id1_col1, vec = c(id1,id1), then returns two rows id1_col2, id1_col2.
}
