% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.match2}
\alias{ez.match2}
\title{similar to ez.match}
\usage{
ez.match2(df, col, vec)
}
\description{
similar to ez.match
}
\note{
Two differences from ez.match:
\cr 1. intersect(unique(vec),df[[col]]) will be used to match.
\cr 2. Duplicates in df[[col]] will be kept in order.
\cr For example, df[[col]] has two rows id1_col2, id1_col1, vec = c(id1,id1), then returns two rows id1_col2, id1_col1.
}
