% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.maxnp}
\alias{ez.maxnp}
\title{maximize np by checking recursively various variable combinations (PCA trick)}
\usage{
ez.maxnp(df, targetVar = NULL, fixedVars = NULL)
}
\arguments{
\item{targetVar}{a single outcome var, will calculate its mean in each completed subsamples}

\item{fixedVars}{vars you must have remained. (does not matter if it includes targetVar or not)}
}
\value{
returns a new data frame
}
\description{
maximize np by checking recursively various variable combinations (PCA trick)
}
