% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.moment}
\alias{ez.moment}
\title{moment}
\usage{
ez.moment(format = "\%Y-\%m-\%d_\%H-\%M-\%S")
}
\arguments{
\item{format}{see example or https://www.stat.berkeley.edu/~s133/dates.html}
}
\value{
returns a string
}
\description{
current date,time in string
}
\examples{
# \%a  Abbreviated weekday 
# \%A  Full weekday
# \%b  Abbreviated month   
# \%B  Full month
# \%c  Locale-specific date and time   
# \%d  Decimal date
# \%H  Decimal hours (24 hour) 
# \%I  Decimal hours (12 hour)
# \%j  Decimal day of the year 
# \%m  Decimal month
# \%M  Decimal minute  
# \%p  Locale-specific AM/PM
# \%S  Decimal second  
# \%U  Decimal week of the year (starting on Sunday)
# \%w  Decimal Weekday (0=Sunday)  
# \%W  Decimal week of the year (starting on Monday)
# \%x  Locale-specific Date    
# \%X  Locale-specific Time
# \%y  2-digit year    
# \%Y  4-digit year
# \%z  Offset from GMT 
# \%Z  Time zone (character)
}

