% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.mv}
\alias{ez.mv}
\title{move}
\usage{
ez.mv(from, to)
}
\arguments{
\item{support}{c('a.txt','b.txt')
to parent folder does not have to exist already
ez.mv('a.txt','folder'), ez.mv('a.txt','folder/a.txt'), ez.mv('a.txt','folder/b.txt')
ez.mv('a','b')-->get b/a, b now has a as subfolder, regardless of b exists or not
                 use ez.rn('a','b') to change name a->b}
}
\description{
move
}
