% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.na.keep}
\alias{ez.keepna}
\alias{ez.na.keep}
\title{keep rows that have a certain number (range) of NAs anywhere/somewhere and delete others}
\usage{
ez.na.keep(df, col = NULL, n = 0, reindex = TRUE)

ez.keepna(df, col = NULL, n = 0, reindex = TRUE)
}
\arguments{
\item{df}{a data frame}

\item{col}{restrict to the columns where you would like to search for NA; eg, 3, c(3), 2:5, "place", c("place","age")
\cr default is NULL, search for all columns}

\item{n}{integer or vector, 0, c(3,5), number/range of NAs allowed.
\cr If a number, the exact number of NAs kept
\cr Range includes both ends 3<=n<=5
\cr Range could be -Inf, Inf}

\item{reindex}{whether to keep original row index or reindex
\cr eg, original rownames() is 1, 2, 3, then drop row 2
\cr if not reindex, new index is 1, 3
\cr if reindex, new index is 1, 2}
}
\value{
returns a new df with rows that have NA(s) removed
}
\description{
keep rows that have a certain number (range) of NAs anywhere/somewhere and delete others

alias ez.na.keep, ez.keepna
}

