% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.nan2na}
\alias{ez.nan2na}
\title{Replaces non-finite values (Inf, -Inf, NaN, NA) with \code{NA}s in array-like objects.}
\usage{
ez.nan2na(x)
}
\arguments{
\item{x}{An array-like object, or vector}
}
\value{
Returns \code{x} with non-finite values (Inf, -Inf, NaN, NA) replaced with \code{NA}s.
}
\description{
Replaces non-finite values (Inf, -Inf, NaN, NA) with \code{NA}s in array-like objects.
}
\note{
is.na() returns TRUE for both NA and NaN, however is.nan() return TRUE for NaN (0/0) and FALSE for NA.
\cr mean(c(NA,NA),na.rm=T)  ---> NaN, mean(c(NA,NA),na.rm=F) ---> NA  sum(c(NA,NA),na.rm=T)  ---> 0, sum(c(NA,NA),na.rm=F) ---> NA
}
\examples{
ez.nan2na(Inf)
ez.nan2na(rep(c(0, -Inf, Inf, NA, NaN), 3))
ez.nan2na(matrix(c(0, Inf, -Inf, 0), 2, 2))
ez.nan2na(array(c(0, -Inf, Inf, 1, NaN), dim = c(2, 3, 4)))
ez.nan2na(ts(rep(c(0, -Inf, Inf), 2), frequency = 5))
ez.nan2na(factor(c('ab','de'))), ez.nan2na(c('ab','de'))  # returns the original
}
\seealso{
\code{\link{is.finite}}, \code{\link{NA}}
}
