% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.open}
\alias{ez.open}
\title{dummpy
# +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# environment, file, folder, os
# +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# options(warn=-1)  # turn off warning
# options(warn=0)  # turn back on warning
open a local file with default program (xlsx etc)}
\usage{
ez.open(filePath)
}
\description{
open a local file with default program (xlsx etc)
}
\note{
\code{\link{browseURL}} does the same
}
