% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.outlier}
\alias{ez.outlier}
\title{univariate outlier cleanup}
\usage{
ez.outlier(x, col = NULL, method = c("z", "mad", "iqr"), cutoff = NULL,
  plot = TRUE, fillout = c("na", "mean", "median"), na.rm = TRUE,
  print2scr = TRUE)
}
\arguments{
\item{x}{a data frame or a vector}

\item{col}{passed to \code{\link{ez.selcol}}. colwise processing
\cr        if x is a data frame, col specified, process that col only.
\cr        if x is a data frame, col unspecified (i.e., NULL default), process all cols
\cr        if x is not a data frame, col is ignored
\cr        could be multiple cols}

\item{method}{z score or IQR (John Tukey)}

\item{cutoff}{abs() > cutoff will be treated as outliers. Default values(if NULL):
\cr z 95% of values fall within 1.96, qnorm(0.025,lower.tail=F), or 3
\cr mad 2.5, which is the standard recommendation, or 5.2
\cr iqr 1.5}

\item{plot}{boxplot and hist before and after outlier processing. Only works for vector (imagine plot for dataframe)}

\item{fillout}{how to process outlier, fill with na, mean, median (columnwise for data frame)}
}
\value{
returns a new data frame or vector
}
\description{
univariate outlier cleanup
}
\note{
univariate outlier approach
The Z-score method relies on the mean and standard deviation of a group of data to measure central
tendency and dispersion. This is troublesome, because the mean and standard deviation are highly
affected by outliers – they are not robust. In fact, the skewing that outliers bring is one of the
biggest reasons for finding and removing outliers from a dataset!
Another drawback of the Z-score method is that it behaves strangely in small datasets – in fact,
the Z-score method will never detect an outlier if the dataset has fewer than 12 items in it.
\cr
\cr
Median absolute deviation, modified z-score. The median and MAD are robust measures of central tendency and dispersion, respectively.
\cr
\cr
Interquartile range method is that, like the modified Z-score method, it uses a robust measure of dispersion.
\cr
}
\examples{
set.seed(1234)
x = rnorm(10)



}
