% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.p.apa}
\alias{ez.p.apa}
\title{format p value according to apa for report}
\usage{
ez.p.apa(pvalue)
}
\arguments{
\item{pvalue}{numeric vector}
}
\value{
character vector (p < .001, p = .003, p = .02)
}
\description{
format p value according to apa for report
}
