% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.pdfon}
\alias{ez.pdfon}
\title{wrapper of \code{\link{pdf}}}
\usage{
ez.pdfon(file = if (onefile) "Rplots.pdf" else "Rplot\%03d.pdf", width,
  height, onefile, family, title, fonts, version, paper, encoding, bg, fg,
  pointsize, pagecentre, colormodel, useDingbats, useKerning, fillOddEven,
  compress)
}
\description{
wrapper of \code{\link{pdf}}
}
\note{
additionally one can use \code{\link[extrafont]{embed_fonts}} to embed fonts after dev.off()
}
\examples{
A4:     width 7(inches) height = 5
Letter: 8.5 x 11
}
\seealso{
\code{\link{ez.export}} \code{\link{ez.pdfoff}}
}

