% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.piechart}
\alias{ez.piechart}
\title{pie chart}
\usage{
ez.piechart(df, cmd, start = 0, direction = 1, color = "color",
  alpha = 1, n.size = 5.5, n.type = 3, ylab = "", xlab = "",
  zlab = NULL, legend.position = "top", legend.direction = "horizontal",
  legend.box = T, legend.size = c(0, 10), xangle = 0, vjust = NULL,
  hjust = NULL)
}
\arguments{
\item{df}{data frame in long format (but be careful that standard error might be inaccurate depending on grouping in the long format)}

\item{cmd}{a single col name}

\item{start}{offset of starting point from 12 o'clock in radians eg, pi/3}

\item{direction}{1, clockwise; -1, anticlockwise}

\item{color}{"bw" or "color"  black/white or colorblind-friendly color}

\item{alpha}{pie alpha value}

\item{n.size}{set to 0 to hide count/percentage}

\item{n.type}{1 = pct; 2 = n; 3 = pct (n); 4 = n (pct)}

\item{ylab}{y label}

\item{xlab}{x label}

\item{zlab}{z/a/fill/legend label, only applicable when there is z provided NULL}

\item{legend.position}{legend position 'top', 'bottom', 'left', 'right', 'none', c(x,y,two-element numeric vector)
\cr         c(0,0) corresponds to the "bottom left" and c(1,1) corresponds to the "top right" position.
\cr         if no z/a (legend) provided, auto force to 'none'}

\item{legend.direction}{horizontal or vertical}

\item{legend.box}{box of legend, T or F}

\item{legend.size}{c(0,10) the first number 0 controls the legend title, 0=hide; the second number controls legend.key.size, legend.text}

\item{xangle}{angle of x text 0}

\item{vjust}{vjust of x text NULL}

\item{hjust}{hjust of x text NULL}
}
\value{
a ggplot object (+theme_apa() to get apa format plot), +scale_y_continuous(limits=c(-5,8),breaks=seq(-5,8,by=2),oob=scales::rescale_none)
\cr see http://stackoverflow.com/a/31437048/2292993 for discussion
}
\description{
pie chart
}
