% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.play}
\alias{ez.play}
\title{Generate a predefine data frame for testing, debugging, playing and etc}
\usage{
ez.play()
}
\description{
Generate a predefine frame for testing, debugging, playing and etc
}
