% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.plot}
\alias{ez.plot}
\title{plot a customized boxplot with jittered stripplot, violin, and mean}
\usage{
ez.plot(df, cmd, violin = FALSE, show.n = TRUE, print.cmd = FALSE)
}
\arguments{
\item{df}{data frame in long format}

\item{cmd}{like "y", "y|x z a", "y|x z" or "y|x" where y is continous, x z a are discrete}

\item{violin}{plot violin or not}

\item{show.n}{show n (number of samples for each level)  or not}

\item{print.cmd}{T/F}
}
\value{
a ggplot object (+theme_apa() to get apa format plot)
}
\description{
plot a customized boxplot with jittered stripplot, violin, and mean
}

