% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.ppq}
\alias{ez.ppq}
\title{ppq}
\usage{
ez.ppq(...)
}
\arguments{
\item{...}{internal for list(...), see example}
}
\value{
returns a data frame
}
\description{
Pin Pin Qi, input (named) vectors of different length and output a data frame with NA appended to the same length
}
\examples{
ez.ppq(a=1:4,b=2:4)
# a  b
# 1  2
# 2  3
# 3  4
# 4 NA
}
