% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.pprint}
\alias{ez.pprint}
\title{colorful print}
\usage{
ez.pprint(string, color = "green")
}
\arguments{
\item{color}{string, eg, bold, italic, underline, strikethrough, black, red, green, yellow, blue, magenta, cyan, white, silver, bgWhite, bgGreen, bgCyan etc}
}
\description{
colorful print
}

