% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.print}
\alias{ez.print}
\title{wrapper of \code{\link{cat}}, cat(string, "\n", sep='')}
\usage{
ez.print(string, file = "", append = TRUE, sep = "", fill = FALSE,
  labels = NULL)
}
\arguments{
\item{string}{a single character vector, other types will be auto converted with as.character() by cat()
\cr        ez.print(factor(4:5)) --> 12
\cr        ez.print(c(pi,'=',3))  --> 3.14159265358979=3
\cr        ez.print(c('var','\\t','p')) --> var\tp}

\item{file}{filepath, default ''=print to console; if provided, save to a file and no output in console}

\item{append}{T/F}

\item{sep}{default empty. a character vector of strings to append after each element.}
}
\value{
each print generates a new line automatically
}
\description{
wrapper of \code{\link{cat}}, cat(string, "\n", sep='')
}
\details{
do not use R \code{\link{print}}, not actually printing \\n
}
\examples{
# cat("hello","world",file="output.txt",sep="\\n",append=TRUE)
}
\seealso{
\code{\link{sprintf}}, \code{\link{ez.log}}, \code{\link{ez.join}}
}
