% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.quotient}
\alias{ez.quotient}
\title{quotient}
\usage{
ez.quotient(m, n)
}
\arguments{
\item{m}{dividend}

\item{n}{divider}
}
\description{
quotient
}
\examples{
11 is given as the dividend and 5 the divider, (ie, 11 divided by 5, or divide 11 by 5), then 2(quotient) and 1(remainder).
}
\seealso{
\code{\link{ez.remainder}}
}
