% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.r}
\alias{ez.r}
\title{r correlation}
\usage{
ez.r(x, col = NULL, type = "pearson")
}
\arguments{
\item{x}{df or matrix}

\item{col}{col included for calculation, ignored if x is not a df}

\item{type}{"pearson" or "spearman"}
}
\value{
r matrix
}
\description{
r correlation
}
