% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.radarmap}
\alias{ez.radarmap}
\title{draw a radar plot}
\usage{
ez.radarmap(df, id, stats = "mean", lwd = 1, angle = 0, fontsize = 0.8,
  facet = FALSE, facetfontsize = 1, color = id, linetype = NULL,
  print.cmd = FALSE)
}
\arguments{
\item{df}{a data frame with a single variable as factor and all othere variables being numeric}

\item{id}{column name as id in quotes ""}

\item{stats}{perform stats (with NA removed) grouped by id, "mean", "median", "sd", "var", "none"; if "none", assume df already has summary stats}

\item{lwd}{line width}

\item{angle}{angle to show the variable names}

\item{fontsize}{fontsize of variable names, if 0, no variable names shown}

\item{facet}{if F show each id level separately; if T show all levels in one plot}

\item{facetfontsize}{fontsize of id level names (only valid when facet=T)}

\item{color}{color for different id levels, if NULL, remain the same for different id levels}

\item{linetype}{linetype for different id levels, if NULL, remain the same for different id levels}

\item{print.cmd}{T/F}
}
\value{
a ggplot object (+theme_apa() to get apa format plot)
}
\description{
variables/axes arranged in the shape of a radar map, scales are rescaled to 0-1
\cr best works with multivariates
}
\note{
As a reminder, the returned ggplot object can be modified post-hoc
}
\references{
\href{http://www.cmap.polytechnique.fr/~lepennec/R/Radar/RadarAndParallelPlots.html}{Erwan Le Pennec - CMAP}
}

