% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.read}
\alias{ez.read}
\title{read csv table, wrapper of \code{\link{read.csv}}}
\usage{
ez.read(file, ..., skip.rows = NULL, tolower = FALSE)
}
\arguments{
\item{skip.rows}{rows to skip (1 based) before read in, eg 1:3}

\item{tolower}{whether to convert all column names to lower case}
}
\value{
returns a data frame
}
\description{
read csv table, wrapper of \code{\link{read.csv}}
}
\examples{
read.table(file, header = FALSE, sep = "",
           dec = ".", numerals = c("allow.loss", "warn.loss", "no.loss"),
           row.names, col.names, as.is = !stringsAsFactors,
           na.strings = "NA", colClasses = NA, nrows = -1,
           skip = 0, check.names = TRUE, fill = !blank.lines.skip,
           strip.white = FALSE, blank.lines.skip = TRUE,
           comment.char = "#",
           allowEscapes = FALSE, flush = FALSE,
           stringsAsFactors = default.stringsAsFactors(),
           fileEncoding = "", encoding = "unknown", text, skipNul = FALSE)

read.csv(file, header = TRUE, sep = ",",
         dec = ".", fill = TRUE, comment.char = "", ...)
}
