% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.readp}
\alias{ez.readp}
\title{read a parquet (a columnar storage file format) file}
\usage{
ez.readp(
  file,
  col_select = NULL,
  as_data_frame = TRUE,
  props = ParquetArrowReaderProperties$create(),
  mmap = TRUE,
  ...
)
}
\description{
read a parquet (a columnar storage file format) file
}
\note{
alias of \code{\link[arrow]{read_parquet}}
}
