% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.readplist}
\alias{ez.readplist}
\title{read a list of parquet (a columnar storage file format) file}
\usage{
ez.readplist(plist, ...)
}
\description{
read a list of parquet (a columnar storage file format) file
}
\note{
Returns a data.frame. Use \code{\link[data.table]{rbindlist}}. ... passed to \code{\link[arrow]{read_parquet}}
}
