% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.readr}
\alias{ez.readr}
\title{read a rds file which contains only one obj}
\usage{
ez.readr(file, refhook = NULL)
}
\value{
obj = ez.readr(file)
}
\description{
read a rds file which contains only one obj
}
\seealso{
alias of \code{\link{readRDS}}
}
