% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.reads2}
\alias{ez.reads2}
\title{wrapper of \code{\link[sjmisc]{read_spss}}}
\usage{
ez.reads2(..., tolower = FALSE)
}
\arguments{
\item{tolower}{whether to convert all column names to lower case}
}
\description{
potentially keep variable labels and value labels, but does not have string space-trimming capability
}

