% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.readt}
\alias{ez.readt}
\title{read csv table, wrapper of \code{\link{read.csv}}}
\usage{
ez.readt(file, ..., skip.rows = NULL, tolower = FALSE, makenames = TRUE)
}
\arguments{
\item{...}{more useful parameters from \code{\link{read.table}} \code{\link{read.csv}}, eg. strip.white, blank.lines.skip, na.strings, stringsAsFactors}

\item{skip.rows}{rows to skip (1 based) before read in, eg 1:3, if negative,eg, -3:-5, counting from the last line}

\item{tolower}{whether to convert all column names to lower case}

\item{makenames}{if F, keep as is. if T, call make.names(unique=TRUE,allow_=TRUE) _ kept, The character "X" is prepended if necessary. All invalid characters are translated to "." .1 .2 etc appended for uniqueness}
}
\value{
returns a data frame
}
\description{
read csv table/text, wrapper of \code{\link{read.csv}}
}
\examples{
read.table(file, header = FALSE, sep = "",
           dec = ".", numerals = c("allow.loss", "warn.loss", "no.loss"),
           row.names, col.names, as.is = !stringsAsFactors,
           na.strings = "NA", colClasses = NA, nrows = -1,
           skip = 0, check.names = TRUE, fill = !blank.lines.skip,
           strip.white = FALSE, blank.lines.skip = TRUE,
           comment.char = "#",
           allowEscapes = FALSE, flush = FALSE,
           stringsAsFactors = default.stringsAsFactors(),
           fileEncoding = "", encoding = "unknown", text, skipNul = FALSE)

read.csv(file, header = TRUE, sep = ",",
         dec = ".", fill = TRUE, comment.char = "", ...)
}
