% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.readxlist}
\alias{ez.readxlist}
\title{read an xlsx file, returns and prints sheet names, wrapper of \code{\link[openxlsx]{getSheetNames}}}
\usage{
ez.readxlist(file, toprint = TRUE)
}
\arguments{
\item{toprint}{print out sheet indices and names, default TRUE}
}
\value{
a list of sheet names
}
\description{
read an xlsx file, returns and prints sheet names, wrapper of \code{\link[openxlsx]{getSheetNames}}
}
\examples{
readxlist(file)
}

