% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.reindex}
\alias{ez.reindex}
\title{reset the index of a data frame from 1...N}
\usage{
ez.reindex(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
returns a new df
}
\description{
internally call row.names(df) <- NULL
}
\examples{
occasionally, the index of a data frame could be broken, eg, after removing a row:
original row.names() is 1, 2, 3, then drop row 2
if not reindex, new index is 1, 3
if reindex, new index is 1, 2
}
\seealso{
\code{\link{ez.rnames}}
}
