% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.relevelfactor}
\alias{ez.relevelfactor}
\title{reset factor levels in a df after its levels have been modified, relevel a factor in order to reflect its new levels}
\usage{
ez.relevelfactor(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{a factor column name, quoted "", eg, "group"}
}
\value{
returns a new df
}
\description{
does not change factor label (set factor order as is)
}

