% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.remainder}
\alias{ez.remainder}
\title{remainder}
\usage{
ez.remainder(m, n)
}
\arguments{
\item{m}{dividend}

\item{n}{divider}
}
\description{
remainder
}
\examples{
11 is given as the dividend and 5 the divider, (ie, 11 divided by 5, or divide 11 by 5), then 2(quotient) and 1(remainder).
}
\seealso{
\code{\link{ez.quotient}}
}

