% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.replace}
\alias{ez.replace}
\title{replace a single value in data frame with another value}
\usage{
ez.replace(df, col, oldval, newval)
}
\arguments{
\item{df}{data frame}

\item{col}{column name in string}

\item{oldval}{old value (e.g., -Inf)}

\item{newval}{new value (e.g., NA)}
}
\value{
returns a new df, old one does not change
}
\description{
see also \code{\link{ez.strreplace}} \code{\link{ez.recode}} \code{\link{ez.recode2}} 
\cr wrapper of df[[col]][which(df[[col]]==oldval)] <- newval
\cr the "==" syntax within "which()" could be modified
}
\seealso{
Other data.transformation.functions: \code{\link{ez.2factor}},
  \code{\link{ez.2label}}, \code{\link{ez.2long}},
  \code{\link{ez.2lower}}, \code{\link{ez.2value}},
  \code{\link{ez.2wide}}, \code{\link{ez.compute}},
  \code{\link{ez.delete}}, \code{\link{ez.del}},
  \code{\link{ez.label.get}}, \code{\link{ez.label.set}},
  \code{\link{ez.leftjoin}}, \code{\link{ez.move}},
  \code{\link{ez.newcol}}, \code{\link{ez.recode2}},
  \code{\link{ez.recode}}, \code{\link{ez.recols}},
  \code{\link{ez.recol}}, \code{\link{ez.rename}},
  \code{\link{ez.rncols}}, \code{\link{ez.rncol}},
  \code{\link{ez.select}}, \code{\link{ez.sort}},
  \code{\link{ez.split}}, \code{\link{ez.unique}},
  \code{\link{ez.values.get}}, \code{\link{ez.values.set}}
}

