% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.replacewhen}
\alias{ez.replacewhen}
\title{replace when}
\usage{
ez.replacewhen(df, ...)
}
\arguments{
\item{df}{df}

\item{...}{pt_num=1220,baby_num=3,baby_name='Bennnnnnn',the first element is used as condition to pinpoint the row(s), the rest as cols to be replaced. Quotes around col names are optional, 'pt_num'=1220.}
}
\value{
returns a new df
}
\description{
replace a df: eg, when pt_num=1220, let baby_num=3,baby_name='Bennnnnnn'
\cr keep data type whenever possible, remove all attr of col (otherwise could be inconsistent)
}
\details{
smilar to \code{\link{ez.recode}}num->num (if get replaced with another num), numeric->char (if get replaced with a char), char->char, factor->factor (factor internally converted to char then back to factor)
\cr wrapper of df[[col]][theRow] <- newval
\cr df[theRow,col]=newval  # this syntax works also, but df[145:146,2,drop=F]=4 says unused arg drop=F
}
\examples{
df=ez.replacewhen(nicu,pt_num=1220,baby_num=3,baby_name='Ben')
}
\seealso{
\code{\link{ez.replace}}
}
