% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.require}
\alias{ez.require}
\title{require a package, if not exist auto install and load}
\usage{
ez.require(pkg, load = TRUE)
}
\arguments{
\item{pkg}{pkg name}

\item{load}{auto load or not (default=TRUE)}
}
\description{
require a package, if not exist auto install and load
}

