% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.rescale01}
\alias{ez.rescale01}
\title{rescale a vector to 0-1}
\usage{
ez.rescale01(x)
}
\value{
if x is not numeric, return original x (unchanged)
}
\description{
rescale a vector to 0-1
}
\note{
step 1: if there is negative value, x - min(x); step 2: x/max(x)
}
