% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.reshape}
\alias{ez.reshape}
\title{reshape matrix}
\usage{
ez.reshape(a, n, m)
}
\value{
reshape(a, n, m) returns the n-by-m matrix whose elements are taken column-wise from a.
\cr An error results if a does not have n*m elements. If m is missing, it will be calculated from n and the size of a.
}
\description{
reshape matrix
}
