% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.rm}
\alias{ez.rm}
\title{remove a file, wrapper of \code{\link{unlink}}}
\usage{
ez.rm(x)
}
\description{
remove a file, wrapper of \code{\link{unlink}}
}
\examples{
support c('a.txt','b.txt')
x, a character vector with the names of the file(s) or directories to be deleted.
Wildcards (normally ‘*’ and ‘?’) are allowed.
0 for success, 1 for failure.
}
