% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.sanitize}
\alias{ez.sanitize}
\title{sanitize column-wise}
\usage{
ez.sanitize(x, col = NULL, procedures = c("toupper", "removeleading0"))
}
\arguments{
\item{x}{a data frame or a vector}

\item{col}{evaluated by \code{\link{ez.selcol}}(x,col). Or, NULL=all cols.}

\item{procedures}{c('toupper','removeleading0')}
}
\value{
returns a new data frame or vector
}
\description{
sanitize column-wise (if not numeric,logical,date), see also \code{\link{ez.clcols}}
}
\seealso{
\code{\link{ez.clcols}}
}

