% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.savep}
\alias{ez.savep}
\alias{ez.writep}
\title{save a df to a parquet (a columnar storage file format) file}
\usage{
ez.savep(x, sink, chunk_size = NULL, version = NULL, compression = NULL,
  compression_level = NULL, use_dictionary = NULL,
  write_statistics = NULL, data_page_size = NULL,
  properties = ParquetWriterProperties$create(x, version = version,
  compression = compression, compression_level = compression_level,
  use_dictionary = use_dictionary, write_statistics = write_statistics,
  data_page_size = data_page_size), use_deprecated_int96_timestamps = FALSE,
  coerce_timestamps = NULL, allow_truncated_timestamps = FALSE,
  arrow_properties = ParquetArrowWriterProperties$create(use_deprecated_int96_timestamps
  = use_deprecated_int96_timestamps, coerce_timestamps = coerce_timestamps,
  allow_truncated_timestamps = allow_truncated_timestamps))

ez.writep(x, sink, chunk_size = NULL, version = NULL, compression = NULL,
  compression_level = NULL, use_dictionary = NULL,
  write_statistics = NULL, data_page_size = NULL,
  properties = ParquetWriterProperties$create(x, version = version,
  compression = compression, compression_level = compression_level,
  use_dictionary = use_dictionary, write_statistics = write_statistics,
  data_page_size = data_page_size), use_deprecated_int96_timestamps = FALSE,
  coerce_timestamps = NULL, allow_truncated_timestamps = FALSE,
  arrow_properties = ParquetArrowWriterProperties$create(use_deprecated_int96_timestamps
  = use_deprecated_int96_timestamps, coerce_timestamps = coerce_timestamps,
  allow_truncated_timestamps = allow_truncated_timestamps))
}
\description{
save a df to a parquet (a columnar storage file format) file
}
\note{
alias of \code{\link[arrow]{write_parquet}}
}
