% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.saver}
\alias{ez.saver}
\alias{ez.writer}
\title{save one or more or all objects to a rda file}
\usage{
ez.saver(file, vars = NULL, ...)

ez.writer(file, vars = NULL, ...)
}
\arguments{
\item{file}{file path to a rda file}

\item{vars}{a vector of obj names. NULL=all. 'var1', c('var1','var2'). If you have want a particular name, you have to assign that name as a variable first. e.g., desiredName=x; ez.saver(file,c('desiredName'))}

\item{...}{see ... for \code{\link{save}} or \code{\link{save.image}}}
}
\value{
returns nothing
}
\description{
save one or more or all objects to a rda file
}
\seealso{
\code{\link{ez.readr}}, \code{\link{ez.savem}}
}
