% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.scatterplot}
\alias{ez.scatterplot}
\title{scatter plot with lattice}
\usage{
ez.scatterplot(df, cmd, loess = FALSE, model = c("lm", "lmrob", "lmRob",
  "rlm"), type = c("partial", "spartial"), scale = FALSE, rp = TRUE,
  rp.x = 0.025, rp.y = 0.05, se = TRUE, layout = NULL, rp.size = 14,
  line.width = 3, point.size = 14, x.axis.size = 16, y.axis.size = 16,
  x.lab.size = 18, y.lab.size = 18, title.size = 20, legend.size = c(0,
  14), line.color = "#BE1B22", line.style = 1, loess.color = "dark grey",
  loess.width = 3, loess.style = 2, point.color = "#0086B8",
  point.shape = 16, point.alpha = 0.9, colors = ez.palette("Zhu"),
  shapes = c(16, 17, 15, 3, 7, 8), ylab = NULL, xlab = NULL,
  x.tick.number = 5, y.tick.number = 5, title = NULL, zlab = NULL,
  legend.box = FALSE, legend.position = "top",
  legend.direction = "horizontal", hack = FALSE, print2scr = TRUE, ...)
}
\arguments{
\item{df}{data frame}

\item{cmd}{like "y~x+a+b", "y~x+a+b|z", "y~x+a+b||z", "y~x+a+b|||z"where y x are continous, z discrete (| one regression line, || multiple regression lines by levels of z--gives interaction p value), +a+b optional for covariates residualization
\cr y~x+a+b|||z plots separated by z groups}

\item{loess}{T/R adds a loess fit (uses panel.loess, the same as type = c("smooth"))}

\item{model}{one of c('lm', 'lmrob', 'lmRob', 'rlm'), robustbase::lmrob--MM-type Estimators; robust::lmRob--automatically chooses an appropriate algorithm. one or more, 'lm' will always be included internally, even if not specified}

\item{type}{one of c('partial','spartial'), plot partial or semi-partial correlation}

\item{scale}{when having covariates, z transform residuals}

\item{rp}{show r (signed) and p values}

\item{rp.x}{r p values x position (0-1, relative to top left, big-->right), ignored if rp=FALSE.}

\item{rp.y}{r p values y position (0-1, relative to top left, big-->up), ignored if rp=FALSE.}

\item{se}{standard error of linear regression line}

\item{rp.size}{r p values font size, ignored if rp=FALSE}

\item{point.size}{if less point, oen can increase size}

\item{legend.size}{c(0,12) the first number 0 controls the legend title, 0=hide; the second number controls legend.key.size, legend.text}

\item{line.color}{only applicable when not auto varied, regression line color.}

\item{point.color}{only applicable when not auto varied}

\item{point.shape}{only applicable when not auto varied}

\item{point.alpha}{if overplot for points, one can reduce alpha}

\item{ylab}{y label. NULL}

\item{xlab}{x label. NULL}

\item{x.tick.number}{integer, how many ticks. can also set xlim, say xlim=c(1,10) via \code{...}}

\item{zlab}{z legend label/title, only applicable when there is z provided. NULL}

\item{legend.box}{box of legend, T or F}

\item{legend.position}{legend position 'none' (I think this is not natively supported, but works) "left", "right", "top", and "bottom"
\cr    c(x,y,two-element fractional numeric vector)     c(0,0) corresponds to the "bottom left" and c(1,1) corresponds to the "top right" position, but within the plot box.}

\item{legend.direction}{horizontal or vertical}

\item{hack}{T/F if more than 1 model specified, plot all in one plot}

\item{...}{other parameters passed to \code{\link[lattice]{xyplot}}. eg, xlim=c(1,10), layout=c(col,row)--a bit
weird. lattice orders panels from the bottom up. To get what you want either use as.table = TRUE or
reverse the ordering in index.cond. eg, as.table=F, index.cond=list(c(3,2,1))}
}
\value{
a lattice plot
}
\description{
scatter plot with lattice
}
\note{
specify auto through param colors, shapes
}
