% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.scatterplot2}
\alias{ez.scatterplot2}
\title{scatter plot with ggplot}
\usage{
ez.scatterplot2(df, cmd, rp.size = 5, rp.x = 0.25, rp.y = 0.99,
  line.color = "#BE1B22", point.color = "#0086B8", point.shape = 16,
  point.alpha = 0.95, point.size = 3, rug.size = 0.25, ylab = NULL,
  xlab = NULL, zlab = NULL, legend.position = "top",
  legend.direction = "horizontal", legend.box = T, legend.size = c(0, 10),
  rp = TRUE, se = TRUE, rug = FALSE, ellipse = FALSE)
}
\arguments{
\item{df}{data frame}

\item{cmd}{like "y~x+a+b", "y~x+a+b|z", "y~x+a+b||z", "y~x+a+b|||z"where y x are continous, z discrete (| one regression line, || multiple regression lines by levels of z--gives interaction p value), +a+b optional for covariates residualization
\cr y~x+a+b|||z an invisible plotlist returned, call ggmultiplot(plotlist)}

\item{rp.size}{r p values font size, ignored if rp=FALSE}

\item{rp.x}{r p values x position (0-1, relative to top left, big-->right), ignored if rp=FALSE. Internally, convert to geom_label(x,y) where x,y refers to actually x,y axis value}

\item{rp.y}{r p values y position (0-1, relative to top left, big-->bottom), ignored if rp=FALSE}

\item{line.color}{only applicable when y~x and y~x|z (ie, not auto varied with aes()), regression line color}

\item{point.color}{only applicable when y~x (ie, not auto varied with aes()). for auto ones, use scale_*_*}

\item{point.shape}{only applicable when y~x (ie, not auto varied with aes()). for auto ones, use scale_*_*}

\item{point.alpha}{if overplot for points, reduce alpha}

\item{point.size}{if less point, increase size}

\item{rug.size}{rug size}

\item{ylab}{y label NULL}

\item{xlab}{x label NULL}

\item{zlab}{z/fill/legend label, only applicable when there is z provided NULL}

\item{legend.position}{legend position 'top', 'bottom', 'left', 'right', 'none', c(x,y,two-element numeric vector)
\cr         c(0,0) corresponds to the "bottom left" and c(1,1) corresponds to the "top right" position.}

\item{legend.direction}{horizontal or vertical}

\item{legend.box}{box of legend, T or F}

\item{legend.size}{c(0,10) the first number 0 controls the legend title, 0=hide; the second number controls legend.key.size, legend.text}

\item{rp}{show r (signed) and p values}

\item{se}{standard error of linear regression line}

\item{rug}{marginal rug indicating univariate distribution}

\item{ellipse}{draw confidence ellipses, powered by stat_ellipse()}
}
\value{
a ggplot object (+theme_apa() to get apa format plot)
}
\description{
scatter plot with ggplot
}
