% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.se}
\alias{ez.se}
\title{standard error of mean}
\usage{
ez.se(x)
}
\arguments{
\item{x}{a vector}
}
\description{
na will be omitted before calculation, the formula is sqrt(var(x,na.rm=TRUE)/length(na.omit(x)))
}

