% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.selcol}
\alias{ez.selcol}
\title{select cols in a df}
\usage{
ez.selcol(df, col = NULL, ...)
}
\arguments{
\item{df}{df}

\item{col}{a string or a vector or NULL (NULL=all cols)
\cr 'c(sample_num,mother_num)' (quoted) or c("sample_num","mother_num") (not quoted)
\cr 1:4 (not quoted)
\cr 'col1:col3' (quoted)
\cr '-(ABCB1_c1236t:pgp_rs2032582)', '-c(neonate_admit_NICU,BDNF)' (quoted)}
}
\value{
returns vector of col names
}
\description{
select cols in a df, evaluated by sprintf('dplyr::select(df,%s, ...)',toString(col)
}
