% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.sprintf}
\alias{ez.sprintf}
\title{alias of \code{\link[glue]{glue}}}
\usage{
ez.sprintf(
  ...,
  .sep = "",
  .envir = parent.frame(),
  .open = "{",
  .close = "}",
  .na = "NA",
  .null = character(),
  .comment = "#",
  .literal = FALSE,
  .transformer = identity_transformer,
  .trim = TRUE
)
}
\description{
alias of \code{\link[glue]{glue}}
}
