% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.strcomp}
\alias{ez.strcomp}
\title{compare if s1 exact the same as s2, but ignore case, space, punctuation}
\usage{
ez.strcomp(s1, s2, icase = TRUE, ispace = TRUE, ipunctuation = TRUE)
}
\arguments{
\item{s}{string, supports vectorization}

\item{i}{ignore case, space, punctuation}
}
\value{
logic
}
\description{
compare if s1 exact the same as s2, but ignore case, space, punctuation
}
\note{
internally strip the space, punct, and convert case, then use "==" to compare
}
