% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.strreplace}
\alias{ez.strreplace}
\title{replace string or string vectors literally}
\usage{
ez.strreplace(s, expr, repstr, ignorecase = FALSE, once = FALSE)
}
\description{
support single string, vectors
\cr case sensitive by default (ignorecase=FALSE), wrapper of sub, gsub
}

