% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.subplot}
\alias{ez.subplot}
\title{subplot, wrapper of \code{\link{par}}}
\usage{
ez.subplot(n, m, ...)
}
\description{
subplot, wrapper of \code{\link{par}}
}
\examples{
subplot(n,m,...) divides the current figure into an n-by-m grid
A vector of the form c(n, m). Subsequent figures will be drawn in an n-by-m array on the device
by columns (mfcol), or rows (mfrow), respectively.
see more ?par
}
