% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.table}
\alias{ez.table}
\title{table xy}
\usage{
ez.table(df, x, y = NULL, prop.r = TRUE, prop.c = TRUE, prop.t = TRUE,
  digits = 2, max.width = 1, expected = FALSE, prop.chisq = FALSE,
  chisq = FALSE, fisher = FALSE, mcnemar = FALSE, resid = FALSE,
  sresid = FALSE, asresid = FALSE, missing.include = FALSE,
  format = "SPSS", dnn = NULL, ...)
}
\description{
(df, x, y) or (x, y), \code{\link[gmodels]{CrossTable}},  auto convert/reset factor
\cr ez.table2 input could be one, two, or more varibles/cols
}
\examples{
# dnn=c('row','col')
}
