% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.table2}
\alias{ez.table2}
\title{table xy}
\usage{
ez.table2(df, x, y = NULL, digits = 2, max.width = 1, expected = FALSE,
  prop.r = FALSE, prop.c = FALSE, prop.t = TRUE, prop.chisq = FALSE,
  chisq = FALSE, fisher = TRUE, mcnemar = FALSE, resid = FALSE,
  sresid = FALSE, asresid = FALSE, missing.include = FALSE,
  format = "SPSS", dnn = NULL, ...)
}
\description{
(df, x, y) or (x, y), \code{\link[gmodels]{CrossTable}},  auto convert to factor
\cr ez.table input could be one, two, or more varibles/cols
}
\examples{
# dnn=c('row','col')
}

