% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.time}
\alias{ez.time}
\title{convert to time}
\usage{
ez.time(x, format = "numeric")
}
\arguments{
\item{x}{a vector of number}

\item{format}{string, 'numeric' or 'times'}
}
\value{
returns a vector of number (class numeric) or time (class times)
}
\description{
convert from a number (read from external files, eg, spss date 64080) to time (17:48:00)/number(0.7416667)
}
\seealso{
\code{\link{ez.date}} \code{\link{ez.is.date}} \code{\link{ez.is.date.convertible}} \code{\link{ez.age}} \code{\link{ez.time}}
}
