% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.transpose}
\alias{ez.transpose}
\title{alias of \code{\link[data.table]{transpose}}}
\usage{
ez.transpose(l, fill = NA, ignore.empty = FALSE)
}
\description{
alias of \code{\link[data.table]{transpose}}
}
