% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.trim}
\alias{ez.trim}
\title{Merge multiple spaces to single space in the middle, and remove trailing/leading spaces}
\usage{
ez.trim(s, how = 4)
}
\arguments{
\item{s}{a string}

\item{how}{a num 1=left only; 2=right only; 3=left and right; 4 (default)=left and right and merge middle}
}
\value{
For portability, whitespace is taken as the character class [ \t\r\n] (space, horizontal tab, line feed, carriage return).
}
\description{
underlying function is \code{\link{gsub}} with regular expression
}
\seealso{
\code{\link{trimws}}
}
