% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.type}
\alias{ez.type}
\title{show the content of a file in read-only mode, wrapper of wrapper of \code{\link{file.show}}}
\usage{
ez.type(path)
}
\description{
show the content of a file in read-only mode, wrapper of wrapper of \code{\link{file.show}}
}

