% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.unload}
\alias{ez.unload}
\title{unload a package, wrapper of detach(pkg, unload=TRUE, character.only = TRUE)}
\usage{
ez.unload(pkg)
}
\arguments{
\item{pkg}{pkg name in string}
}
\description{
unload a package, wrapper of detach(pkg, unload=TRUE, character.only = TRUE)
}
