% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.values.set}
\alias{ez.values.set}
\title{set value labels, wrapper of \code{\link{sjmisc_set_labels}}}
\usage{
ez.values.set(x, valuelabels, force.labels = FALSE, force.values = FALSE,
  ...)
}
\value{
returns a new changed df
}
\description{
set value labels, wrapper of \code{\link{sjmisc_set_labels}}
}
\examples{
#1 4 5 9 do not have to all appear in x
# notice the particular order and symbol: "strongly agree" <- 1
set_labels(x, c("strongly agree"=1,
               "totally disagree"=4,
               "refused"=5,
               "missing"=9))
}
\seealso{
\code{\link[tidyr]{gather}}, \code{\link[tidyr]{spread}}, \code{\link[tidyr]{separate}}, \code{\link[tidyr]{unite}}
\cr \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}
\cr \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{arrange}}
\cr \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{mutate}}
\cr \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{right_join}}, \code{\link[dplyr]{inner_join}}, \code{\link[dplyr]{full_join}}, \code{\link[dplyr]{semi_join}}, \code{\link[dplyr]{anti_join}}
\cr \code{\link[dplyr]{intersect}}, \code{\link[dplyr]{union}}, \code{\link[dplyr]{setdiff}}
\cr \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{bind_cols}}

Other data transformation functions: \code{\link{ez.2char}},
  \code{\link{ez.2factor}}, \code{\link{ez.2label}},
  \code{\link{ez.2long}}, \code{\link{ez.2value}},
  \code{\link{ez.2wide}}, \code{\link{ez.compute}},
  \code{\link{ez.label.get}}, \code{\link{ez.label.set}},
  \code{\link{ez.leftjoin}}, \code{\link{ez.move}},
  \code{\link{ez.num}}, \code{\link{ez.recode}},
  \code{\link{ez.recols}}, \code{\link{ez.recol}},
  \code{\link{ez.replace}}, \code{\link{ez.select}},
  \code{\link{ez.sort}}, \code{\link{ez.split}},
  \code{\link{ez.str}}, \code{\link{ez.unique}},
  \code{\link{ez.values.get}}
}
