% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.vi}
\alias{ez.vi}
\title{print sorted uniques of a df col or a vector (NA last) and other information}
\usage{
ez.vi(x, printn = 35, printcn = 600, order = "as")
}
\arguments{
\item{printn}{print first n and last n (useful for loooong vector). If 2n >= total length, print all. Inf=all}

\item{printcn}{Counts/Levels print n}

\item{order}{vector order for printing out, 'as','az','za'}
}
\description{
vi (view everything print out), vv (view format vector), vx (view excel), View (built-in). print sorted uniques of a df col or a vector (NA last) and other information
}
