% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.view}
\alias{ez.view}
\title{show information about a data frame or similar object (like spss variable view)}
\usage{
ez.view(x, show.frq = T, show.prc = T, sort.by.name = F, ...)
}
\description{
wrapper of \code{\link[sjPlot]{view_df}}; can make the html bigger by openning in internet browser
}
\seealso{
\code{\link{ez.info}} or \code{\link{ez.show}}
}

