% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.vv}
\alias{ez.vv}
\title{format a vector for easy manual copy/processing.}
\usage{
ez.vv(vec, printn = Inf, order = "as", quote = NULL,
  print2screen = TRUE)
}
\arguments{
\item{vec}{a vector}

\item{printn}{print first n and last n (useful for loooong vector). If 2n >= total length, print all. Inf=all}

\item{order}{vector order for printing out, 'as','az','za'}

\item{quote}{TRUE/FALSE, whether add a quote around each element (switch for string or number). NULL = auto (F for numeric, T otherwise)}
}
\value{
nothing, only print out. 
\cr By default in R when you type a variable name, you get [1] "rs171440fwd" "rs1800497fwd"
\cr now with this function you get 'rs171440fwd','rs1800497fwd','rs180043'
}
\description{
vi (view everything print out), vv (view format vector), vx (view excel), View (built-in).
}
\seealso{
\code{\link{ez.print}} \code{\link{ez.pprint}}
}
