% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.vx}
\alias{ez.vx}
\title{view the overview of a data frame or similar object (like spss variable view, but with much more information)}
\usage{
ez.vx(df, temp = NULL, id = NULL, file = NULL, width = 300,
  incomparables = FALSE, debug = NULL, ...)
}
\arguments{
\item{df}{a data frame}

\item{temp}{when file is NULL, the name prefix for the temp excel file. If prefix not provided through this param, auto generate one}

\item{id}{a single col name in string or number (eg, 'age' or 3), that serves as (potentially unique) id, except which duplicated rows will be checked against. If NULL, rownames() will be auto used}

\item{file}{a file name, if NULL, a temp generated, will save more detailed variable information to an excel file}

\item{width}{controls if too many factor levels to print, eg 300. NULL=unlimited}
}
\value{
returns a list $row, $col, $dat (input data frame), $pth (file path)
}
\description{
vc (c(vv)), vi (view everything print out), vv (view format vector), vx (view excel), View (built-in).
}
\note{
when file=NULL (which stores to a temp file), debug=NULL, use getOption('debug') which is TRUE if not set up
\cr when file=NULL, debug provided, overwrites getOption('debug')
\cr when file provided, any debug is ignored
\cr Bottom line: file > param debug > option debug
\cr Updated: as of Thu, Nov 30 2017, not any more a wrapper of \code{\link[sjPlot]{view_df}}; can make the html bigger by openning in internet browser
}
