% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.warn}
\alias{ez.warn}
\title{get/set warn error mode}
\usage{
ez.warn(warnErrorMode = NULL, print2scr = TRUE)
}
\arguments{
\item{warnErrorMode}{1/0/NULL. if 1/0, set mode; if NULL, get mode, wrapper of getOption('warn')}
}
\value{
if 1/0 passed in, return invisible NULL
\cr if NULL passed in, return the current status 1/0/NULL (corresponding to on/off/Not set)
}
\description{
get/set warn error mode
}
