% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.wherena}
\alias{ez.wherena}
\title{visualize where all the NAs of a dataframe are}
\usage{
ez.wherena(df, id = NULL, color = "red", angle = 270, basesize = 9,
  xsize = 1, ysize = 1)
}
\arguments{
\item{df}{data frame in wide format}

\item{id}{a column name as id, will be shown as y axis, quoted "", eg, subject ID; if id not given, internally add row# as id}

\item{color}{color of missing values, eg, "red"}

\item{angle}{the x axis label angle, default=270 (vertical), suggests 330 if label is not too long}

\item{basesize}{base font size}

\item{xsize}{x axis label font relative size}

\item{ysize}{y axis label font relative size}
}
\value{
a ggplot object (+theme_apa() to get apa format plot)
}
\description{
you might get In apply(df, 2, as.numeric) : NAs introduced by coercion
\cr but it does not matter, the purpose here is to change everything to numeric
\cr so that 0 or 1 can be assigned for plotting
}

