% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.whos}
\alias{ez.whos}
\title{alias of \code{\link{sessionInfo}}, \code{\link{ez.who}}}
\usage{
ez.whos(package = NULL)
}
\description{
alias of \code{\link{sessionInfo}}, \code{\link{ez.who}}
}
\seealso{
\code{\link{objects}}
}
