% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.writes}
\alias{ez.writes}
\title{alias of \code{\link{sjmisc_write_spss}}, \code{\link{ez.saves}}}
\usage{
ez.writes(x, path, enc.to.utf8 = TRUE)
}
\description{
potentially keep variable labels and value labels
}

