% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.x}
\alias{ez.x}
\title{view df in excel (saved to a temp file, auto delete the temp file)}
\usage{
ez.x(..., temp = NULL, debug = NULL)
}
\arguments{
\item{...}{one or more df. df1, df2 becomes Sheet1, Sheet2 etc}
}
\value{
returns nothing
}
\description{
view df in excel (saved to a temp file, auto delete the temp file)
}
\note{
when debug provided, overwrites getOption('debug')
}
