% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.xl}
\alias{ez.xl}
\title{view df in excel (saved to a temp file, auto delete the temp file)}
\usage{
ez.xl(df, temp = NULL, debug = NULL, label = TRUE)
}
\arguments{
\item{temp}{when file is NULL, the name prefix for the temp excel file. If prefix not provided through this param, auto generate one}

\item{label}{T/F. call ez.2label()}
}
\value{
returns nothing
}
\description{
view df in excel (saved to a temp file, auto delete the temp file)
}
\note{
when debug provided, overwrites getOption('debug'). A trick to use: let View=ez.xl, then you can use GUI to click.
}
