% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.xlcolconv}
\alias{ez.xlcolconv}
\title{Excel-style column name converter: number from or to letters}
\usage{
ez.xlcolconv(col)
}
\arguments{
\item{col}{number or letters, input a number (27), output letters ('AA'), vice versa}
}
\description{
Excel-style column name converter: number from or to letters. Vectorized.
}
