% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.zresidize}
\alias{ez.zresidize}
\title{residualize variable, returning it changed in-place (i.e under its original column name)}
\usage{
ez.zresidize(data, var, covar, model = "lm", scale = TRUE, ...)
}
\arguments{
\item{var}{dependent var}

\item{covar}{covariates}

\item{model}{'lm', 'lmrob' (MM-type Estimators, robustbase), 'lmRob' (automatically chooses, robust), 'rlm' (MASS)}

\item{scale}{unstandarize or standardize residual}

\item{...}{additional param passed to the specified model}
}
\value{
dataframe with var residualized in place (i.e under its original column name). If covar have NA, then that row for var will be NA.
}
\description{
residualize variable, returning it changed in-place (i.e under its original column name)
}
