% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R ez.R
\name{fast_alpha}
\alias{fast_alpha}
\title{Fast Alpha}
\usage{
fast_alpha(x)
}
\arguments{
\item{x}{Vector of type character}
}
\description{
Quick function to remove non-alphanumeric character
}
\examples{
x = c("!"#$\%&/String")
fast_alpha(x)
}
