% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R ez.R
\name{get_total}
\alias{get_total}
\title{Get the total of numeric variables}
\usage{
get_total(df)
}
\arguments{
\item{df}{Object of type dataframe

This function calculates the total of all numerical variables, there must be at least one variable of any type other than numerical.}
}
\description{
Get the total of numeric variables
}
\examples{
get_total(iris)
}
