% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggpass}
\alias{ggpass}
\title{df %>% ggplot1() %>% ggpass() %>% ggplot2()}
\usage{
ggpass(pp)
}
\description{
df %>% ggplot1() %>% ggpass() %>% ggplot2()
}
\note{
Assuming pp has $df (preferred to $data), or $data. Make sure $df/$data is what you intend to plot
}

