% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{matlabcolor}
\alias{matlabcolor}
\alias{matlabcolor2}
\title{change plot continous color to matlab like}
\usage{
matlabcolor(n = 100)

matlabcolor2(n = 100)
}
\arguments{
\item{n}{how many colors, e.g., 100 (default)}
}
\description{
change plot continous color to matlab like
}
\examples{
p + matlabcolor(), p + matlabcolor2()
}
