% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R ez.R
\name{melt_date}
\alias{melt_date}
\title{melt data frame by date}
\usage{
melt_date(data, start, end, id)
}
\arguments{
\item{data}{Object of type dataframe}

\item{start}{date type value}

\item{end}{date type value}

\item{id}{Character value that will be used as id to melt the data frame from start to end by dates}
}
\description{
This function returns a data frame in long format, where the dates are completed from start to end by some id.
}
\examples{
data_example = data.frame(id = c(1, 2),
start = as.Date(c("2021-01-01", "2021-01-10")),
end = as.Date(c("2021-01-10", "2021-01-15")))
melt_date(data = data_example, start = start, end = end, id = id)
}
