% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R ez.R
\name{recent_files}
\alias{recent_files}
\title{Most recent file}
\usage{
recent_files()
}
\description{
Returns the name of the most recent file in the directory
}
\examples{
getwd() # Get of the directory
recent_files()
}
