% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{data}
\name{RMN}
\alias{RMN}
\title{creat shorthand for Times New Roman cross-platform}
\format{
An object of class \code{character} of length 1.
}
\usage{
RMN
}
\description{
creat shorthand for Times New Roman cross-platform
}
\keyword{datasets}
