% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{round2}
\alias{round2}
\title{round2 always round half up (R's native round/sprintf may round .055 -> .05)}
\usage{
round2(x, n)
}
\description{
round2 always round half up (R's native round/sprintf may round .055 -> .05)
}
