% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R ez.R
\name{seq_dates}
\alias{seq_dates}
\title{Date sequence}
\usage{
seq_dates(date1, date2)
}
\arguments{
\item{date1}{Vector of type character or date}

\item{date2}{Vector of type character or date}
}
\description{
This function returns a sequence of dates.
}
\examples{
seq_dates("2021-01-01","2021-01-31")

seq_dates("20210101","20210131")

}
