% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{show.color}
\alias{show.color}
\title{show color}
\usage{
show.color(v, name, label = T)
}
\arguments{
\item{v}{color vector to show, if not provided, show all colors in ez.palette}

\item{name}{display name for the color v (optional)}

\item{label}{T/F display actual color code for each v}
}
\description{
show color
}
\examples{
show.color(ez.palette('Zhu'),'Current'), show.color("#89DAF7")
}
