% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjmisc18.R
\name{sjmisc_get_label}
\alias{sjmisc_get_label}
\title{sjmisc 1.8 hack}
\usage{
sjmisc_get_label(x, def.value = NULL)
}
\arguments{
\item{x}{a df: (efc), or a single var: (efc$e42dep)}

\item{def.value}{the value returned when no label attribute for a single var. if x is a df, def.value is not effective, instead returns '' if no label for a col in df}
}
\description{
sjmisc 1.8 hack, variable labels
}
