% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjmisc18.R
\name{sjmisc_get_labels}
\alias{sjmisc_get_labels}
\title{sjmisc 1.8 hack}
\usage{
sjmisc_get_labels(x, attr.only = FALSE, include.values = NULL,
  include.non.labelled = FALSE)
}
\description{
sjmisc 1.8 hack, value labels
}
\note{
include.values, how values are returned: NULL = unnamed vector; as named vector = "as.name" or "n"; "as.prefix" or "p", c("[1] Caucasian", "[3] Asian")
\cr attr.only = T, only search attributes, T desired (if F, also search factor levels, character strings)
\cr include.non.labelled, if T, values without labels will also be included in the returned labels
\cr returns NULL if no value labels attribute
}
