% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjmisc18.R
\name{sjmisc_read_spss}
\alias{sjmisc_read_spss}
\title{sjmisc 1.8 hack}
\usage{
sjmisc_read_spss(path, enc = NA, attach.var.labels = FALSE,
  atomic.to.fac = FALSE, keep.na = FALSE, option = "haven")
}
\description{
sjmisc 1.8 hack
}
