% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjmisc18.R
\name{sjmisc_set_label}
\alias{sjmisc_set_label}
\title{sjmisc 1.8 hack}
\usage{
sjmisc_set_label(x, lab, attr.string = NULL)
}
\arguments{
\item{df}{data frame}

\item{varname}{variable name with quote ""}

\item{label}{explanatory string, if "" or NULL, essentially clear label attribute}
}
\value{
returns a new changed df
}
\description{
sjmisc 1.8 hack, variable labels
}
