% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjmisc18.R
\name{sjmisc_set_labels}
\alias{sjmisc_set_labels}
\title{sjmisc 1.8 hack}
\usage{
sjmisc_set_labels(x, labels, force.labels = FALSE, force.values = TRUE)
}
\description{
sjmisc 1.8 hack, value labels
}
\note{
The usage of the original function is a bit confusing. Here, I summarize:
\cr 1) Always use a named vector for valuelabels, in which case force.labels and force.values will be ignored
\cr valuelabels:
\cr 1 4 5 9 do not have to all appear in x, that is, you can assign more labels than values (or vice versa, assign less labels)
\cr notice the particular order and symbol: "strongly agree" <- 1
\cr sjmisc_set_labels(x, c("strongly agree"=1,"totally disagree"=4,"refused"=5,"missing"=9))
\cr This format is flipped from the results directly from ez.getlabels
\cr
\cr 2) when valuelabels="" or NULL, essentially clear value labels attribute
\cr
\cr 3) if no exisiting attr, using haven style: labels
\cr
\cr 4) though x could be df, I generally use x as var
}
