% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_ezr.R
\name{start_ezr}
\alias{start_ezr}
\title{Start ezr}
\usage{
start_ezr(
  data_for_ezr = NULL,
  sigfig = 3,
  select_list_max = 1e+05,
  ezr_saved_analysis_file_name = "ezr_saved_analysis.csv",
  ezr_run_analysis_file_name = "ezr_run_analysis.csv"
)
}
\arguments{
\item{data_for_ezr}{a data object (a data frame or a data.table)}

\item{sigfig}{number of significant digits to round to}

\item{select_list_max}{maximum number of variable names to display
for dropdown menus}

\item{ezr_saved_analysis_file_name}{name of the .csv file on which
saved analysis will be recorded (default = "ezr_saved_analysis.csv")}

\item{ezr_run_analysis_file_name}{name of the .csv file on which
all conducted analyses will be recorded
(default = "ezr_run_analysis.csv")}
}
\value{
There will be no output from this function. Rather, the ezr
program will open on a new tab or window of the local machine's
web browser
}
\description{
Starts the ezr program on the local machine
}
\examples{
if (interactive()) {start_ezr(data = mtcars)}
}
