% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R ez.R
\name{style_elegant_xlsx}
\alias{style_elegant_xlsx}
\title{Style elegant xlsx}
\usage{
style_elegant_xlsx(sheet, name, data)
}
\arguments{
\item{sheet}{Sheet name}

\item{name}{file name}

\item{data}{Database name}
}
\description{
This function expands a dataframe to a desired dimension.
}
\examples{
style_elegant_xlsx("example","iris.xlsx",iris)
}
