% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_apa}
\alias{theme_apa}
\title{ggplot2 Theme for APA Publications}
\usage{
theme_apa(plot.box = FALSE, lab.size = 18, text.size = 16)
}
\arguments{
\item{plot.box}{logical.  If \code{TRUE} a full box surrounds the plot area.  If \code{FALSE} only the x and y axis are shown.}

\item{lab.size}{default 18}

\item{text.size}{default 16}
}
\description{
ggplot2 Theme for APA Publications

A ggplot2 theme with no background and Times New Roman font (legend size not affected).
}
\note{
In order for R (at least on Mac) to recognize Times New Roman font, \href{https://github.com/wch/extrafont/}{extrafont} required
}
\seealso{
\code{\link[ggplot2]{theme}} \code{\link{theme_apa_nosize}}
}
\author{
Jerry modified from \href{https://github.com/trinker/plotflow}{trinker/plotflow}
}
