% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_apa_nosize}
\alias{theme_apa_nosize}
\title{ggplot2 Theme for APA Publications}
\usage{
theme_apa_nosize(plot.box = FALSE)
}
\arguments{
\item{plot.box}{logical.  If \code{TRUE} a full box surrounds the plot area.  If \code{FALSE} only the x and y axis are shown.}
}
\description{
A ggplot2 theme with no background and Times New Roman font (do not change font size).
}
\note{
In order for R (at least on Mac) to recognize Times New Roman font, \href{https://github.com/wch/extrafont/}{extrafont} required
}
\seealso{
\code{\link[ggplot2]{theme}} \code{\link{theme_apa}}
}
\author{
Jerry modified from \href{https://github.com/trinker/plotflow}{trinker/plotflow}
}
