% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R ez.R
\name{to_num}
\alias{to_num}
\title{Convert to number}
\usage{
to_num(x)
}
\arguments{
\item{x}{Vector of numeric type or character}
}
\description{
This function returns a numeric value or vector, which can be integer or decimal, containing non-numeric characters that prevent transformation.
}
\examples{
x = c("1","2","3.1","4,2","5,0")
to_num(x)
}
