% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R ez.R
\name{train_test}
\alias{train_test}
\title{Create training and test df}
\usage{
train_test(df, p)
}
\arguments{
\item{df}{Object of type dataframe}

\item{p}{Value between 0 and 1 that contains the probability of separating the df}
}
\description{
This function returns a list with two df, which are separated with probability p.
}
\examples{

train_test(iris,0.8)
}
