% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{z.2long}
\alias{z.2long}
\title{reconstruct to long format}
\usage{
z.2long(df, indexname, index, measurename = NULL, measure = NULL,
  drop = NULL, ...)
}
\arguments{
\item{indexname}{variable name for timing/repetition/index variable, such as "session"}

\item{index}{level name for each timing/repetition/index point, such as c("1,2"), c("Pre, Post")}

\item{measurename}{variable name for the measurement, such as "BDI"}

\item{measure}{column names that are the repeated measures, such as c("BDI_Pre","BDI_Post")}

\item{drop}{variables to drop before reshaping}
}
\value{
returns a new df, label attributes seem to be intact.
}
\note{
refer to my spss syntax 'Time(2) | Measure1(Pre1 Post1) | Measure2(Pre2 Post2) +/- Subject'
}
\examples{
df <- data.frame(
    id = 1:10,
    time = as.Date('2009-01-01') + 0:9,
    Q3.2.1. = rnorm(10, 0, 1),
    Q3.2.2. = rnorm(10, 0, 1),
    Q3.2.3. = rnorm(10, 0, 1),
    Q3.3.1. = rnorm(10, 0, 1),
    Q3.3.2. = rnorm(10, 0, 1),
    Q3.3.3. = rnorm(10, 0, 1)
)
z.2long(df,
       "loop_number",c(1:3),
       c('Q3.2','Q3.3'),
       list(c("Q3.2.1.","Q3.2.2.","Q3.2.3."),c("Q3.3.1.","Q3.3.2.","Q3.3.3.")))
}
\seealso{
\code{\link[tidyr]{gather}}, \code{\link[tidyr]{spread}}, \code{\link[tidyr]{separate}}, \code{\link[tidyr]{unite}}
\cr \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}
\cr \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{arrange}}
\cr \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{mutate}}
\cr \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{right_join}}, \code{\link[dplyr]{inner_join}}, \code{\link[dplyr]{full_join}}, \code{\link[dplyr]{semi_join}}, \code{\link[dplyr]{anti_join}}
\cr \code{\link[dplyr]{intersect}}, \code{\link[dplyr]{union}}, \code{\link[dplyr]{setdiff}}
\cr \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{bind_cols}}

Other data.transformation.functions: \code{\link{z.2factor}},
  \code{\link{z.2label}}, \code{\link{z.2lower}},
  \code{\link{z.2value}}, \code{\link{z.2wide}},
  \code{\link{z.compute}}, \code{\link{z.delete}},
  \code{\link{z.del}}, \code{\link{z.label.get}},
  \code{\link{z.label.set}}, \code{\link{z.leftjoin}},
  \code{\link{z.move}}, \code{\link{z.newcol}},
  \code{\link{z.recode2}}, \code{\link{z.recode}},
  \code{\link{z.recols}}, \code{\link{z.recol}},
  \code{\link{z.rename}}, \code{\link{z.rncols}},
  \code{\link{z.rncol}}, \code{\link{z.select}},
  \code{\link{z.sort}}, \code{\link{z.split}},
  \code{\link{z.unique}}, \code{\link{z.values.get}},
  \code{\link{z.values.set}}
}

