% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{z.2value}
\alias{z.2value}
\title{wrapper of \code{\link[sjmisc]{to_value}}}
\usage{
z.2value(x, start.at = NULL, keep.labels = TRUE, ...)
}
\arguments{
\item{x}{factor or a data frame with factors. May also be a character vector.}

\item{start.at}{starting index, i.e. the lowest numeric value of the variable's value range. By default, this argument is NULL, hence the lowest value of the returned numeric variable corresponds to the lowest factor level (if factor is numeric) or to 1 (if factor levels are not numeric).}
}
\value{
returns a numeric variable
}
\description{
continous number<--categorical string/number
}
\details{
opposite of \code{\link{z.2factor}}, \code{\link{z.2label}}
}
\examples{
# starting at 1
dummy <- factor(c("D", "F", "H"))
to_value(dummy)
# [1] 1 2 3
# attr(,"labels")
# D F H
# 1 2 3
}
\seealso{
\code{\link[tidyr]{gather}}, \code{\link[tidyr]{spread}}, \code{\link[tidyr]{separate}}, \code{\link[tidyr]{unite}}
\cr \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}
\cr \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{arrange}}
\cr \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{mutate}}
\cr \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{right_join}}, \code{\link[dplyr]{inner_join}}, \code{\link[dplyr]{full_join}}, \code{\link[dplyr]{semi_join}}, \code{\link[dplyr]{anti_join}}
\cr \code{\link[dplyr]{intersect}}, \code{\link[dplyr]{union}}, \code{\link[dplyr]{setdiff}}
\cr \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{bind_cols}}

Other data.transformation.functions: \code{\link{z.2factor}},
  \code{\link{z.2label}}, \code{\link{z.compute}},
  \code{\link{z.delete}}, \code{\link{z.del}},
  \code{\link{z.label.get}}, \code{\link{z.label.set}},
  \code{\link{z.label}}, \code{\link{z.leftjoin}},
  \code{\link{z.move}}, \code{\link{z.newcol}},
  \code{\link{z.recode2}}, \code{\link{z.recode}},
  \code{\link{z.recols}}, \code{\link{z.recol}},
  \code{\link{z.rename}}, \code{\link{z.rncols}},
  \code{\link{z.rncol}}, \code{\link{z.select}},
  \code{\link{z.sort}}, \code{\link{z.split}},
  \code{\link{z.unique}}, \code{\link{z.values.get}},
  \code{\link{z.values.set}}
}

