% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{z.2wide}
\alias{z.2wide}
\title{reconstruct to wide format}
\usage{
z.2wide(df, id, indexname, measure = NULL, drop = NULL, ...)
}
\arguments{
\item{indexname}{variable name for timing/repetition/index variable, such as "session"}

\item{measure}{column names that are the repeated measures, such as c("BDI_Pre","BDI_Post")}

\item{drop}{variables to drop before reshaping}
}
\value{
returns a new df. Because of the structural change, lable attributes would be lost and can NOT be copied with \code{\link[sjmisc]{copy_labels}}--not a subset.
}
\note{
refer to my spss syntax 'SUBJID * Time [School] - Measure2'
}
\examples{
set.seed(10)
df <- data_frame(
    Person = rep(c("greg", "sally", "sue"), each=2),
    Time = rep(c("Pre", "Post"), 3),
    Score1 = round(rnorm(6, mean = 80, sd=4), 0),
    Score2 = round(jitter(Score1, 15), 0),
    Score3 = 5 + (Score1 + Score2)/2
)
df = data.frame(df)
z.2wide(df,
        "Person",
        "Time",
        c("Score1","Score2","Score3"))
}
\seealso{
\code{\link[tidyr]{gather}}, \code{\link[tidyr]{spread}}, \code{\link[tidyr]{separate}}, \code{\link[tidyr]{unite}}
\cr \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}
\cr \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{arrange}}
\cr \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{mutate}}
\cr \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{right_join}}, \code{\link[dplyr]{inner_join}}, \code{\link[dplyr]{full_join}}, \code{\link[dplyr]{semi_join}}, \code{\link[dplyr]{anti_join}}
\cr \code{\link[dplyr]{intersect}}, \code{\link[dplyr]{union}}, \code{\link[dplyr]{setdiff}}
\cr \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{bind_cols}}

Other data.transformation.functions: \code{\link{z.2factor}},
  \code{\link{z.2label}}, \code{\link{z.2long}},
  \code{\link{z.2lower}}, \code{\link{z.2value}},
  \code{\link{z.compute}}, \code{\link{z.delete}},
  \code{\link{z.del}}, \code{\link{z.label.get}},
  \code{\link{z.label.set}}, \code{\link{z.leftjoin}},
  \code{\link{z.move}}, \code{\link{z.newcol}},
  \code{\link{z.recode2}}, \code{\link{z.recode}},
  \code{\link{z.recols}}, \code{\link{z.recol}},
  \code{\link{z.rename}}, \code{\link{z.rncols}},
  \code{\link{z.rncol}}, \code{\link{z.select}},
  \code{\link{z.sort}}, \code{\link{z.split}},
  \code{\link{z.unique}}, \code{\link{z.values.get}},
  \code{\link{z.values.set}}
}

