% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{z.clean}
\alias{z.clean}
\title{clear, clean  console, workspace, plot}
\usage{
z.clean(area = 0)
}
\arguments{
\item{area}{0 all, 1 console only, 2 workspace only, 3 plot only}
}
\description{
clear, clean  console, workspace, plot
}

