% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{z.clear}
\alias{z.clear}
\title{clear, clean  console, workspace, plot}
\usage{
z.clear(area = 0)
}
\arguments{
\item{area}{0 all, 
1 console only
2 workspace only
3 plot only
'var' particular var
c('var1','var2') particular vars}
}
\description{
clear, clean  console, workspace, plot
}

