% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc.R
\name{z.cnames}
\alias{z.cnames}
\title{all column names, alias of \code{\link{names}}, there are also names(), colnames(), rownames(), row.names() but no col.names()}
\usage{
z.cnames()
}
\description{
all column names, alias of \code{\link{names}}, there are also names(), colnames(), rownames(), row.names() but no col.names()
}

