% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{z.corrmap}
\alias{z.corrmap}
\title{plot a correlation matrix map}
\usage{
z.corrmap(df, corr.type = "pearson", sig.level = 0.05, insig = "blank",
  method = "color", tl.col = "black", tl.cex = 0.4, col = NULL, ...)
}
\arguments{
\item{df}{data frame in wide format, should be all numeric}

\item{corr.type}{"pearson" or "spearman"}

\item{sig.level}{sig.level}

\item{insig}{how to treat insig values, one of "pch","p-value","blank", "n"}

\item{...}{see \code{\link[corrplot]{corrplot}} for more parameters}
}
\description{
a wrapper of \code{\link[corrplot]{corrplot}}; the correlation and p values are calculated with \code{\link[Hmisc]{rcorr}}
}

