% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc.R
\name{z.del}
\alias{z.del}
\title{delete one or many cols, may use \code{\link[dplyr]{select}} instead}
\usage{
z.del(df, del)
}
\arguments{
\item{del}{sth like 'Month' or c('Month','Day')}
}
\value{
returns a new df, old one does not change
}
\description{
delete one or many cols, may use \code{\link[dplyr]{select}} instead
}

