% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{z.describe}
\alias{z.describe}
\title{plot a customized boxplot with jittered stripplot, violin, and mean}
\usage{
z.describe(df, cmd)
}
\arguments{
\item{df}{data frame in long format}

\item{cmd}{like "y|x z" or "y|x" where y is continous, x z are discrete}
}
\value{
a ggplot object (+theme_apa() to get apa format plot)
}
\description{
plot a customized boxplot with jittered stripplot, violin, and mean
}

