% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{z.execute}
\alias{z.execute}
\title{alias of \code{\link{system}}}
\usage{
z.execute(command, intern = FALSE, ignore.stdout = FALSE,
  ignore.stderr = FALSE, wait = TRUE, input = NULL,
  show.output.on.console = TRUE, minimized = FALSE, invisible = TRUE)
}
\description{
alias of \code{\link{system}}
}

