% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{z.export}
\alias{z.export}
\title{wrapper of \code{\link{dev.copy2pdf}}}
\usage{
z.export(filename = "RPlot.pdf", pdf.width = 8.5, pdf.height = 11, ...)
}
\description{
wrapper of \code{\link{dev.copy2pdf}}
}
\examples{
A4:     width 7(inches) height = 5
Letter: 8.5 x 11
}
\seealso{
\code{\link{pdf}}
}

