% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc.R
\name{z.find}
\alias{z.find}
\title{similar to matlab find, find non-zero in a vector}
\usage{
z.find(v)
}
\arguments{
\item{v}{logical input}
}
\value{
returns index vector
}
\description{
similar to matlab find, find non-zero in a vector
}
\examples{
z.find(a>1)
}

