% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{z.frame}
\alias{z.frame}
\title{alias of data.frame}
\usage{
z.frame(..., row.names = NULL, check.rows = FALSE, check.names = TRUE,
  stringsAsFactors = default.stringsAsFactors())
}
\description{
alias of data.frame
}
\examples{
sx = c("F", "F", "F", "M", "M", "M")
ht = c(69, 64, 67, 68, 72, 71)
wt = c(148, 132, 142, 149, 167, 165)
people = data.frame(sx, ht, wt)
}

