% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{z.github}
\alias{z.github}
\title{alias of \code{\link{devtools::install_github}}}
\usage{
z.github(repo, username = NULL, ref = "master", subdir = NULL,
  auth_token = github_pat(quiet), host = "api.github.com", force = FALSE,
  quiet = FALSE, ...)
}
\description{
alias of \code{\link{devtools::install_github}}
}

