% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{z.heatmap}
\alias{z.heatmap}
\title{plot a heatmap with values shown}
\usage{
z.heatmap(df, id, show.values = T)
}
\arguments{
\item{df}{data frame in wide format, like a correlation matrix}

\item{id}{a column name as id, will be shown as x axis, quoted ""}

\item{show.values}{whether to show values, default TRUE}
}
\value{
a ggplot object (+theme_apa() to get apa format plot)
}
\description{
plot a heatmap with values shown
}

