% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc.R
\name{z.import}
\alias{z.import}
\title{alias of \code{\link{library}}}
\usage{
z.import(package, help, pos = 2, lib.loc = NULL, character.only = FALSE,
  logical.return = FALSE, warn.conflicts = TRUE, quietly = FALSE,
  verbose = getOption("verbose"))
}
\description{
alias of \code{\link{library}}
}

