% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{z.join}
\alias{z.join}
\title{mimic python join, supports vectorization, wrapper of \code{\link{paste}}}
\usage{
z.join(sep = "", ...)
}
\arguments{
\item{sep}{default is nothing}
}
\description{
mimic python join, supports vectorization, wrapper of \code{\link{paste}}
}
\examples{
z.join('','a',' x ','b') # "a x b"
}

