% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{z.joinpath}
\alias{z.joinpath}
\title{alias of \code{\link{file.path}}}
\usage{
z.joinpath(..., fsep = .Platform$file.sep)
}
\description{
alias of \code{\link{file.path}}
}
\examples{
paste(..., sep=.Platform$file.sep)
}

