% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{z.log}
\alias{z.log}
\title{Prints output to both terminal and a file (log.txt) globally.}
\usage{
z.log(file = "log.txt", mode = "a", status = TRUE)
}
\arguments{
\item{mode}{a=append; w=overwrite}
}
\description{
Prints output to both terminal and a file (log.txt) globally.
}

