% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{z.mv}
\alias{z.mv}
\title{move}
\usage{
z.mv(from, to)
}
\description{
move
}
\examples{
support c('a.txt','b.txt')
to parent folder does not have to exist already
z.mv('a.txt','folder'), z.mv('a.txt','folder/a.txt'), z.mv('a.txt','folder/b.txt')
z.mv('a','b')-->get b/a, b now has a as subfolder, regardless of b exists or not
                 use z.rn('a','b') to change name a->b
}

