% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{z.na.remove}
\alias{z.na.remove}
\title{remove rows that have an NA anywhere/somewhere or a certain number of NAs}
\usage{
z.na.remove(df, col = NULL, n = 0)
}
\arguments{
\item{df}{a data frame}

\item{col}{restrict to the columns where you would like to search for NA; eg, 3, c(3), 2:5, "place", c("place","age")}

\item{n}{integer, remove rows that have >n NAs (ie, max number of NAs allowed)}
}
\value{
returns a new df with rows that have NA(s) removed
}
\description{
remove rows that have an NA anywhere/somewhere or a certain number of NAs
}
\note{
if neither col,n provided, search all columns for NA
\cr if col provided, n is ignored
\cr to use n, call the function like this: (df,n=3), where n=0 means searching all columns
}

