% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{z.names}
\alias{z.names}
\title{all column names, alias of \code{\link{names}}, there are also names(), colnames(), rownames(), row.names() but no col.names()}
\usage{
z.names()
}
\description{
all column names, alias of \code{\link{names}}, there are also names(), colnames(), rownames(), row.names() but no col.names()
}
\seealso{
\code{\link{nrow}}, \code{\link{ncol}}, \code{\link{dim}}, \code{\link{length}}
\cr \code{\link{z.len}}, \code{\link{z.size}}
\cr \code{\link{names}}, \code{\link{colnames}}, \code{\link{rownames}}, \code{\link{row.names}},
\cr \code{\link{z.rnames}}, \code{\link{z.cnames}}, \code{\link{z.names}}
}

