% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc.R
\name{z.newcol}
\alias{z.newcol}
\title{create a new col, may use \code{\link[dplyr]{mutate}} instead}
\usage{
z.newcol(df, newColName, defaultVal = NA)
}
\arguments{
\item{newColName}{''}

\item{defaultVal}{NA (default)}
}
\value{
returns a new df, old one does not change
}
\description{
create a new col, may use \code{\link[dplyr]{mutate}} instead
}
\seealso{
Other data.transformation.functions: \code{\link{z.rncol}}
}

