% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{z.num}
\alias{z.num}
\title{alias of \code{\link{type.convert}}}
\usage{
z.num(x, na.strings = "NA", as.is = FALSE, dec = ".",
  numerals = c("allow.loss", "warn.loss", "no.loss"))
}
\description{
Convert a character vector to logical, integer, numeric, complex or factor as appropriate.
}

