% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{z.pdfoff}
\alias{z.pdfoff}
\title{wrapper of \code{\link{dev.off}}}
\usage{
z.pdfoff(which = dev.cur())
}
\description{
wrapper of \code{\link{dev.off}}
}
\note{
additionally one can use \code{\link[extrafont]{embed_fonts}} to embed fonts after dev.off()
}
\examples{
A4:     width 7(inches) height = 5
Letter: 8.5 x 11
}
\seealso{
\code{\link{z.export}} \code{\link{z.pdfon}}
}

