% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{z.pdfon}
\alias{z.pdfon}
\title{wrapper of \code{\link{pdf}}}
\usage{
z.pdfon(file = ifelse(onefile, "Rplots.pdf", "Rplot\%03d.pdf"), width, height,
  onefile, family, title, fonts, version, paper, encoding, bg, fg, pointsize,
  pagecentre, colormodel, useDingbats, useKerning, fillOddEven, compress)
}
\description{
wrapper of \code{\link{pdf}}
}
\examples{
A4:     width 7(inches) height = 5
Letter: 8.5 x 11
}
\seealso{
\code{\link{z.export}} \code{\link{z.pdfoff}}
}

