% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc.R
\name{z.rand}
\alias{z.rand}
\title{random number, alias of \code{\link{runif}}}
\usage{
z.rand(n, min = 0, max = 1)
}
\description{
random number, alias of \code{\link{runif}}
}
\examples{
runif(n, min = 0, max = 1)
}

