% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{z.range}
\alias{z.range}
\title{similar to python range (python: left inclusive, right exclusive), wrapper of \code{\link{seq}}, may also consider 1:3}
\usage{
z.range(start, stop, step = 1)
}
\value{
returns a vector (both ends inclusive)
}
\description{
similar to python range (python: left inclusive, right exclusive), wrapper of \code{\link{seq}}, may also consider 1:3
}
\examples{
z.range(1,3) # 1 2 3, equivalent to 1:3
}

