% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{z.reads2}
\alias{z.reads2}
\title{alias of \code{\link[sjmisc]{read_spss}}}
\usage{
z.reads2(..., tolower = FALSE)
}
\arguments{
\item{tolower}{whether to convert all column names to lower case}
}
\description{
potentially keep variable labels and value labels.
}

