% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc.R
\name{z.readx}
\alias{z.readx}
\title{read an xlsx file}
\usage{
z.readx(file, sheetIndex = 1, ...)
}
\description{
read an xlsx file
}
\examples{
read.xlsx(file, sheetIndex, sheetName=NULL, rowIndex=NULL,
          startRow=NULL, endRow=NULL, colIndex=NULL,
          as.data.frame=TRUE, header=TRUE, colClasses=NA,
          keepFormulas=FALSE, encoding="unknown", ...)
colClasses: Only numeric, character, Date, POSIXct, column types are accepted
colClasses=c("Date", "character","integer", rep("numeric", 2),  "POSIXct")
}

