% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc.R
\name{z.recol}
\alias{z.recol}
\title{reorder a single col (sort of, see below)}
\usage{
z.recol(df, movecommand)
}
\arguments{
\item{movecommand}{sth like "v17, v18 before v3; v6, v16 last; v5 first", supports before/after, last/first}
}
\description{
reorder a single col (sort of, see below)
}

