% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc.R
\name{z.recols}
\alias{z.recols}
\title{reorder all cols}
\usage{
z.recols(df, newColOrder)
}
\arguments{
\item{newColOrder}{c('','',''), number of cols must match}
}
\value{
returns a new df, old one does not change
}
\description{
reorder all cols
}

