% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc.R
\name{z.regexpi}
\alias{z.regexpi}
\title{find with regular expression (case insensitive)}
\usage{
z.regexpi(s, pat, ignorecase = TRUE, once = FALSE, split = FALSE)
}
\arguments{
\item{s}{string}

\item{pat}{regular expression pattern}

\item{once}{match once or many times}

\item{split}{split or not (see below)}
}
\value{
returns a list ($start, $end, $match, $split)
\cr $start and $end as numeric vectors indicating the start and end positions of the matches.
\cr $match contains each exact match
\cr $split contains the character vector of splitted strings (split by the matched pattern, ie, the rest that do not match).
\cr If no match is found and split=FALSE, all components will be NULL
\cr If no match is found and split=TRUE, all components will be NULL except that split will contain the whole string
}
\description{
find with regular expression (case insensitive)
}

