% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc.R
\name{z.regexprep}
\alias{z.regexprep}
\title{replace string using regular expression (case sensitive)}
\usage{
z.regexprep(s, expr, repstr, ignorecase = FALSE, once = FALSE)
}
\description{
replace string using regular expression (case sensitive)
}

