% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{z.regexprepi}
\alias{z.regexprepi}
\title{replace string using regular expression (case insensitive)}
\usage{
z.regexprepi(s, expr, repstr, ignorecase = TRUE, once = FALSE)
}
\description{
replace string using regular expression (case insensitive)
}

