% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{z.rn}
\alias{z.rn}
\title{rename}
\usage{
z.rn(from, to)
}
\description{
rename
}
\examples{
support c('a.txt','b.txt'), c('d.txt','e.txt')
to parent folder must exist already; otherwise error
in case new name exists
      if old and new both folders, move old to new as subfolder
      if old and new both files, overwrite the new file with old file without prompt
}

