% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc.R
\name{z.rnames}
\alias{z.rnames}
\title{all row names, alias of \code{\link{row.names}}, there are also names(), colnames(), rownames(), row.names() but no col.names()}
\usage{
z.rnames(x)
}
\description{
all row names, alias of \code{\link{row.names}}, there are also names(), colnames(), rownames(), row.names() but no col.names()
}

