% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc.R
\name{z.rncol}
\alias{z.rncol}
\title{rename a single col}
\usage{
z.rncol(df, ...)
}
\arguments{
\item{...}{c("oldColName"="newColName")) or c(oldColName="newColName"))}
}
\value{
returns a new df, old one does not change
}
\description{
alias of \code{\link[reshape]{rename}}
}
\seealso{
Other data.transformation.functions: \code{\link{z.newcol}}
}

