% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc.R
\name{z.rncols}
\alias{z.rncols}
\title{rename all cols}
\usage{
z.rncols(df, newColNames)
}
\arguments{
\item{newColName}{c('','',''), number of cols must match}
}
\value{
returns a new df, old one does not change
}
\description{
rename all cols
}

