% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{z.save}
\alias{z.save}
\title{wrapper of write.csv, but with row.names removed}
\usage{
z.save(x, file = "RData.csv", row.names = FALSE, ...)
}
\description{
wrapper of write.csv, but with row.names removed
}
\examples{
(x, file="RData.csv", row.names=FALSE, append = FALSE, quote = TRUE, sep = ",",
            na = "NA", dec = ".",
            col.names = TRUE, qmethod = c("escape", "double"),
            fileEncoding = "")
dec: decimal point
}

