% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{z.saves}
\alias{z.saves}
\title{alias of \code{\link[sjmisc]{write_spss}}, \code{\link{z.writes}}}
\usage{
z.saves(x, path, enc.to.utf8 = TRUE)
}
\description{
potentially keep variable labels and value labels
}

