% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{z.show}
\alias{z.show}
\title{show information about a data frame or other object}
\usage{
z.show(...)
}
\description{
show information about a data frame or other object
}
\examples{
Compactly display the internal structure of an R object, a diagnostic function and an alternative to summary
(and to some extent, dput). Ideally, only one line for each ‘basic’ structure is displayed.
It is especially well suited to compactly display the (abbreviated) contents of (possibly nested) lists.
The idea is to give reasonable output for any R object. It calls args for (non-primitive) function objects.
}

