% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{z.values.get}
\alias{z.values.get}
\title{get value labels, wrapper of \code{\link[sjmisc]{get_labels}}}
\usage{
z.values.get(x, include.values = NULL, attr.only = T,
  include.non.labelled = F, ...)
}
\value{
returns a list $varname
}
\details{
see also \code{\link[sjmisc]{get_values}}
}
\examples{

}
\seealso{
\code{\link[tidyr]{gather}}, \code{\link[tidyr]{spread}}, \code{\link[tidyr]{separate}}, \code{\link[tidyr]{unite}}
\cr \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}
\cr \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{arrange}}
\cr \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{mutate}}
\cr \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{right_join}}, \code{\link[dplyr]{inner_join}}, \code{\link[dplyr]{full_join}}, \code{\link[dplyr]{semi_join}}, \code{\link[dplyr]{anti_join}}
\cr \code{\link[dplyr]{intersect}}, \code{\link[dplyr]{union}}, \code{\link[dplyr]{setdiff}}
\cr \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{bind_cols}}

Other data.transformation.functions: \code{\link{z.2factor}},
  \code{\link{z.2label}}, \code{\link{z.2long}},
  \code{\link{z.2lower}}, \code{\link{z.2value}},
  \code{\link{z.2wide}}, \code{\link{z.compute}},
  \code{\link{z.delete}}, \code{\link{z.del}},
  \code{\link{z.label.get}}, \code{\link{z.label.set}},
  \code{\link{z.leftjoin}}, \code{\link{z.move}},
  \code{\link{z.newcol}}, \code{\link{z.recode2}},
  \code{\link{z.recode}}, \code{\link{z.recols}},
  \code{\link{z.recol}}, \code{\link{z.rename}},
  \code{\link{z.rncols}}, \code{\link{z.rncol}},
  \code{\link{z.select}}, \code{\link{z.sort}},
  \code{\link{z.split}}, \code{\link{z.unique}},
  \code{\link{z.values.set}}
}

