% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{z.ver}
\alias{z.ver}
\title{print the version of a package}
\usage{
z.ver(pkg = NULL)
}
\arguments{
\item{pkg}{package name in quotes, default is NULL}
}
\value{
if pkg not provided, prints only R version
}
\description{
print the version of a package
}

