% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{z.view}
\alias{z.view}
\title{show information about a data frame or similar object (like spss variable view)}
\usage{
z.view(x, showFreq = T, showPerc = T, sortByName = F, ...)
}
\description{
wrapper of \code{\link[sjPlot]{view_df}}; can make the html bigger by openning in internet browser
}

