% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{z.wherena}
\alias{z.wherena}
\title{visualize where all the NAs of a dataframe are}
\usage{
z.wherena(df, id, color = "red", angle = 270, basesize = 9, xsize = 1,
  ysize = 1)
}
\arguments{
\item{df}{data frame in wide format}

\item{id}{a column name as id, will be shown as y axis, quoted "", eg, subject ID}

\item{color}{color of missing values, eg, "red"}

\item{angle}{the x axis label angle, default=270 (vertical), suggests 330 if label is not too long}

\item{basesize}{base font size}

\item{xsize}{x axis label font relative size}

\item{ysize}{y axis label font relative size}
}
\value{
a ggplot object (+theme_apa() to get apa format plot)
}
\description{
visualize where all the NAs of a dataframe are
}

