% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{z.who}
\alias{z.who}
\title{alias of \code{\link{sessionInfo}}, \code{\link{z.whos}}}
\usage{
z.who(package = NULL)
}
\value{
Print version information about R, the OS and attached or loaded packages.
}
\description{
alias of \code{\link{sessionInfo}}, \code{\link{z.whos}}
}
\seealso{
\code{\link{objects}}
}

