% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmisc.R
\name{z.whos}
\alias{z.whos}
\title{alias of \code{\link{objects}}}
\usage{
z.whos(name, pos = -1L, envir = as.environment(pos), all.names = FALSE,
  pattern, sorted = TRUE)
}
\value{
returns/gives the names of the objects in the specified environment
}
\description{
alias of \code{\link{objects}}
}

