% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.2label}
\alias{ez.2label}
\title{number ef[(0/1)]|   {attr number g(0/1) / factor attr number h[0/1]}-->factor char i[male/female]   |char jk[(male/female)]}
\usage{
ez.2label(x, col = NULL, add.non.labelled = TRUE, drop.is_na = FALSE, ...)
}
\arguments{
\item{x}{a data frame or a vector/col}

\item{col}{if x is a data frame, col is specified (e.g., "cond"), convert that col only
\cr        if x is a data frame, col is unspecified (i.e., NULL default), convert all possible cols in x
\cr        if x is not a data frame, col is ignored}

\item{drop.is_na}{ignore is_na attr, if yes, treat as NA}
}
\value{
returns a factor with string as its levels or a data frame with changed col(s)
}
\description{
e=e, f=f, g/h/i->i, j=j, k=k
}
\details{
Both value and variable label attributes will be removed when converting variables to factors.
\cr wrapper of \code{\link[sjmisc]{to_label}}
}
\examples{
e=c(1,2); f=factor(1:2) 
g=c(0,0,1,1,1,0); attr(g,'value.labels') <- c(boy=1,girl=0)
h=factor(c(0,0,1,1,1,0)); attr(h,'labels') <- c(boy=1,girl=0)
i=factor(c('girl','girl','boy','boy','boy','girl'))
j=c('x','y'); k=factor(c('x','y'))

ez.2label: e=e, f=f, g/h/i->i, j=j, k=k
ez.2value: e=e<-f, g<-(0,1+attr)g/(0,1-attr)h//(1,2+attr)i, (1,2+attr)<-j/k
ez.2factor: ef->f, g/h->h, i=i, factor[x,y,z]<j/k
}
\seealso{
Other data.transformation.functions: \code{\link{ez.2char}},
  \code{\link{ez.2factor}}, \code{\link{ez.2long}},
  \code{\link{ez.2lower}}, \code{\link{ez.2value}},
  \code{\link{ez.2wide}}, \code{\link{ez.compute}},
  \code{\link{ez.del}}, \code{\link{ez.label.get}},
  \code{\link{ez.label.set}}, \code{\link{ez.leftjoin}},
  \code{\link{ez.move}}, \code{\link{ez.newcol}},
  \code{\link{ez.num}}, \code{\link{ez.recode2}},
  \code{\link{ez.recode}}, \code{\link{ez.recols}},
  \code{\link{ez.recol}}, \code{\link{ez.rename}},
  \code{\link{ez.replace}}, \code{\link{ez.rncols}},
  \code{\link{ez.rncol}}, \code{\link{ez.select}},
  \code{\link{ez.sort}}, \code{\link{ez.split}},
  \code{\link{ez.str}}, \code{\link{ez.unique}},
  \code{\link{ez.values.get}}, \code{\link{ez.values.set}}
}

