% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.2label}
\alias{ez.2label}
\title{number ef[(0/1)]|   {attr number g(0/1) / factor attr number h[0/1]}-->factor char i[male/female]   |char jk[(male/female)]}
\usage{
ez.2label(x, col = NULL, add.non.labelled = TRUE,
  drop.missing.value = FALSE, ...)
}
\arguments{
\item{x}{a data frame or a vector}

\item{col}{internally evaluated by eval('dplyr::select()')
\cr        if x is a data frame, col is specified (e.g., "cond"), convert that col only
\cr        if x is a data frame, col is unspecified (i.e., NULL default), convert all possible cols in x
\cr        if x is not a data frame, col is ignored}

\item{add.non.labelled}{logical, if TRUE, values without associated value label (ie, more values than "labels", hence not all values are labelled.) will also be converted to labels (as is). If FALSE, non-labelled values are converted to NA}

\item{drop.missing.value}{if TRUE, all types of missing value codes are converted (is_na attr) into NA before x is converted as factor. If FALSE, ignore is_na attr, missing values will be left as their original codes.}
}
\value{
returns a factor with string as its levels or a data frame with changed col(s)
}
\description{
e=e, f=f, g/h/i->i, j=j, k=k
}
\details{
Both value and variable label attributes will be removed when converting variables to factors.
\cr wrapper of \code{\link[sjmisc]{to_label}}
}
\examples{
e=c(1,2); f=factor(1:2) 
g=c(0,0,1,1,1,0); attr(g,'value.labels') <- c(boy=1,girl=0)
h=factor(c(0,0,1,1,1,0)); attr(h,'labels') <- c(boy=1,girl=0)
i=factor(c('girl','girl','boy','boy','boy','girl'))
j=c('x','y'); k=factor(c('x','y'))

ez.2label: e=e, f=f, g/h/i->i, j=j, k=k
ez.2value: e=e, f(0,1)<-f[1,2], g=g, h(0,1-attr)<-h, g(0,1+attr)<-i/j/k
ez.2factor: ef->f, g/h->h, i=i, factor[x,y,z]<j/k
}
\seealso{
Other data.transformation.functions: \code{\link{ez.2char}},
  \code{\link{ez.2factor}}, \code{\link{ez.2long}},
  \code{\link{ez.2lower}}, \code{\link{ez.2value}},
  \code{\link{ez.2wide}}, \code{\link{ez.compute}},
  \code{\link{ez.del}}, \code{\link{ez.label.get}},
  \code{\link{ez.label.set}}, \code{\link{ez.leftjoin}},
  \code{\link{ez.move}}, \code{\link{ez.newcol}},
  \code{\link{ez.num}}, \code{\link{ez.recode}},
  \code{\link{ez.recols}}, \code{\link{ez.recol}},
  \code{\link{ez.rename}}, \code{\link{ez.replace}},
  \code{\link{ez.rncols}}, \code{\link{ez.rncol}},
  \code{\link{ez.select}}, \code{\link{ez.sort}},
  \code{\link{ez.split}}, \code{\link{ez.str}},
  \code{\link{ez.unique}}, \code{\link{ez.values.get}},
  \code{\link{ez.values.set}}
}

