% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.anovas}
\alias{ez.anovas}
\title{a series of one-way anova, for many y and many x; if many y and many x at the same time, returns a list, see note for details}
\usage{
ez.anovas(df, y, x, pthreshold = 0.05, showerror = F, print2screen = T,
  viewresults = F, plot = T, facet = "cols", pmethods = c("bonferroni",
  "fdr"), ...)
}
\arguments{
\item{df}{a data frame
\cr NA in df will be auto excluded in aov(), reflected by degree_of_freedom}

\item{y}{internally evaluated by eval('dplyr::select()'), a vector of continous variables c('var1','var2'), or a single variable 'var1', if it is a factor, auto converts to numeric (internally call ez.2value(df[[yy]]), (eg, names(select(beta,Gender:dmce)))}

\item{x}{internally evaluated by eval('dplyr::select()'), a vector of categorical variables, or a single categorical variable}

\item{pthreshold}{default .05, print/output results whenever p < pthreshold, could be 1 then get all}

\item{showerror}{whether show error message when error occurs, default F}

\item{plot}{T/F, the dash line is bonferroni p = 0.05}

\item{facet}{one of 'cols', 'rows', 'wrap', valid only if plot=T}

\item{pmethods}{c('bonferroni','fdr'), type p.adjust.methods for all methods. even though pthreshold only shows a few sig results, this correction applies for all possible tests that have been done.}

\item{...}{dots passed to ez.2value(df[[yy]],...)}
}
\value{
an invisible data frame with x,y,p,means and print results out on screen; results can then be saved using ez.savex(results,'results.xlsx')
\cr the means column in excel can be split into mulitiple columns using Data >Text to Columns
\cr degree_of_freedom: from F-statistic
}
\description{
df = ez.2value(df,y,...); df = ez.2factor(df,x); aov(df[[yy]]~df[[xx]])
}
\note{
if many y and x at the same time, returns a list. $xlist for ez.savexlist(xlist), $plist (if plot = T) for ggmultiplot(plotlist = plist,cols=3)
}
