% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.anovas}
\alias{ez.anovas}
\title{a series of one-way anova, for many y and many x}
\usage{
ez.anovas(df, y, x, pthreshold = 0.05, showerror = F)
}
\arguments{
\item{df}{a data frame}

\item{y}{a vector of continous variables c('var1','var2'), or a single variable 'var1', if it is a factor, auto converts to numeric (internally call ez.2value(df[[yy]]), (eg, names(select(beta,Gender:dmce)))}

\item{x}{a vector of categorical variables, or a single categorical variable}

\item{pthreshold}{default .05, print/output results whenever p < pthreshold, could be 1 then get all}

\item{showerror}{whether show error message when error occurs, default F}

\item{...}{dots passed to ez.2value(df[[yy]],...)}
}
\value{
an invisible data frame with x,y,p,means and print results out on screen; results can then be saved using ez.savex(results,'results.xlsx')
\cr the means column in excel can be split into mulitiple columns using Data >Text to Columns
}
\description{
aov(ez.2value(df[[yy]])~df[[xx]])
}

