% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.append}
\alias{ez.append}
\title{append a row to an exisiting data frame}
\usage{
ez.append(df, newrow, print2screen = TRUE)
}
\arguments{
\item{df}{df to be appended}

\item{newrow}{a vector, e.g.,  c("Ted", 25)}

\item{print2screen}{whether to print the new row to string (auto separated by tab), default TRUE}
}
\value{
returns a new df, old passed df does not change
}
\description{
could be slow, commonly use together with \code{\link{ez.header}}
}
\examples{
typically within a loop:  results = ez.append(results,c(var,p,R,U))
}

