% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.barplot}
\alias{ez.barplot}
\title{barplot with ggplot}
\usage{
ez.barplot(df, cmd, bar_gap = 0.7, bar_width = 0.7, error_size = 0.7,
  error_gap = 0.7, error_width = 0.3, error_direction = "both",
  ylab = NULL, xlab = NULL, zlab = NULL, zpos = "top", xangle = 0,
  vjust = NULL, hjust = NULL)
}
\arguments{
\item{df}{data frame in long format}

\item{cmd}{like "y|x, y|x z" where y (axis) is continous, x (axis) z (legend) are discrete}
}
\value{
a ggplot object (+theme_apa() to get apa format plot)
}
\description{
barplot with ggplot
}

