% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.boxcox}
\alias{ez.boxcox}
\title{box-cox power transformation}
\usage{
ez.boxcox(y, col = NULL, na.rm = FALSE, plot = TRUE, print2scr = TRUE,
  value = TRUE, value.force = FALSE, value.method = c("boxcox",
  "tukey.modified"), value.lambda = c("rounded", "raw"), ...)
}
\arguments{
\item{y}{a data frame or a vector}

\item{col}{passed to \code{\link{ez.selcol}}
\cr        if x is a data frame, col specified, process that col only.
\cr        if x is a data frame, col unspecified (i.e., NULL default), process all cols
\cr        if x is not a data frame, col is ignored
\cr        could be multiple cols}

\item{na.rm}{rm na from y,x (pairwise), if not, NA stays as is. applicable only if y is a vector.}

\item{plot}{boxcox plot. applicable only if y is a vector.}

\item{print2scr}{print out transformation parameters, only when there is an actual transformation}

\item{value}{return transformed y, or a list of transformation parameters. If y is a data frame, value will be always (internally) set to TRUE}

\item{value.force}{transform regardless or only if p.lambda rounded is less than .05. otherwise returns original y}

\item{value.method}{boxcox (scaled tukey) or modified tukey, both methods keep the ordering. see \code{\link[car]{bcPower}}}

\item{value.lambda}{use rounded lambda, one of c(1, 0, -1, 0.5, 0.33, -0.5, -0.33, 2, -2) or raw/calculated lambda}
}
\value{
returns transformed y, or a list of transformation parameters, otherwise original y (depending on value, value.force). 
\cr If y is a data frame, always return a potentially transformed data frame, and no printout, no plot
}
\description{
box-cox power transformation
}
\note{
Box and Cox (1964) generally deals with non-negative responses. This funciton can handle (a few) negative responses (Hawkins and Weisberg (2017))
\cr while allowing for the transformed data to be interpreted similarly to the interpretation of Box- Cox transformed values.
\cr essentially estimate/add a number (ie, gamma) to y to make it positive
}
