% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.color}
\alias{ez.color}
\title{show some help info on color}
\usage{
ez.color()
}
\description{
show some help info on color
}
\examples{
RColorBrewer::display.brewer.all()
cols <- RColorBrewer::brewer.pal(8,"Set3")
colorRampPalette(brewer.pal(9,"Blues"))(100)
}

