% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.compare}
\alias{ez.compare}
\title{compare two vectors, two dataframes}
\usage{
ez.compare(lh, rh, ...)
}
\description{
compare two vectors, two dataframes
}
\note{
nrow() for data frame, length() for vector
}

