% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.corrmap}
\alias{ez.corrmap}
\title{plot a correlation matrix map}
\usage{
ez.corrmap(df, corr.type = "pearson", sig.level = 0.05, insig = "blank",
  method = "color", tl.col = "black", tl.cex = 0.4, col = NULL, ...)
}
\arguments{
\item{df}{data frame in wide format, should be all numeric}

\item{corr.type}{"pearson" or "spearman", pairwise deletion for NA}

\item{sig.level}{sig.level}

\item{insig}{how to treat insig values, one of "pch"(show x),"p-value","blank", "n"(no change, as is)}

\item{...}{see \code{\link[corrplot]{corrplot}} for more parameters}
}
\value{
returns a list (r, p) r: a matrix representing the corrmap (p > sig.level, set to NA/blank), p: all raw p values
}
\description{
a wrapper of \code{\link[corrplot]{corrplot}}; the correlation and p values are calculated with \code{\link[Hmisc]{rcorr}}
}

