% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.countplot}
\alias{ez.countplot}
\title{plot count data}
\usage{
ez.countplot(
  df,
  cmd,
  position = "both",
  color = "color",
  colors = ez.palette("Zhu"),
  alpha = 1,
  n.size = 5.5,
  n.type = 3,
  width = 0.7,
  ylab = NULL,
  xlab = NULL,
  zlab = NULL,
  legend.position = "top",
  legend.direction = "horizontal",
  legend.box = T,
  legend.size = c(0, 10),
  xangle = 0,
  vjust = NULL,
  hjust = NULL,
  facet = "cols",
  theme.apa = TRUE
)
}
\arguments{
\item{df}{data frame in long format (but be careful that standard error might be inaccurate depending on grouping in the long format)}

\item{cmd}{like "x", c("x1", "x2", "x3"), "x|z", "x|z a" where x z a are all discrete.}

\item{position}{so far can only be "stack", "fill", "both". ('dodge' not supported yet)}

\item{color}{"bw" or "color"  black/white or colorblind-friendly color}

\item{alpha}{bar alpha value}

\item{n.size}{set to 0 to hide count/percentage}

\item{n.type}{1 = n for stack, pct for fill; 2 = pct for stack, n for fill; 3 = n (pct) for stack, pct (n) for fill; 4 = pct (n) for stack, n (pct) for fill}

\item{width}{the width of bar/bin itself}

\item{ylab}{y label NULL}

\item{xlab}{x label NULL}

\item{zlab}{z/a/fill/legend label, only applicable when there is z provided NULL}

\item{legend.position}{legend position 'top', 'bottom', 'left', 'right', 'none', c(x,y,two-element numeric vector)
\cr         c(0,0) corresponds to the "bottom left" and c(1,1) corresponds to the "top right" position.
\cr         if no z/a (legend) provided, auto force to 'none'}

\item{legend.direction}{horizontal or vertical}

\item{legend.box}{box of legend, T or F}

\item{legend.size}{c(0,10) the first number 0 controls the legend title, 0=hide; the second number controls legend.key.size, legend.text}

\item{xangle}{angle of x text 0}

\item{vjust}{vjust of x text NULL}

\item{hjust}{hjust of x text NULL}

\item{facet}{one of 'cols', 'rows', 'wrap'}
}
\value{
a ggplot object (+theme_apa() to get apa format plot), +scale_y_continuous(limits=c(-5,8),breaks=seq(-5,8,by=2),oob=scales::rescale_none)
\cr see http://stackoverflow.com/a/31437048/2292993 for discussion
}
\description{
plot count data, eg, ez.countplot(iris, 'Species'). See also \code{\link{ez.piechart}} \code{\link{ez.hist}}
}
