% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.dprime}
\alias{ez.dprime}
\title{Dprime and Other Signal Detection Theory indices.}
\usage{
ez.dprime(n_hit, n_fa, n_miss, n_cr, adjusted = TRUE)
}
\arguments{
\item{n_hit}{Number of hits.}

\item{n_fa}{Number of false alarms.}

\item{n_miss}{Number of misses.}

\item{n_cr}{Number of correct rejections.}

\item{adjusted}{Should it use the Hautus (1995) adjustments for extreme values (hit rate of 1 and false alarm rate of 0). see more at \href{https://stats.stackexchange.com/questions/134779}{stackexchange}}
}
\description{
Computes Signal Detection Theory indices (%correct, d', beta, A', B''D, c, c').
}
\examples{
n_hit <- 9
n_fa <- 2
n_miss <- 1
n_cr <- 7

indices <- dprime(n_hit, n_fa, n_miss, n_cr)


df <- data.frame(Participant = c("A", "B", "C"),
    n_hit = c(1, 2, 5),
    n_fa = c(6, 8, 1))

indices <- dprime(n_hit=df$n_hit,
    n_fa=df$n_fa,
    n_targets=10,
    n_distractors=10,
    adjusted=FALSE)


}
\author{
Jerry modified from \href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
