% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.duplicated}
\alias{ez.duplicated}
\title{duplicated}
\usage{
ez.duplicated(x, col = NULL, vec = TRUE, dim = 1, incomparables = FALSE,
  value = FALSE, ...)
}
\arguments{
\item{x}{a data frame or a vector/col of any data type (factor, char, numeric)}

\item{col}{restrict to the columns where you would like to search for duplicates; e.g., 3, c(3), 2:5, "place", c("place","age")
\cr if x is a data frame, col is specified (e.g., "cond"), check that col only
\cr if x is a data frame, col is unspecified (i.e., NULL default), check all cols in x
\cr if x is not a data frame, col is ignored}

\item{vec}{TRUE/FALSE, if TRUE, returns a vector of TRUE/FALSE indicating duplicates; 
\cr if FALSE, returns a df with one column 'Duplicated' of TRUE/FALSE
\cr This is useful for binding with other data frames}

\item{dim}{1=find duplicated rows, 2=find duplicated cols. dim has no effect when x is a vector}

\item{incomparables}{a vector of values that cannot be compared. FALSE is a special value, meaning that all values can be compared, 
and may be the only value accepted for methods other than the default. It will be coerced internally to the same type as x.}

\item{value}{TRUE/FALSE, if TRUE, returns actual duplicated values, instead of logicals. 
The returned data type is the same as the original (data frame->data frame, factor->factor, etc, because only slicing based on logicals). 
Ignore/Overwrite vec=T/F.}
}
\value{
return depends, see vec above (By default, missing values are regarded as equal, to avoid that, pass incomparables=NA)
\cr this is different from the built-in R \code{\link{duplicated}}
\cr x <- c(1, 1, 4, 5, 4, 6)  duplicated(x) returns [1] FALSE TRUE FALSE FALSE TRUE FALSE
\cr but ez.duplicated(x) returns [1] TRUE TRUE TRUE FALSE TRUE FALSE
\cr Also, the function has a trick, so that duplicated cols could be checked, while the native duplicated cannot directly apply to cols. See https://stackoverflow.com/questions/9818125/
}
\description{
find the duplicated rows/cols in a data frame or duplicated elements in a vector of any data type (factor, char, numeric)
}

