% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R, R/os.R
\name{ez.env}
\alias{ez.env}
\title{switch env}
\usage{
ez.env(env = NULL)

ez.env(env = NULL)
}
\arguments{
\item{env}{NULL=print out current env (.libPaths()[1]); use symlink to trick (see code)}

\item{env}{NULL=print out current env (.libPaths()[1]); use symlink to trick (see code)}
}
\value{
returns NULL

returns NULL
}
\description{
switch env

switch env
}

