% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.env}
\alias{ez.env}
\alias{ez.envr}
\title{switch env}
\usage{
ez.env(env = NULL)

ez.envr(env = NULL)
}
\arguments{
\item{env}{NULL=print out current env (.libPaths()[1]); use symlink to trick (see code)}
}
\value{
returns NULL
}
\description{
switch env
}
