% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.es.t.paired.m12s12r}
\alias{ez.es.t.paired.m12s12r}
\title{calculate effect size}
\usage{
ez.es.t.paired.m12s12r(m12, s12, r)
}
\arguments{
\item{m12}{the mean of differences that equals the difference of means (m1-m2), available in SPSS paired samples Test output table}

\item{s12}{the standard deviation of the difference score, available in SPSS paired samples Test output table}

\item{r}{correlation}
}
\value{
returns invisible
}
\description{
calculate effect size
}
