% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.eval}
\alias{ez.eval}
\title{wrapper of \code{\link{eval}}}
\usage{
ez.eval(cmd)
}
\arguments{
\item{cmd}{an R cmd in text, e.g., constructed with sprintf()}
}
\value{
this function simply evaluates the cmd in the caller's envir, so the actual return depends on the caller/cmd
}
\description{
wrapper of \code{\link{eval}}
}

