% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.eval}
\alias{ez.eval}
\alias{ez.evaluate}
\title{wrapper of \code{\link{eval}}}
\usage{
ez.eval(cmd, env = parent.frame())

ez.evaluate(cmd, env = parent.frame())
}
\arguments{
\item{cmd}{an R cmd in text, e.g., constructed with sprintf()}

\item{env}{caller's envir, could be environment(), .GlobalEnv (same as globalenv()), default is parent.frame()
if default parent.frame() does not work properly, try passing environment()
tricky/hard to hard-code/default it in this wrapper function, R does not support call stack well.}
}
\value{
this function evaluates the cmd in the caller's envir, so the actual return depends on the caller/cmd
}
\description{
wrapper of \code{\link{eval}}
}
\examples{
ez.eval('z=9',environment())
}
