% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.export}
\alias{ez.export}
\title{wrapper of \code{\link{dev.copy2pdf}}}
\usage{
ez.export(filename = "RPlot.pdf", pdf.width = 11, pdf.height = 8.5, ...)
}
\description{
wrapper of \code{\link{dev.copy2pdf}}
}
\examples{
A4:     width 7(inches) height = 5
Letter: 8.5 x 11 (portrait default ie, 11 x 8.5)
}
\seealso{
\code{\link{pdf}} \code{\link{ez.pdfon}} \code{\link{ez.pdfoff}}
}
