% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.factorder}
\alias{ez.factorder}
\title{change factor level order in a df}
\usage{
ez.factorder(df, col, ord = "as")
}
\arguments{
\item{df}{data frame}

\item{col}{a factor column name, quoted "", eg, "group"}

\item{ord}{"az","za"--alphabetic;
\cr "as"--as is, appearance;
\cr c("small","medium","large")--specified level order
\cr "col2" --another column in az
\cr "col2:az" --another column in az
\cr "col2:za" --another column in za}
}
\value{
returns a new df
}
\description{
does not change factor label; only changes the order of printing out
}

