% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.factorname}
\alias{ez.factorname}
\title{change factor level names in a df}
\usage{
ez.factorname(df, col, newLevelNames)
}
\arguments{
\item{df}{data frame}

\item{col}{a factor column name, quoted "", eg, "group"}

\item{newLevelNames}{new level names coresponding to levels(x), eg, c("one","two","three")}
}
\value{
returns a new df
}
\description{
change factor level names in a df
}
\references{
\href{http://www.cookbook-r.com/Manipulating_data/Renaming_levels_of_a_factor/}{Cookbook R: Renaming levels of a factor}
}

