% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.factorname}
\alias{ez.factorname}
\title{change factor level names in a df}
\usage{
ez.factorname(x, col, newLevelNames)
}
\arguments{
\item{x}{data frame or vector, factor or non-factor}

\item{col}{a single column name, quoted string, ignored when x is not a data frame}

\item{newLevelNames}{new level names coresponding to levels(x), eg, c("one","two","three")}
}
\value{
returns a new df, factor (non-factor->factor)
}
\description{
change factor level names in a df
}
\references{
\href{http://www.cookbook-r.com/Manipulating_data/Renaming_levels_of_a_factor/}{Cookbook R: Renaming levels of a factor}
}
