% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.fishers}
\alias{ez.fishers}
\title{a series of fisher.test, for many y and many x; if many y and many x at the same time, returns a list}
\usage{
ez.fishers(df, y, x, showerror = T, viewresult = F, plot = T, cols = 3,
  pmethods = c("bonferroni", "fdr"), labsize = 2, textsize = 1.5,
  titlesize = 3, width = 300)
}
\arguments{
\item{df}{a data frame, if its column is factor, auto converts to numeric (internally call ez.2factor(df))
\cr NA in df will be auto excluded in fisher.test(), reflected by total}

\item{y}{internally evaluated by eval('dplyr::select()'), a vector of outcome variables c('var1','var2'), or a single variable 'var1'}

\item{x}{internally evaluated by eval('dplyr::select()'), a vector of predictors, or a single predictor, (eg, names(select(beta,Gender:dmce)), but both mulitple/single x, only simple regression)}

\item{showerror}{whether show error message when error occurs}

\item{plot}{T/F, the dash line is bonferroni p = 0.05}

\item{cols}{number of columns for ggmultiplot. NULL=auto calculate}

\item{pmethods}{c('bonferroni','fdr'), type p.adjust.methods for all methods. This correction applies for all possible tests that have been/could be done.}

\item{width}{width for toString(countTable,width=width)}
}
\value{
an invisible list of data frame
}
\description{
df=ez.2factor(df,c(x,y)), fisher.test(df[[xx]],df[[yy]])
}
\note{
odds ratio only exist for 2x2 table, otherwise 0 (arbitrary assigned by jerry)
}
