% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.fullpath}
\alias{ez.fullpath}
\alias{ez.fp}
\title{Returns the full path with resolving}
\usage{
ez.fullpath(path)

ez.fp(path)
}
\description{
Returns the full path with resolving
}
\note{
Resolve ~, %WinEnvVar%, * or ? wildcard, relative path \cr
Mac: ~  --> /Users/jerry \cr
Windows: ~  --> C:\\Users\\Jerry\\Documents \cr
Mac environment variable not supported (because it cannot load .bash_profile--check Sys.getenv()
and tricker to parse a variable name in path)
}
\examples{
ez.fp('\%WINAPPS\%/RStudio')
}
