% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.hist}
\alias{ez.hist}
\title{plot continous data}
\usage{
ez.hist(x, cmd, bins = 60, density = FALSE, xline = NULL,
  color = "color", alpha = 0.5, ylab = NULL, xlab = NULL, zlab = NULL,
  legend.position = "top", legend.direction = "horizontal",
  legend.box = T, legend.size = c(0, 10), xangle = 0, vjust = NULL,
  hjust = NULL, facet = "cols", ...)
}
\arguments{
\item{x}{df or vector, if vector cmd ignored}

\item{cmd}{like "x, x|z, x|z a" where x z a are all discrete}

\item{bins}{number of bins (alternatively pass bidwidth that depends on the scale of variable)}

\item{density}{when true ignore bins, plot density (sort of smoothed histogram); when false plot normal histogram}

\item{xline}{draw red line(s) on x axis, eg, 2.5 or c(2.5,4). NULL=mean(df[[xx]])}

\item{color}{"bw" or "color"  black/white or colorblind-friendly color}

\item{alpha}{bar alpha value}

\item{ylab}{y label NULL}

\item{xlab}{x label NULL}

\item{zlab}{z/a/fill/legend label, only applicable when there is z provided NULL}

\item{legend.position}{legend position 'top', 'bottom', 'left', 'right', 'none', c(x,y,two-element numeric vector)
\cr         c(0,0) corresponds to the "bottom left" and c(1,1) corresponds to the "top right" position.
\cr         if no z/a (legend) provided, auto force to 'none'}

\item{legend.direction}{horizontal or vertical}

\item{legend.box}{box of legend, T or F}

\item{legend.size}{c(0,10) the first number 0 controls the legend title, 0=hide; the second number controls legend.key.size, legend.text}

\item{xangle}{angle of x text 0}

\item{vjust}{vjust of x text NULL}

\item{hjust}{hjust of x text NULL}

\item{facet}{one of 'cols', 'rows', 'wrap'}

\item{print.cmd}{T/F}
}
\value{
a ggplot object (+theme_apa() to get apa format plot), +scale_y_continuous(limits=c(-5,8),breaks=seq(-5,8,by=2),oob=scales::rescale_none)
\cr see http://stackoverflow.com/a/31437048/2292993 for discussion
}
\description{
plot continous data
}
