% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.is.numeric.like}
\alias{ez.is.numeric.like}
\title{Check numeric}
\usage{
ez.is.numeric.like(
  x,
  naAsTrue = TRUE,
  na.strings = c("", ".", "NA", "na", "N/A", "n/a", "NaN", "nan")
)
}
\arguments{
\item{x}{a str vector, or a factor of str vector, or numeric vector. x will be coerced and trimws.}

\item{naAsTrue}{whether NA (including actual NA and na.strings) will be treated as numeric like}

\item{na.strings}{case sensitive strings that will be treated to NA.}
}
\value{
a logical vector (vectorized).
}
\description{
check if a str obj is actually numeric
}
\note{
Using regular expression
\cr TRUE for any actual numeric c(3,4,5,9.9) or c("-3","+4.4",   "-42","4L","9L",   "1.36e4","1.36E4",    NA, "NA", "","NaN", NaN):
\cr positive or negative numbers with no more than one decimal c("-3","+4.4") OR
\cr positive or negative integers (e.g., c("-42","4L","39L")) OR
\cr positive or negative numbers in scientific notation c("1.36e4","1.36e-4","1.36e+4","1.36E4")
\cr NA, or na.strings
}
