% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.joinpath}
\alias{ez.joinpath}
\alias{ez.jp}
\title{join path}
\usage{
ez.joinpath(...)

ez.jp(...)
}
\description{
join path
}
\note{
\code{\link{file.path}} cannot handle/ignore trailing slash,
internally trim double+ and trailing slashes, though some OS treats multiple slashes as one
alternative: paste(..., sep=.Platform$file.sep)
}
