% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjmisc18.R
\name{ez.labels.swap}
\alias{ez.labels.swap}
\alias{ez.swaplabels}
\title{swap value labels format, to be consistent, haven style ("labels", mode 2) is preferred to foreign ("value.labels", mode 1)}
\usage{
ez.labels.swap(df, mode = 2)

ez.swaplabels(df, mode = 2)
}
\arguments{
\item{mode}{1: foreign, "value.labels";  2: haven, "labels"
\cr 
\cr 1: foreign: variable.labels (as df attributes), value.labels
\cr 2: haven: label = variable label, labels = value labels}
}
\description{
swap value labels format, to be consistent, haven style ("labels", mode 2) is preferred to foreign ("value.labels", mode 1)
}
