% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.log}
\alias{ez.log}
\title{Prints/Directs output to both terminal and a file (log.txt) globally, wrapper of \code{\link{sink}}}
\usage{
ez.log(file = "log.txt", mode = "a", status = TRUE, timestamp = TRUE)
}
\arguments{
\item{mode}{a=append; w=overwrite}

\item{status}{T=open the redirection/file, F=close the redirection}

\item{timestamp}{T=insert timestamp at the beginning and end, F=otherwise}
}
\value{
nothing
}
\description{
recommend to use logon(), logoff()
}
\examples{
# logging on: log("thelog.txt")
# logging off: log(status=False)  # no need to pass in the file parameter
}
\seealso{
\code{\link{ez.print}}
}

