% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.logistics}
\alias{ez.logistics}
\title{a series of simple logistic regression, for many y and many x; if many y and many x at the same time, returns a list}
\usage{
ez.logistics(df, y, x, covar = NULL, showerror = T, viewresult = F,
  plot = T, cols = 3, pmethods = c("bonferroni", "fdr"), labsize = 2,
  textsize = 1.5, titlesize = 3, ...)
}
\arguments{
\item{df}{a data frame, if its column is factor, auto converts to numeric (internally call ez.2value(df))
\cr NA in df will be auto excluded in glm(), reflected by degree_of_freedom}

\item{y}{internally evaluated by eval('dplyr::select()'), a vector of outcome variables c('var1','var2'), or a single variable 'var1'}

\item{x}{internally evaluated by eval('dplyr::select()'), a vector of predictors, or a single predictor, (eg, names(select(beta,Gender:dmce)), but both mulitple/single x, only simple regression)}

\item{covar}{NULL=no covar, internally evaluated by eval('dplyr::select()'), a vector of covariates c('var1','var2'), or a single variable 'var1'}

\item{showerror}{whether show error message when error occurs}

\item{plot}{T/F, the dash line is bonferroni p = 0.05}

\item{cols}{number of columns for gmultiplot. NULL=auto calculate}

\item{pmethods}{c('bonferroni','fdr'), type p.adjust.methods for all methods. This correction applies for all possible tests that have been/could be done.}

\item{...}{dots passed to ez.2value(df,...)}
}
\value{
an invisible list of data frame
\cr odds_ratio: odds ratio=exp(b), one unit increase in x result in the odds of being 1 for y "OR" times the odds of being 0 for y
\cr so that the variances of dependent and independent variables are 1.
\cr Therefore, standardized coefficients refer to how many standard deviations a dependent variable will change,
\cr per standard deviation increase in the predictor variable.
\cr
\cr degree_of_freedom
}
\description{
df=ez.2value(df,y,...), df[[xx]]=ez.2value(df[[xx]],...), glm(df[[yy]]~df[[xx]],family=binomial)
}
