% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.lsd}
\alias{ez.lsd}
\title{lsd}
\usage{
ez.lsd(path = ".", pattern = NULL, full = FALSE, hidden = FALSE)
}
\value{
a char vector with 0, 1 or more elements
}
\description{
lsd
}
\note{
if hidden=FALSE, not include .folders
}
