% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.lsd}
\alias{ez.lsd}
\title{lsd}
\usage{
ez.lsd(path = ".", pattern = NULL, hidden = FALSE)
}
\description{
lsd
}
\examples{
default: hidden=FALSE, not include .folders
all.files--hidden files, include.dirs--subdirs, no..--. and .. folders
}

