% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.num}
\alias{ez.2num}
\alias{ez.num}
\title{sort of a wrapper of \code{\link{type.convert}}, alias ez.2num}
\usage{
ez.num(x, ...)

ez.2num(x, ...)
}
\arguments{
\item{x}{a character vector, data frame, list, or a factor}
}
\value{
returns a converted vector
\cr with \code{\link{ez.2value}} if x is a factor with chars, will be converted to 1 2 3 etc, see its example
\cr \code{\link{ez.num}} keeps the same char as is
}
\description{
Convert a character vector, data frame to logical, integer, numeric, complex or factor when appropriate.
}
\details{
see \url{http://stackoverflow.com/a/22701462/2292993}
}
\examples{
ez.num(c(1,'2','a',3))  # -> still factor
ez.num(c(1,'2',3)) # -> int
}
\seealso{
\code{\link{ez.2value}}

Other data.transformation.functions: \code{\link{ez.2character}},
  \code{\link{ez.2factor}}, \code{\link{ez.2label}},
  \code{\link{ez.2long}}, \code{\link{ez.2lower}},
  \code{\link{ez.2value}}, \code{\link{ez.2wide}},
  \code{\link{ez.compute}}, \code{\link{ez.del}},
  \code{\link{ez.label.get}}, \code{\link{ez.label.set}},
  \code{\link{ez.leftjoin}}, \code{\link{ez.move}},
  \code{\link{ez.newcol}}, \code{\link{ez.recode2}},
  \code{\link{ez.recode}}, \code{\link{ez.recols}},
  \code{\link{ez.recol}}, \code{\link{ez.rename}},
  \code{\link{ez.replace}}, \code{\link{ez.rncols}},
  \code{\link{ez.rncol}}, \code{\link{ez.select}},
  \code{\link{ez.sort}}, \code{\link{ez.split}},
  \code{\link{ez.unique}}, \code{\link{ez.values.get}},
  \code{\link{ez.values.set}}
}

