% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.num}
\alias{ez.num}
\title{sort of a wrapper of \code{\link{type.convert}}}
\usage{
ez.num(x, ...)
}
\arguments{
\item{x}{a character vector, or a factor}
}
\value{
returns a converted vector
\cr with \code{\link{ez.2value}} if x is a factor with chars, will be converted to 1 2 3 etc, see its example
\cr \code{\link{ez.num}} keeps the same char as is
}
\description{
Convert a character vector to logical, integer, numeric, complex or factor as appropriate.
}
\details{
see \url{http://stackoverflow.com/a/22701462/2292993}
}
\seealso{
\code{\link{ez.2value}}
}

