% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.open}
\alias{ez.open}
\title{open a local file with default program (xlsx etc)}
\usage{
ez.open(filePath)
}
\description{
open a local file with default program (xlsx etc)
}
\note{
\code{\link{browseURL}} does the same
}
