% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.open}
\alias{ez.open}
\title{open a local file, or web url with associated programs, alias of \code{\link{browseURL}}}
\usage{
ez.open(url, browser = getOption("browser"), encodeIfNeeded = FALSE)
}
\description{
open a local file, or web url with associated programs, alias of \code{\link{browseURL}}
}

