% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.p.apa}
\alias{ez.p.apa}
\title{format p value according to apa for report}
\usage{
ez.p.apa(pvalue, prefix = 0, pe = F)
}
\arguments{
\item{pvalue}{numeric vector}

\item{prefix}{-1,0,1,2}

\item{pe}{affects only p < .001. if T, would be sth like 3.14e-04; otherwise < .001}
}
\value{
character vector prefix -1 (****,***,**,*,ns); 0 (< .001, .003, .02); 1 (< .001, = .003, = .02); 2 (p < .001, p = .003, p = .02)
}
\description{
format p value according to apa for report
}
