% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.pdfoff}
\alias{ez.pdfoff}
\title{wrapper of \code{\link{dev.off}}}
\usage{
ez.pdfoff(which = dev.cur())
}
\description{
wrapper of \code{\link{dev.off}}
}
\examples{
A4:     width 7(inches) height = 5
Letter: 8.5 x 11
}
\seealso{
\code{\link{ez.export}} \code{\link{ez.pdfon}}
}
