% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.pdfon}
\alias{ez.pdfon}
\title{wrapper of \code{\link{pdf}}, direct all plotting to a pdf file(s)}
\usage{
ez.pdfon(file = if (onefile) "Rplots.pdf" else "Rplot\%03d.pdf", width,
  height, onefile, family, title, fonts, version, paper, encoding, bg, fg,
  pointsize, pagecentre, colormodel, useDingbats, useKerning, fillOddEven,
  compress)
}
\description{
wrapper of \code{\link{pdf}}, direct all plotting to a pdf file(s)
}
\examples{
A4:     width 7(inches) height = 5
Letter: 8.5 x 11
ez.pdfon("Living_All.pdf",width=10,height=10,onefile=T)
ez.plot(jd.result,"ageout_rural|group")
ez.plot(jd.result,"n_rural|group")
ez.pdfoff()
}
\seealso{
\code{\link{ez.export}} \code{\link{ez.pdfoff}}
}
