% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.pprint}
\alias{ez.pprint}
\title{print out a vector for easy manual copy/processing.}
\usage{
ez.pprint(vec, quote = TRUE)
}
\arguments{
\item{vec}{a vector}

\item{quote}{TRUE/FALSE, whether add a quote around each element (switch for string or number)}
}
\value{
nothing, only print out. 
\cr By default in R when you type a variable name, you get [1] "rs171440fwd" "rs1800497fwd"
\cr now with this function you get 'rs171440fwd','rs1800497fwd','rs180043'
}
\description{
print out a vector for easy manual copy/processing.
}
\seealso{
\code{\link{ez.print}}
}

