% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.print}
\alias{ez.print}
\title{wrapper of \code{\link{cat}}}
\usage{
ez.print(...)
}
\value{
each print generates a new line automatically
}
\description{
wrapper of \code{\link{cat}}
}
\details{
do not use R \code{\link{print}}, not actually printing \\n
}
\seealso{
\code{\link{sprintf}}
}

