% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.print}
\alias{ez.print}
\title{wrapper of \code{\link{cat}}}
\usage{
ez.print(..., sep = "")
}
\arguments{
\item{...}{could be char, num, vector, or mixed (no need to convert type) e.g., 
\cr        but be careful with factors (in number)--convert with as.character(factor variable)
\cr        ez.print('a=',3)
\cr        ez.print('a=',c(3,'b'))
\cr        ez.print(var,'\\t',p)
\cr        ez.print(c(var,'\\t',p))}

\item{sep}{default empty}
}
\value{
each print generates a new line automatically
}
\description{
wrapper of \code{\link{cat}}
}
\details{
do not use R \code{\link{print}}, not actually printing \\n
}
\seealso{
\code{\link{sprintf}}, \code{\link{ez.log}}, \code{\link{ez.join}}
}
