% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.print}
\alias{ez.print}
\title{wrapper of \code{\link{cat}}}
\usage{
ez.print(..., file = "", sep = "", fill = FALSE, labels = NULL,
  append = FALSE)
}
\arguments{
\item{...}{could be char, num, vector, or mixed (no need to convert type) e.g.,
\cr        but be careful with factors (in number)--convert with as.character(factor variable)
\cr        ez.print('a=',3)
\cr        ez.print('a=',c(3,'b'))
\cr        ez.print(var,'\\t',p)
\cr        ez.print(c(var,'\\t',p))}

\item{file}{filepath if saving to a file. If printed to a file, no output in console}

\item{sep}{default empty. a character vector of strings to append after each element.}

\item{append}{T/F}
}
\value{
each print generates a new line automatically
}
\description{
wrapper of \code{\link{cat}}
}
\details{
do not use R \code{\link{print}}, not actually printing \\n
}
\examples{
# cat("hello","world",file="output.txt",sep="\\n",append=TRUE)
}
\seealso{
\code{\link{sprintf}}, \code{\link{ez.log}}, \code{\link{ez.join}}
}
