% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.r}
\alias{ez.r}
\title{r correlation}
\usage{
ez.r(x, col = NULL, type = "pearson", print2scr = T)
}
\arguments{
\item{x}{df or matrix}

\item{col}{col included for calculation, ignored if x is not a df}

\item{type}{"pearson" or "spearman"}

\item{print2scr}{print attention message, if NA present in r matrix; no effect if there is no NA}
}
\value{
r matrix
}
\description{
r correlation
}
\note{
underlying uses \code{\link[Hmisc]{rcorr}}, which uses pairwise deletion. Pairs with fewer than 2 non-missing values, or too small variance, have the r values set to NA
\cr In the r matrix you get, you may get some NAs in the marix without warning)
}
