% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.rand}
\alias{ez.rand}
\title{random number, alias of \code{\link{runif}}}
\usage{
ez.rand(n, min = 0, max = 1)
}
\arguments{
\item{runif(n, }{min = 0, max = 1)}
}
\description{
random number, alias of \code{\link{runif}}
}
