% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.reads}
\alias{ez.reads}
\title{wrapper of \code{\link[sjmisc]{read_spss}}}
\usage{
ez.reads(path, tofactor = TRUE, keepna = FALSE, tolower = FALSE, ...)
}
\arguments{
\item{path}{File path to the data file}

\item{tofactor}{if TRUE, factor variables imported from the dataset (which are imported as atomic) will be converted to factors.}

\item{keepna}{if TRUE, user-defined missing values will be left as their original codes. If FALSE (default), corresponding values are converted to NA.}

\item{tolower}{whether to convert all column names to lower case}
}
\description{
potentially keep variable labels and value labels, internally trim string space; but would not convert value labels to factor levels (i.e., gender 1/2->male/female)
}

