% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.reads2}
\alias{ez.reads2}
\title{read spss .sav file with foreign package (maybe for viewing purpose only, instead of processing)}
\usage{
ez.reads2(file, valuelabel = TRUE, tolower = FALSE)
}
\arguments{
\item{valuelabel}{logic
\cr               if True: gender=Male,Female, gender is a factor with two levels "Male/Female"
\cr               if False: gender=1,2, gender is a number with attributes "Male/Female"
\cr               (char will always be converted to factor regardless of valuelabel)}

\item{tolower}{whether to convert all column names to lower case}
}
\description{
cannot trim string space (trim.factor.names, trim_values in read.spss not working??), but can convert value labels to factor levels for easy viewing (i.e., gender 1/2->male/female) when valuelabel=TRUE (see below), see more at \code{\link[foreign]{read.spss}}
}
\examples{
(file, valuelabel=TRUE,tolower=FALSE)

alternatively, one can use SPSS R plugin to pass data between SPSS and R.
}

