% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.readx}
\alias{ez.readx}
\title{read an xlsx file, wrapper of \code{\link[openxlsx]{read.xlsx}}}
\usage{
ez.readx(file, sheet = 1, tolower = FALSE, stringsAsFactors = TRUE,
  makenames = TRUE, ...)
}
\arguments{
\item{tolower}{whether to convert all column names to lower case}

\item{stringsAsFactors}{T/F}

\item{makenames}{if F, keep as is. if T, call make.names(unique=TRUE,allow_=TRUE) _ kept, The character "X" is prepended if necessary. All invalid characters are translated to "." .1 .2 etc appended for uniqueness}
}
\value{
when stringsAsFactors=T, in the returned data frame, string to factor
\cr number stored as text in excel (->string) -> factor
}
\description{
uses openxlsx package which does not require java and is much faster, but has a slightly different interface/parameters from xlsx package. internally trim (leading and trailing) string spaces
}
\examples{
read.xlsx(xlsxFile, sheet = 1, startRow = 1, colNames = TRUE,
         rowNames = FALSE, detectDates = FALSE, skipEmptyRows = TRUE,
         rows = NULL, cols = NULL, check.names = FALSE, namedRegion = NULL)
}
