% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.readx2}
\alias{ez.readx2}
\title{read an xlsx file, wrapper of \code{\link[xlsx]{read.xlsx}} from the xlsx package, internally trim (leading and trailing) string spaces}
\usage{
ez.readx2(file, sheetIndex = 1, tolower = FALSE, stringsAsFactors = TRUE,
  blanksAsNA = TRUE, na.strings = c("", "."), makenames = TRUE,
  detectDates = TRUE, ...)
}
\arguments{
\item{tolower}{whether to convert all column names to lower case}

\item{stringsAsFactors}{T/F}

\item{blanksAsNA}{T/F, converts factor or character vector elements to NA if matching na.strings}

\item{na.strings}{only applicable if blanksAsNA=T. e.g., c('','.','NA','na','N/A','n/a','NaN','nan')}

\item{makenames}{if F, keep as is. if T, call make.names(unique=TRUE,allow_=TRUE) _ kept, The character "X" is prepended if necessary. All invalid characters are translated to "." .1 .2 etc appended for uniqueness
\cr in fact, makenames is ignored, because xlsx::read.xlsx internally deals with col names. I still keep it for consistency with ez.readx}

\item{detectDates}{If TRUE, attempt to recognise dates and perform conversion.
\cr in fact, detectDates is ignored, because xlsx::read.xlsx internally deals with dates. I still keep it for consistency with ez.readx}
}
\value{
when stringsAsFactors=T, in the returned data frame, string to factor
\cr number stored as text in excel (->string) -> factor
\cr dates and datetimes will be auto converted, duplicate or illegal col names as variables will also be auto dealt with
}
\description{
read an xlsx file, wrapper of \code{\link[xlsx]{read.xlsx}} from the xlsx package, internally trim (leading and trailing) string spaces
}
\examples{
read.xlsx(file, sheetIndex, sheetName=NULL, rowIndex=NULL,
          startRow=NULL, endRow=NULL, colIndex=NULL,
          as.data.frame=TRUE, header=TRUE, colClasses=NA,
          keepFormulas=FALSE, encoding="unknown", ...)
colClasses: Only numeric, character, Date, POSIXct, column types are accepted
colClasses=c("Date", "character","integer", rep("numeric", 2),  "POSIXct")
}
