% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.readxlist}
\alias{ez.readxlist}
\title{read all sheets in an xlsx file, optionally prints sheet names, same syntax as \code{\link{ez.readx}}, wrapper of \code{\link[openxlsx]{getSheetNames}} from the openxlsx package}
\usage{
ez.readxlist(file, print2screen = TRUE, tolower = FALSE,
  stringsAsFactors = TRUE, blanksAsNA = TRUE, na.strings = c("", "."),
  makenames = TRUE, ...)
}
\arguments{
\item{print2screen}{print out sheet indices and names, default TRUE}

\item{tolower}{whether to convert all column names to lower case}

\item{stringsAsFactors}{T/F}

\item{blanksAsNA}{T/F, converts factor or character vector elements to NA if matching na.strings}

\item{na.strings}{only applicable if blanksAsNA=T. e.g., c('','.','NA','na','N/A','n/a','NaN','nan')}

\item{makenames}{if F, keep as is. if T, call make.names(unique=TRUE,allow_=TRUE) _ kept, The character "X" is prepended if necessary. All invalid characters are translated to "." .1 .2 etc appended for uniqueness}
}
\value{
a list of sheet as data frame. To get sheetnames, names(result)
\cr when stringsAsFactors=T, in the returned data frame, string to factor
\cr number stored as text in excel (->string) -> factor
}
\description{
read all sheets in an xlsx file, optionally prints sheet names, same syntax as \code{\link{ez.readx}}, wrapper of \code{\link[openxlsx]{getSheetNames}} from the openxlsx package
}
\examples{
readxlist(file, sheet = 1, startRow = 1, colNames = TRUE,
         rowNames = FALSE, detectDates = FALSE, skipEmptyRows = TRUE,
         rows = NULL, cols = NULL, check.names = FALSE, namedRegion = NULL)
}
