% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.regexprep}
\alias{ez.regexprep}
\title{replace string or string vectors using regular expression (case sensitive)}
\usage{
ez.regexprep(s, expr, repstr, ignorecase = FALSE, once = FALSE)
}
\description{
replace string or string vectors using regular expression (case sensitive)
}
\seealso{
\code{\link{ez.strreplace}}
}

