% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.regressions}
\alias{ez.regressions}
\title{a series of simple regression, for many y and many x}
\usage{
ez.regressions(df, y, x, pthreshold = 0.05, showerror = F)
}
\arguments{
\item{df}{a data frame, if its column is factor, auto converts to numeric (internally call ez.2value(df))}

\item{y}{a vector of outcome variables c('var1','var2'), or a single variable 'var1'}

\item{x}{a vector of predictors, or a single predictor, (eg, names(select(beta,Gender:dmce)))}

\item{pthreshold}{default .05, print/output results whenever p < pthreshold, could be 1 then get all}

\item{showerror}{whether show error message when error occurs, default F}
}
\value{
an invisible data frame with y,x,p,b and print results out on screen; results can then be saved using ez.savex(results,'results.xlsx')
}
\description{
lm(scale(df[[yy]])~scale(df[[xx]]))
}

