% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{ez.replace}
\alias{ez.replace}
\title{replace a single value in data frame with another value}
\usage{
ez.replace(df, col, oldval, newval)
}
\arguments{
\item{df}{data frame}

\item{col}{column name in string}

\item{oldval}{old value (e.g., -Inf, NA)}

\item{newval}{new value (e.g., NA)}
}
\value{
returns a new df, old one does not change
}
\description{
smilar to \code{\link{ez.recode}} numeric->char, char->char, factor->factor
\cr wrapper of df[[col]][which(df[[col]]==oldval)] <- newval
\cr the "==" syntax within "which()" could be modified
}
\examples{
data=data.frame(a=factor(c(1,2)))
ez.replace(data,'a',1,3) \%>\% .$a
ez.replace(data,'a',1,'abc') \%>\% .$a
          # a was factor with level (1,2), now is factor with level (2,3), (2 abc)

data=data.frame(a=c('r1','r2'))
ez.replace(data,'a','r1',3) \%>\% .$a
          # a was factor with level ('r1','r2'), now is factor with level ('3','r2')

data=data.frame(a=c('r1','r2'),stringsAsFactors = F)
ez.replace(data,'a','r1',3) \%>\% .$a
ez.replace(data,'a','r1',NA) \%>\% .$a 
          # a was char, now is still char, (NA counted here as char)

data=data.frame(a=c(1,2))
ez.replace(data,'a',1,111) \%>\% .$a
ez.replace(data,'a',1,NA) \%>\% .$a
          # a was numeric, now is still numeric, (NA counted here as numeric)

ez.replace(data,'a',1,'111') \%>\% .$a
ez.replace(data,'a',1,'abc') \%>\% .$a
          # a was numeric, now is character
}
\seealso{
\code{\link{ez.strreplace}} \code{\link{ez.recode}} \code{\link{ez.recode2}}

Other data.transformation.functions: \code{\link{ez.2factor}},
  \code{\link{ez.2label}}, \code{\link{ez.2long}},
  \code{\link{ez.2lower}}, \code{\link{ez.2value}},
  \code{\link{ez.2wide}}, \code{\link{ez.compute}},
  \code{\link{ez.del}}, \code{\link{ez.label.get}},
  \code{\link{ez.label.set}}, \code{\link{ez.leftjoin}},
  \code{\link{ez.move}}, \code{\link{ez.newcol}},
  \code{\link{ez.recode2}}, \code{\link{ez.recode}},
  \code{\link{ez.recols}}, \code{\link{ez.recol}},
  \code{\link{ez.rename}}, \code{\link{ez.rncols}},
  \code{\link{ez.rncol}}, \code{\link{ez.select}},
  \code{\link{ez.sort}}, \code{\link{ez.split}},
  \code{\link{ez.unique}}, \code{\link{ez.values.get}},
  \code{\link{ez.values.set}}
}

