% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.repo}
\alias{ez.repo}
\title{print out or set the repo}
\usage{
ez.repo(repo = NULL)
}
\arguments{
\item{repo}{NULL=print out current repo; 'default'='https://cran.rstudio.com/'; else like '2016-08-01'='https://mran.revolutionanalytics.com/snapshot/2016-08-01'}
}
\value{
returns NULL
}
\description{
print out or set the repo
}

