% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.savep}
\alias{ez.savep}
\alias{ez.writep}
\title{save a df to a parquet (a columnar storage file format) file}
\usage{
ez.savep(
  x,
  sink,
  chunk_size = NULL,
  version = "2.4",
  compression = default_parquet_compression(),
  compression_level = NULL,
  use_dictionary = NULL,
  write_statistics = NULL,
  data_page_size = NULL,
  use_deprecated_int96_timestamps = FALSE,
  coerce_timestamps = NULL,
  allow_truncated_timestamps = FALSE
)

ez.writep(
  x,
  sink,
  chunk_size = NULL,
  version = "2.4",
  compression = default_parquet_compression(),
  compression_level = NULL,
  use_dictionary = NULL,
  write_statistics = NULL,
  data_page_size = NULL,
  use_deprecated_int96_timestamps = FALSE,
  coerce_timestamps = NULL,
  allow_truncated_timestamps = FALSE
)
}
\description{
save a df to a parquet (a columnar storage file format) file
}
\note{
alias of \code{\link[arrow]{write_parquet}}
}
