% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ez.saver}
\alias{ez.saver}
\alias{ez.writer}
\title{save one (only one) object to a rds file}
\usage{
ez.saver(object, file = "", ascii = FALSE, version = NULL,
  compress = TRUE, refhook = NULL)

ez.writer(object, file = "", ascii = FALSE, version = NULL,
  compress = TRUE, refhook = NULL)
}
\description{
save one (only one) object to a rds file
}
\note{
x=5; saveRDS(x, "my.rds")
\cr 
\cr usage for save (you have to always explicitly specify file=):
\cr      1) save(x, file='my.rda'), save(x, y, file='my.rda'), save('x', y, file='my.rda')
\cr      2) save(list='x', file='my.rda'), save(list=c('x', 'y'), file='my.rda')
\cr      the variable name will also be the same as when you save
\cr      if you want a different name, you have to assign variable to that name first
\cr 
\cr save.image(file='my.rda')
\cr 
\cr load(file)
}
\seealso{
\code{\link{saveRDS}}
}
