% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.scatterplot}
\alias{ez.scatterplot}
\title{scatter plot with ggplot}
\usage{
ez.scatterplot(df, cmd, rp.size = 5, rp.x = 0.95, rp.y = 0.95,
  point.alpha = 0.95, point.size = 3, rug.size = 0.5, ylab = NULL,
  xlab = NULL, zlab = NULL, legend_position = "top",
  legend_direction = "horizontal", legend_box = T, legend_size = c(0, 10),
  rp = TRUE, se = TRUE, rug = TRUE, ellipse = FALSE)
}
\arguments{
\item{df}{data frame}

\item{cmd}{like "y~x", "y~x|z", "y~x||z" where y x are continous, z discrete (| one regression line, || multiple regression lines by levels of z)}

\item{rp}{show r squared and p values}

\item{se}{standard error of linear regression line}

\item{rug}{marginal rug indicating univariate distribution}

\item{ellipse}{draw confidence ellipses, powered by stat_ellipse()}
}
\value{
a ggplot object (+theme_apa() to get apa format plot)
}
\description{
scatter plot with ggplot
}

