% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.splitpath}
\alias{ez.splitpath}
\title{splitpath}
\usage{
ez.splitpath(path)
}
\description{
splitpath
}
\examples{
# ez.splitpath(path)
# returns
# $dir
# [1] "/Users/jerry/Downloads"

# $file
# [1] "026999397379845a"

# $ext
# '.pdf
}
