% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.strreplace}
\alias{ez.strreplace}
\title{replace string or string vectors literally}
\usage{
ez.strreplace(s, expr, repstr, once = FALSE)
}
\description{
support single string, vectors
\cr case sensitive! wrapper of sub, gsub
}
\seealso{
\code{\link{ez.recode}}, \code{\link{ez.recode2}}, \code{\link{ez.replace}}, \code{\link{ez.replacewhen}}, \code{\link{ez.2label}}, \code{\link{ez.factorname}}, \code{\link{ez.strreplace}}, \code{\link{ez.strrev}}, \code{\link{ez.regexprep}}, \code{\link{ez.regexprepi}}
}
