% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.strreplace}
\alias{ez.strreplace}
\title{replace string or string vectors literally}
\usage{
ez.strreplace(s, expr, repstr, once = FALSE)
}
\description{
support single string, vectors
\cr case sensitive! wrapper of sub, gsub
}
\seealso{
\code{\link{ez.regexprep}} \code{\link{ez.regexprepi}}
}

