% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.time}
\alias{ez.time}
\title{convert to time}
\usage{
ez.time(x, ori = "SPSS", format = "\%H:\%M", out.type = "numeric", ...)
}
\arguments{
\item{x}{a vector of number or character, hms/difftime}

\item{ori}{one of 'Excel', 'SPSS' (ignored if x is character)}

\item{format}{input format, see examples (more formats at \code{\link{strptime}}). ignored if x is numeric/hms}

\item{out.type}{string, 'numeric' (fractions of a day) or 'times'/'time'}
}
\value{
returns a vector of number (class numeric) or time (class times). class times can be passed to as.character(.), or substr(.,1,5), or as.numeric(.), see more at \code{\link[chron]{chron}} chron stores as a fraction of a day, if you do as.numeric
}
\description{
convert from a time (numeric, character/factor, hms/difftime) to class (chron) times (17:48:00--h:m:s military only) or to class numeric (0.7416667--fractions of a day), call \code{\link[chron]{times}}: 
\cr \cr SPSS military time (civilian time not supported) specified in date format, 17:48 read into R as 64080 (seconds of a day, foreign package), or as hms/difftime (same as SPSS, haven package), or 
\cr \cr Excel military/civilian time specified in time format, 5:48:00 PM, 17:48:00 read into R as 0.7416667 (fractions of a day), or
\cr \cr Pass directly a string "17:48:00" (format specified by param format)
}
\examples{
ez.time("5:20:31 am",format="\%I:\%M:\%S \%p",out.type="time")  # 5, 05; am AM aM Am OK. \%I = 12 hr 
ez.time("15:32", format="\%H:\%M",out.type="time")  # 5 05 OK.
ez.time("5:2", format="\%H:\%M",out.type="time")
}
\seealso{
\code{\link{ez.date}} \code{\link{ez.is.date}} \code{\link{ez.is.date.convertible}} \code{\link{ez.age}} \code{\link{ez.time}}
}
