% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ez.trim}
\alias{ez.trim}
\title{Merge Multiple spaces to single space, and remove trailing/leading spaces}
\usage{
ez.trim(s)
}
\arguments{
\item{s}{a string}
}
\description{
underlying function is \code{\link{gsub}} with regular expression
}
\examples{
 "Hi        buddy        what's up    Bro"  --> "Hi buddy what's up bro"
}
\seealso{
\code{\link{trimws}}
}

