% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.vi}
\alias{ez.vi}
\title{print sorted uniques of a df col or a vector (NA last) and other information}
\usage{
ez.vi(x, printn = 35, plot = TRUE, ...)
}
\arguments{
\item{printn}{print first n and last n (useful for loooong vector). If 2n >= total length, print all. Inf=all}

\item{plot}{plot single vector (call generic/default \code{\link[graphics]{plot}})}
}
\description{
vc (c(vv)), vi (view everything print out), vv (view format vector), vx (view excel), View (built-in).
}
