% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.vi}
\alias{ez.vi}
\alias{vi}
\title{print sorted uniques of a df col or a vector (NA last) and other information}
\usage{
ez.vi(x)

vi(x)
}
\description{
vi (view everything print out), vv (view format vector), vx (view excel), View (built-in). print sorted uniques of a df col or a vector (NA last) and other information
}
