% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ez.view}
\alias{ez.view}
\alias{xview}
\alias{yview}
\title{view the overview of a data frame or similar object (like spss variable view, but with much more information)}
\usage{
ez.view(df, id = NULL, file = NULL, width = 300, characterize = TRUE,
  incomparables = FALSE, debug = NULL, ...)

xview(df, id = NULL, file = NULL, width = 300, characterize = TRUE,
  incomparables = FALSE, debug = NULL, ...)

yview(x, title)
}
\arguments{
\item{df}{a data frame}

\item{id}{a single col name in string or number (eg, 'age' or 3), that serves as (potentially unique) id, except which duplicated rows will be checked against. If NULL, rownames() will be auto used}

\item{file}{a file name, if NULL, a temp generated, will save more detailed variable information to an excel file}

\item{width}{controls if too many factor levels to print, eg 300. NULL=unlimited}

\item{characterize}{T/F count the element freq of character cols or not}
}
\value{
returns a list $row, $col, $dat (input data frame), $pth (file path)
}
\description{
Updated: as of Thu, Nov 30 2017, not any more a wrapper of \code{\link[sjPlot]{view_df}}; can make the html bigger by openning in internet browser
}

