% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os.R
\name{ez.who}
\alias{ez.who}
\title{alias of \code{\link{sessionInfo}}, \code{\link{ez.whos}}}
\usage{
ez.who(package = NULL)
}
\description{
alias of \code{\link{sessionInfo}}, \code{\link{ez.whos}}
}
\seealso{
\code{\link{objects}}
}
