% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ez.xyplot}
\alias{ez.xyplot}
\title{mimic xyplot with ggplot (slightly horizontally jittered)}
\usage{
ez.xyplot(df, cmd)
}
\arguments{
\item{df}{data frame in long format}

\item{cmd}{like "y|x,g" or "y|x z,g" where y is continous, x z are discrete, g is individual/grouping variable
\cr 'FinalMem|Attention, SubjectID'     'FinalMem|Attention Condition, SubjectID'}
}
\value{
a ggplot object (+theme_apa() to get apa format plot)
}
\description{
mimic xyplot with ggplot (slightly horizontally jittered)
}

