% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggpass}
\alias{ggpass}
\title{pass through a ggplot object, see examples}
\usage{
ggpass(pp)
}
\description{
pass through a ggplot object, see examples
}
\note{
Assuming pp has $df (preferred to $data), or $data. Make sure $df/$data is what you intend to plot
}
\examples{
df \%>\% ggplot1() \%>\% ggpass() \%>\% ggplot2()
#however, if "+" is used in ggplot, use the following format
df \%>\% {ggplot1(.,)+xlab('')} \%>\% ggpass() \%>\% {ggplot2(.,)+ylab('')}
}
