% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{view}
\alias{view}
\title{print sorted uniques of a df col or a vector (NA last) and other information}
\usage{
view(x)
}
\description{
print sorted uniques of a df col or a vector (NA last) and other information
}

