% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{z.export}
\alias{z.export}
\title{alias of pdf}
\usage{
z.export(file = ifelse(onefile, "Rplots.pdf", "Rplot\%03d.pdf"), width,
  height, onefile, family, title, fonts, version, paper, encoding, bg, fg,
  pointsize, pagecentre, colormodel, useDingbats, useKerning, fillOddEven,
  compress)
}
\description{
alias of pdf
}
\examples{
first z.export(filename) to create a new canvas
then plot all following graphs to it and
finally dev.off()/z.clean(3) to unlink
pdf(file = ifelse(onefile, "Rplots.pdf", "Rplot\%03d.pdf"),
width, height, onefile, family, title, fonts, version,
paper, encoding, bg, fg, pointsize, pagecentre, colormodel,
useDingbats, useKerning, fillOddEven, compress)
i = 100
fileName = z.join('plot', i, '.pdf')
}

