% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{z.export}
\alias{z.export}
\title{wrapper of \code{\link{dev.copy2pdf}}}
\usage{
z.export(filename = "RPlot.pdf", pdf.width = 7, pdf.height = 5, ...)
}
\description{
wrapper of \code{\link{dev.copy2pdf}}
}
\seealso{
\code{\link{pdf}}
}

