% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{z.heatmap}
\alias{z.heatmap}
\title{plot a heatmap with values shown}
\usage{
z.heatmap(df, id, show.values = T, angle = 270, colors = c("blue",
  "white", "red"))
}
\arguments{
\item{df}{data frame in wide format, like a correlation matrix}

\item{id}{a column name as id, will be shown as x axis, quoted ""}

\item{show.values}{whether to show values, default TRUE}

\item{angle}{the x axis label angle, default=270 (vertical), suggests 330 if label is not too long}

\item{colors}{low, middle, high; default c("blue", "white", "red")}
}
\value{
a ggplot object (+theme_apa() to get apa format plot)
}
\description{
plot a heatmap with values shown
}

