% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{z.heatmap}
\alias{z.heatmap}
\title{plot a heatmap with values shown}
\usage{
z.heatmap(df, id, show.values = F, remove.zero = T, angle = 270,
  colors = c("blue", "white", "red"), xsize = 1, ysize = 1,
  legend.position = "right")
}
\arguments{
\item{df}{data frame in wide format, like a correlation matrix}

\item{id}{a column name as id, will be shown as x axis, quoted ""}

\item{show.values}{whether to show values in addition to color in the plot}

\item{remove.zero}{remove the leading 0 as in correlation, 0.02->.02 (0.00 becomes "<.01")
\cr only works when show.value=T}

\item{angle}{the x axis label angle, default=270 (vertical), suggests 330 if label is not too long}

\item{colors}{low, middle, high}

\item{xsize}{x axis label font size}

\item{ysize}{y axis label font size}

\item{legend.position}{"bottom", "top", "left", "right", "none"}
}
\value{
a ggplot object (+theme_apa() to get apa format plot)
}
\description{
plot a heatmap with values shown
}

