% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{z.reads}
\alias{z.reads}
\title{alias of \code{\link[sjmisc]{read_spss}}}
\usage{
z.reads(path, enc = NA, attach.var.labels = FALSE, atomic.to.fac = FALSE,
  keep.na = FALSE, option = "haven")
}
\description{
potentially keep variable labels and value labels.
}

