% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{z.reads2}
\alias{z.reads2}
\title{read spss .sav file with foreign package}
\usage{
z.reads2(file, valuelabel = TRUE, tolower = FALSE)
}
\description{
read spss .sav file with foreign package
}
\examples{
(file, valuelabel=TRUE,tolower=FALSE)
label means variable value label, not variable label
if TRUE, varialbe becomes a factor, and its levels becomes character labels
ie, without label False gender=1,2; with label True gender=Male,Female
tolower, whether convert all variable names to lower case

alternatively, one can use SPSS R plugin to pass data between SPSS and R.
}

