% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{z.recols}
\alias{z.recols}
\title{reorder all cols}
\usage{
z.recols(df, newColOrder)
}
\arguments{
\item{newColOrder}{c('','',''), number of cols must match}
}
\value{
returns a new df, old one does not change
}
\description{
reorder all cols
}
\seealso{
\code{\link[tidyr]{gather}}, \code{\link[tidyr]{spread}}, \code{\link[tidyr]{separate}}, \code{\link[tidyr]{unite}}
\cr \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}
\cr \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{arrange}}
\cr \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{mutate}}
\cr \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{right_join}}, \code{\link[dplyr]{inner_join}}, \code{\link[dplyr]{full_join}}, \code{\link[dplyr]{semi_join}}, \code{\link[dplyr]{anti_join}}
\cr \code{\link[dplyr]{intersect}}, \code{\link[dplyr]{union}}, \code{\link[dplyr]{setdiff}}
\cr \code{\link[dplyr]{bind_rows}}, \code{\link[dplyr]{bind_cols}}

Other data.transformation.functions: \code{\link{z.compute}},
  \code{\link{z.delete}}, \code{\link{z.del}},
  \code{\link{z.leftjoin}}, \code{\link{z.move}},
  \code{\link{z.newcol}}, \code{\link{z.recode}},
  \code{\link{z.recol}}, \code{\link{z.rename}},
  \code{\link{z.rncols}}, \code{\link{z.rncol}},
  \code{\link{z.select}}, \code{\link{z.sort}},
  \code{\link{z.split}}, \code{\link{z.unique}}
}

